% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCase.R
\name{addCase}
\alias{addCase}
\title{Add observed case(s).}
\usage{
addCase(case = 1, type = "observed", token = "both",
  text.size = 0.5, col = "red", pos = 1)
}
\arguments{
\item{case}{Numeric. Vector of case ID(s).}

\item{type}{Character. Type of case: "observed" or "expected".}

\item{token}{Character. Type of token to plot: "point", "id" or "both".}

\item{text.size}{Numeric. Size of case ID text.}

\item{col}{Character. Color.}

\item{pos}{Numeric. Text position.}
}
\description{
Add case(s), as "address" or "fatalities" as points or IDs, to a plot.
}
\examples{
snowMap(add.cases = FALSE)
addCase(1)

snowMap(add.cases = FALSE)
addCase(100)
}
