% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCases.R
\name{addCases}
\alias{addCases}
\title{Add observed cases by walking neighborhood.}
\usage{
addCases(case = 1, type = "observed", token = "both",
  text.size = 0.5, col = "red", ...)
}
\arguments{
\item{case}{Numeric. Vector of case ID(s).}

\item{type}{Character. Type of case: "observed" or "expected".}

\item{token}{Character. Type of token to plot: "point", "id" or "both".}

\item{text.size}{Numeric. Size of case ID text.}

\item{col}{Character. Color.}

\item{...}{Additional plotting parameters.}
}
\description{
Add cases, as "address" or "fatalities" as points or IDs, to a plot.
}
\examples{
\dontrun{

snowMap(add.cases = FALSE)
addCase(1)

snowMap(add.cases = FALSE)
addCase(100)
}
}
