% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowMap.R
\name{addFrame}
\alias{addFrame}
\title{Add map border to plot.}
\usage{
addFrame(vars = c("x", "y"), col = "black", ...)
}
\arguments{
\item{vars}{Character. Coordinate.}

\item{col}{Character. Color}

\item{...}{Additional plotting parameters.}
}
\description{
Add map border to plot.
}
