% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLandmarks.R
\name{addLandmarks}
\alias{addLandmarks}
\title{Add landmarks.}
\usage{
addLandmarks()
}
\value{
Add base R points and text to a graphics plot.
}
\description{
Adds Golden Square, Lion Brewery, St James Workhouse, St Luke's Church (Henry Whitehead), Soho Square, Falconberg Court & Mews, 18 Sackville (John Snow residence) and 28 Dean Street (Karl Marx residence) to an existing plot. Falconberg Court & Mews are technically an isolate that is not part of the network of roads and is, consequently, technically unreachable.
}
\examples{
snowMap(add.landmarks = FALSE)
addLandmarks()
}
\seealso{
\code{\link[cholera]{snowMap}}, \code{\link[cholera]{addWhitehead}}
}

