% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLandmarkSquares.R
\name{addLandmarkSquares}
\alias{addLandmarkSquares}
\title{Add Golden and Soho Squares to plot.}
\usage{
addLandmarkSquares(latlong = FALSE, text.size = 0.5, text.col = "black")
}
\arguments{
\item{latlong}{Logical. Use estimated longitude and latitude.}

\item{text.size}{Numeric. cex for text labels.}

\item{text.col}{Character. col for text labels.}
}
\value{
Base R points and text.
}
\description{
Add Golden and Soho Squares to plot.
}
\examples{
snowMap()
addLandmarkSquares()
}
