% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMilePosts.R
\name{addMilePosts}
\alias{addMilePosts}
\title{Add walking path mileposts.}
\usage{
addMilePosts(pump.subset = NULL, pump.select = NULL,
  milepost.interval = 50, multi.core = FALSE)
}
\arguments{
\item{pump.subset}{Numeric. Vector of pumps to select (subset) from neighborhoods defined by "pump.select". Negative selection possible. NULL selects all pumps in "pump.select".}

\item{pump.select}{Numeric. Numeric vector of pumps to define pump neighborhoods (i.e. the "population"). Negative selection possible. NULL selects all pumps.}

\item{milepost.interval}{Numeric. Milepost interval in meters.}

\item{multi.core}{Logical or Numeric. TRUE uses parallel::detectCores(). FALSE uses one, single core. You can also specify the number logical cores. On Window, only "multi.core = FALSE" is available.}
}
\value{
A list of data frames.
}
\description{
in-progress prototype.
Return coordinates, post labels and post angles.
}
