% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNeighborhoodSnow.R
\name{addNeighborhoodSnow}
\alias{addNeighborhoodSnow}
\title{Adds Snow's graphical annotation of the Broad Street pump walking neighborhood.}
\usage{
addNeighborhoodSnow(latlong = FALSE, type = "area.polygons",
  non.snow.cases = FALSE, alpha.level = 1/3, polygon.type = "solid",
  polygon.col = NULL, polygon.lwd = NULL)
}
\arguments{
\item{latlong}{Logical.}

\item{type}{Character. Type of annotation plot: ""area.points", "area.polygons", or "roads".}

\item{non.snow.cases}{Logical. Plot anchor cases outside Snow neighborhood.}

\item{alpha.level}{Numeric. Alpha level transparency for area plot: a value in [0, 1].}

\item{polygon.type}{Character. "border" or "solid".}

\item{polygon.col}{Character.}

\item{polygon.lwd}{Numeric.}
}
\description{
Adds Snow's graphical annotation of the Broad Street pump walking neighborhood.
}
\examples{
\dontrun{
plot(neighborhoodVoronoi())
addNeighborhoodSnow()
}
}
