% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLandmarks.R
\name{addPlaguePit}
\alias{addPlaguePit}
\title{Add plague pit (Marshall Street).}
\usage{
addPlaguePit(col = "black", lty = "solid")
}
\arguments{
\item{col}{Character. Color of circle. Default is "black".}

\item{lty}{Character. Type of line for circle. Default is "solid".}
}
\value{
Adds a polygon, based on multiple base R line segments, to a graphics plot.
}
\description{
Draws a polygon for the plague pit located around Marshall Street. From Vestry Report map.
}
\examples{
snowMap(add.landmarks = FALSE)
addPlaguePit()
}
\seealso{
\code{\link[cholera]{addLandmarks}}, \code{\link[cholera]{addWhitehead}}
}

