% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPump.R
\name{addPump}
\alias{addPump}
\title{Add selected pump(s) to plot.}
\usage{
addPump(pump.select = NULL, vestry = FALSE, col = NULL, pch = 24,
  label = TRUE, pos = 1, ...)
}
\arguments{
\item{pump.select}{Numeric or Integer. Vector of water pump numerical ID(s). With \code{vestry = TRUE}, whole number(s) between 1 and 14. With "vestry = FALSE", whole number(s) between 1 and 13. See \code{cholera::pumps.vestry }and \code{cholera::pumps} for IDs and details about specific pumps. NULL plots all pumps. Negative selection allowed.}

\item{vestry}{Logical. TRUE for the 14 pumps from Vestry Report. FALSE for the original 13 pumps.}

\item{col}{Character. Color of pump points.}

\item{pch}{Numeric. Shape of point character.}

\item{label}{Logical. TRUE adds text label.}

\item{pos}{Numeric. Position of label.}

\item{...}{Additional plotting parameters.}
}
\description{
Add selected pump(s) to plot.
}
