% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowMap.R
\name{addRoads}
\alias{addRoads}
\title{Add all streets and roads to plot.}
\usage{
addRoads(vars = c("x", "y"), col = "gray")
}
\arguments{
\item{vars}{Character. Coordinates.}

\item{col}{Character. Color}
}
\description{
Add all streets and roads to plot.
}
