% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSnow.R
\name{addSnow}
\alias{addSnow}
\title{Adds Snow's Annotation  of the Broad Street pump walking neighborhood.}
\usage{
addSnow(streets = TRUE, color = "dodgerblue")
}
\arguments{
\item{streets}{Logical. TRUE plots streets. FALSE plots orthogonal area.}

\item{color}{Character. Color for neighborhood annotation.}
}
\description{
Reproduces Snow's graphic annotation in the Vestry Report.
}
\examples{
plot(neighborhoodVoronoi())
addSnow()
}
