% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTimePosts.R
\name{addTimePosts}
\alias{addTimePosts}
\title{Add walking path timeposts.}
\usage{
addTimePosts(pump.subset = NULL, pump.select = NULL,
  timepost.interval = 60, walking.speed = 5, multi.core = FALSE)
}
\arguments{
\item{pump.subset}{Numeric. Vector of pumps to select (subset) from neighborhoods defined by "pump.select". Negative selection possible. NULL selects all pumps in "pump.select".}

\item{pump.select}{Numeric. Numeric vector of pumps to define pump neighborhoods (i.e. the "population"). Negative selection possible. NULL selects all pumps.}

\item{timepost.interval}{Numeric. Timepost interval in seconds.}

\item{walking.speed}{Numeric. Walking speed in km/hr.}

\item{multi.core}{Logical or Numeric. TRUE uses parallel::detectCores(). FALSE uses one, single core. You can also specify the number logical cores. On Window, only "multi.core = FALSE" is available.}
}
\value{
A list of data frames.
}
\description{
Return coordinates, post labels and post angles: in-progress prototype.
}
