% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVoronoiB.R
\name{addVoronoiB}
\alias{addVoronoiB}
\title{Add Voronoi cells (latlong).}
\usage{
addVoronoiB(pump.select = NULL, vestry = FALSE)
}
\arguments{
\item{pump.select}{Numeric. Default is NULL; all pumps are used. Otherwise, selection by a vector of numeric IDs: 1 to 13 for \code{pumps}; 1 to 14 for \code{pumps.vestry}. Exclusion (negative selection) is possible (e.g., -6).}

\item{vestry}{Logical. \code{FALSE} for original 13 pumps. TRUE for 14 pumps in Vestry Report.}
}
\description{
Add Voronoi cells (latlong).
}
