% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anchor.case}
\alias{anchor.case}
\title{Anchor or base case of each stack of fatalities.}
\format{\describe{
    \item{\code{case}}{numerical case ID}
    \item{\code{anchor.case}}{numerical case ID of anchor.case}
}}
\usage{
anchor.case
}
\description{
Data frame that links a case to its stack. which is the case at the base of a stack. For use with \code{\link{caseLocator}}.
}
\section{Notes}{
 \code{\link{unstackFatalities}} documents the code for these data.
}

\keyword{datasets}
