% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{border}
\alias{border}
\title{Numeric IDs of line segments that create the map's border frame.}
\format{
\describe{
    \item{\code{border}}{numerical ID}
}
}
\usage{
border
}
\description{
Vector of ordered numbers that identify the line segments that make up the frame of the map. For use with sp::Polygon().
}
\keyword{datasets}
