% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caseDistance.R
\name{caseDistance}
\alias{caseDistance}
\title{Compute distance between case fatalities (meters).}
\usage{
caseDistance(a = 19, b = 263, latlong = FALSE)
}
\arguments{
\item{a}{Numeric. Case ID.}

\item{b}{Numeric. Case ID.}

\item{latlong}{Logical.}
}
\description{
Compute distance between case fatalities (meters).
}
