% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cholera.R
\docType{package}
\name{cholera-package}
\alias{cholera-package}
\title{cholera: amend, augment and aid analysis of John Snow's cholera map}
\description{
Amend, augment and aid the analysis of John Snow's cholera map.
}
