\name{cholera}
\Rdversion{0.1}
\alias{cholera-package}

\docType{data}

\title{
Introduction
}

\description{
In his map of the 1854 London cholera outbreak, John Snow used a stack of horizontal bars to represent the number of fatalities at individual address. In 1992, Dodson and Tobler digitized Snow's map (\code{http://www.ncgia.ucsb.edu/pubs/snow/snow.html}). This effort is probably the most complete representation of the data in Snow's map. While the original data is no longer available, the data are preserved in the Michael Friendly's \code{HistData} package.

The \code{cholera} package amends and augments Dodson and Tobler's data. It ixes three apparent coding errors, provides two version of the "unstacked" data, appends road names to the street data, adds fatalities time series data, and provides three versions of pump neighborhoods.
}

\details{
The coding errors stem from three apparently misplaced cases in Dodson and Tobler's data. While the data record 578 bars, only 575 have unique x-y coordinates. There are three pairs with identical coordinates: 1) 93 and 214; 2) 91 and 241; and 3) 209 and 429. Within Snow's scheme of stacking bars, this cannot occur: each bar in a stack should have a unique x-y coordinate. For this reason, duplicates are likely to be coding errors. Fortunately, a careful comparison of Snow's map and the map generated by Dodson and Tobler's data reveals that there are three "missing" bars in the latter. An expedient solution would be to use the duplicate observations to fill in for the "missing" ones. A explanation of why this solution is both plausible and reasonable is discussed in \code{vignette}("duplicate.missing.cases")

"Unstacked" versions of the data are valuable to analysts because the position of bars in Snow's map do not always reflect the actual physical location of a case. In fact, 44 percent of bars do not: those coordinate only serve to align bars within a stack. This duality not only undermines the visual interpretation of the map, it also undermines efforts to analyze the data within. To address these problems, the package includes two versions of unstacked data: \code{fatalities.unstacked} assigns the same x-y coordinate to all bars in a stack, that of the case at the base of a stack; \code{fatalities.address} changes the unit of observation from the case to the address, again the x-y coordinate of the case at a stack's base, and turns the number of bars in the stack into an attribute of the address. The \code{vignette}("unstacking.fatalities") describes the unstacking algorithm.

Appending the actual names of roads to the streets data is helpful not only for further corrections or amendments, but can also aid further analysis. These data are available in \code{roads}. See the \code{vignette}("road.names") for details.

To help assess the effect of the removal of the handle from the Broad Street pump (08 September 1854), two sets of time series data (one derived from the "official" statistics included in the Vestry Report, one included as part of Snow's contribution to the report) are included.

Three types of pump neighborhoods are includes: Snow's Broad Street pump neighborhood and 13 neighborhoods computed by Voronoi tessellation and shortest walking path.

According to Dodson and Tobler (1994), the scale of the map is approximately 1:2000, and the coordinate units are meters. By my calculation, one unit on the map is approximately 177 feet. See the \code{vignette}("road.names") for details.
}

\references{
Henry Brody et. al. 2000. "Map-making and myth-making in Broad Street: the London cholera epidemic, 1854." \emph{The Lancet}. 356(July 1): 64-68.

Crane et. al. 1855. \emph{Report On The Cholera Outbreak In The Parish Of St. James, Westminster, During The Autumn Of 1854}. London: John Churchill.

Michael Friendly. 2016. \code{HistData}: Data Sets from the History of Statistics and Data Visualization. R package version 0.7-8. https://CRAN.R-project.org/package=HistData

Steven Johnson. 2006. \emph{The Ghost Map: The Story of London's Most Terrifying Epidemic and How it Changed Science, Cities and the Modern World}. New York: Riverhead Books.

Tom Koch. 2000. \emph{Cartographies of Disease: Maps, Mapping, and Medicine}. Redlands, CA: ESRI Press.

John Snow et. al. 1936. \emph{Snow on Cholera: Being a Reprint of Two Papers}. New York: The Common Wealth Fund.

Waldo Tobler. 1994. \emph{Snow's Cholera Map}. \code{http://www.ncgia.ucsb.edu/pubs/snow/snow.html}. Note that the original data files are no longer available.

Edward R. Tufte. 1983. \emph{The Visual Display of Quantitative Information}. Cheshire, CT: Graphics Press.

Edward R. Tufte. 1997. \emph{Visual Explanations: Images and Quantities, Evidence and Narrative}. Cheshire, CT: Graphics Press.
}

\seealso{
  \code{library}(HistData)

  \code{\link[cholera]{fatalities}}

  \code{\link[cholera]{fatalities.address}}

  \code{\link[cholera]{fatalities.unstacked}}

  \code{\link[cholera]{roads}}

  \code{\link[cholera]{time.series.snow}}

  \code{\link[cholera]{time.series.vestry}}

  \code{\link[cholera]{caseLocator}}

  \code{\link[cholera]{streetNameLocator}}

  \code{\link[cholera]{streetNumberLocator}}

  \code{vignette("duplicate.missing.cases")}

  \code{vignette("road.names")}

  \code{vignette("unstacking.fatalities")}
}

\keyword{datasets}
\keyword{spatial}
