% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifierAudit.R
\name{classifierAudit}
\alias{classifierAudit}
\title{Test if observed case "belongs" to segment.}
\usage{
classifierAudit(case = 483, segment = "326-2")
}
\arguments{
\item{case}{Numeric or Integer. Numeric ID of observed case.}

\item{segment}{Character. Segment ID. See cholera::road.segments}
}
\value{
Logical TRUE or FALSE
}
\description{
Uses orthogonal projector from an observed case to see if it intersects a segment.
}
\note{
This is a diagnostics function. It is not a guarantee of correct classification.
}
