% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unstack.R
\name{classifierAuditPlot}
\alias{classifierAuditPlot}
\title{Plot whether case is orthogonal to segment.}
\usage{
classifierAuditPlot(case = 483, segment = "326-2", radius = 0.1)
}
\arguments{
\item{case}{Numeric or Integer. Numeric ID of (anchor) case.}

\item{segment}{Character. Segment ID from cholera::road.segments}

\item{radius}{Numeric. Controls the degree of zoom.}
}
\value{
A base R graphic.
}
\description{
Plot whether case is orthogonal to segment.
}
\note{
segment "326-2" is part of Little Windmill Street. A diagnostics; not a guarantee of correct classification.
}
