% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliffordStreet.R
\name{cliffordStreet}
\alias{cliffordStreet}
\title{Estimate native coordinates for missing Clifford Street segment.}
\usage{
cliffordStreet()
}
\description{
Segment between Old Burlington and Saville Row.
}
