% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deldirPolygons.R
\name{deldirPolygons}
\alias{deldirPolygons}
\title{Extract vertices of Delauny triangles and Dirichelet (Voronoi) tiles.}
\usage{
deldirPolygons(sites, rw.data = NULL, rw = NULL, type = "tiles")
}
\arguments{
\item{sites}{Object. Data frame of sites to compute Delauny triangulation and Dirichelet (Voronoi) tessellation with variables "x" and "y".}

\item{rw.data}{Object. Data frame of secondary source of data to set the rectangular window or bounding box: observations, cases, etc. with variables "x" and "y".}

\item{rw}{Numeric. Alternative to rw.data: vector of corners to define the rectangular window or bounding box: xmin, xmax, ymin, ymax. For \code{deldir::deldir()}.}

\item{type}{Character. "tiles" (tessellation) or "triangles" (triangulation) vertices.}
}
\description{
For use with functions that use polygons.
}
\note{
This function relies on the 'deldir' package.
}
\examples{
polygon.vertices <- deldirPolygons(cholera::pumps)
cholera::snowMap()
invisible(lapply(polygon.vertices, polygon))

polygon.vertices <- deldirPolygons(cholera::pumps, cholera::roads)
cholera::snowMap()
invisible(lapply(polygon.vertices, polygon))

polygon.vertices <- deldirPolygons(cholera::pumps, cholera::roads,
 type = "triangles")
cholera::snowMap()
invisible(lapply(polygon.vertices, polygon))
}
