% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceTime.R
\name{distanceTime}
\alias{distanceTime}
\title{Convert distance to time.}
\usage{
distanceTime(x, unit = "minute", meter.unit = 54, speed = 3)
}
\arguments{
\item{x}{Numeric. Nominal map distance.}

\item{unit}{Character. Unit of measurement: "hour" or "minute".}

\item{meter.unit}{Numeric. Estimate of meters per map unit: 54.}

\item{speed}{Numeric. Walking speed: 3 meters per hour.}
}
\value{
An R vector.
}
\description{
Convert distance to time.
}
