% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclideanDistance2.R
\name{euclideanDistance2}
\alias{euclideanDistance2}
\title{Compute Euclidean distance between pumps or between cases.}
\usage{
euclideanDistance2(origin, destination, vestry = FALSE, unit = NULL,
  type = "pumps")
}
\arguments{
\item{origin}{Numeric or Integer. Numeric ID of origin.}

\item{destination}{Numeric or Integer. Numeric ID of destination.}

\item{vestry}{Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 in the original map.}

\item{unit}{Character. Unit of measurement: "meter" or "yard". Default is NULL, which returns the map's native scale. See \code{vignette("roads")} for information on unit distances.}

\item{type}{Character "cases" or "pumps"}
}
\value{
An R vector.
}
\description{
For case to pump(s), use \code{\link{euclideanDistance}}.
}
\note{
For "type = "cases"", "origin" and "destination" must be a number between 1 and 578. To compute distance, the function uses the coordinates of the case's "address" (i.e., its "anchor case"). For type = "pumps", "origin" and "destination" must be numbers between 1 and 14 for vestry = TRUE; for vestry = FALSE
}
\examples{
# Pairwise Euclidean distance between pumps.

# pairs <- combn(cholera::pumps$id, 2, simplify = FALSE)
#
# vapply(pairs, function(x) {
#   euclideanDistance2(x[1], x[2])
# }, numeric(1L))
}
