% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclideanPathB.R
\name{euclideanPathB}
\alias{euclideanPathB}
\title{Compute path of the Euclidean distance between cases and/or pumps (Beta)}
\usage{
euclideanPathB(origin = 1, destination = NULL, type = "case-pump",
  observed = TRUE, vestry = FALSE, unit = "meter",
  time.unit = "second", walking.speed = 5)
}
\arguments{
\item{origin}{Numeric or Character. Numeric ID of case or pump. Character landmark name.}

\item{destination}{Numeric or Character. Numeric ID(s) of case(s) or pump(s). Exclusion is possible via negative selection (e.g., -7). Default is \code{NULL}, which returns closest pump or "anchor" case. Character landmark name.}

\item{type}{Character "case-pump", "cases" or "pumps".}

\item{observed}{Logical. Use observed or "simulated" expected data.}

\item{vestry}{Logical. \code{TRUE} uses the 14 pumps from the Vestry Report. \code{FALSE} uses the 13 pumps from the original map.}

\item{unit}{Character. Unit of distance: "meter", "yard" or "native". "native" returns the map's native scale. See \code{vignette("roads")} for information on unit distances.}

\item{time.unit}{Character. "hour", "minute", or "second".}

\item{walking.speed}{Numeric. Default is 5 km/hr.}
}
\value{
An R list with 3 data frames: x-y coordinates for the origin and destination, and a summary of results.
}
\description{
Compute path of the Euclidean distance between cases and/or pumps (Beta)
}
\note{
The function uses a case's "address" (i.e., "anchor" case of a stack) to compute distance. Time is computed using \code{distanceTime()}.
}
