% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearsonResiduals.R
\name{expectedWalkingLength}
\alias{expectedWalkingLength}
\title{Compute Total Length of Roads by Neighborhood (prototype)}
\usage{
expectedWalkingLength(x)
}
\arguments{
\item{x}{An object created by \code{neighborhoodWalking()}.}
}
\description{
Compute Total Length of Roads by Neighborhood (prototype)
}
