% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fatalities.address}
\alias{fatalities.address}
\title{"Unstacked" Dodson and Tobler's cholera data: address as unit of observation.}
\format{A data frame with 4 variables for 321 addresses
 \describe{
    \item{\code{anchor.case}}{numerical case ID of address}
    \item{\code{x}}{x-coordinate}
    \item{\code{y}}{y-coordinate}
    \item{\code{case.count}}{number of fatalities at address}
}}
\usage{
fatalities.address
}
\description{
An "unstacked" version of the \code{fatalities} dataset. It changes the unit of observation from the case (bar) to the "address", the x-y coordinate of the case at the base of a stack, and makes the number of fatalities an attribute of the "address".
}
\section{Notes}{
 \code{\link[cholera]{unstackFatalities}} documents code for these data.
}

\seealso{
\code{vignette}("duplicate.missing.cases")

\code{vignette}("unstacking.fatalities")

\code{\link[cholera]{caseLocator}}

\code{\link[cholera]{streetNameLocator}}

\code{\link[cholera]{streetNumberLocator}}
}
\keyword{datasets}
