% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fatalities.address2}
\alias{fatalities.address2}
\title{"Unstacked" amended cholera data with address as unit of observation.}
\format{A data frame with 4 variables for 321 addresses
 \describe{
    \item{\code{anchor}}{numerical case ID of address}
    \item{\code{x}}{x-coordinate}
    \item{\code{y}}{y-coordinate}
    \item{\code{case.count}}{number of fatalities at address}
    \item{\code{longitude}}{x-coordinate}
    \item{\code{latitude}}{y-coordinate}
}}
\usage{
fatalities.address2
}
\description{
An "unstacked" version of the \code{fatalities} dataset. It changes the unit of observation from the case (bar) to the "address", the x-y coordinates of the case at the base of a stack, and makes the number of fatalities an attribute of the "address".
}
\note{
sel <- fatalities2$case %in% fatalities.address2$anchor

identical(fatalities2[sel, "longitude"], fatalities.address2$longitude)

identical(fatalities2[sel, "latitude"], fatalities.address2$latitude)
}
\seealso{
\code{\link{caseLocator}}

\code{\link{streetNameLocator}}

\code{\link{streetNumberLocator}}
}
\keyword{datasets}
