% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fatalities}
\alias{fatalities}
\title{Amended Dodson and Tobler's cholera data.}
\format{A data frame with 3 variable that records the position and the nearest pump for the 578 bars on Snow's map.

 \describe{
    \item{\code{case}}{numeric case ID}
    \item{\code{x}}{x-coordinate}
    \item{\code{y}}{y-coordinate}
}}
\usage{
fatalities
}
\description{
An amended version of Dodson and Tobler's digitization of John Snow's map of the 1854 London cholera outbreak. It removes 3 duplicate observations and imputes the location for 3 "missing" observation. This information is also available in HistData::Snow.deaths2 (>= ver. 0.7-8).
}
\seealso{
\code{vignette}("duplicate.missing.cases")

\code{\link{fatalityLocator}}

\code{\link[cholera]{streetNameViewer}}

\code{\link[cholera]{streetNumberViewer}}
}
\keyword{datasets}

