% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fatalities.unstacked}
\alias{fatalities.unstacked}
\title{Amended fatalities data with all cases in a stack sharing the coordinates of the "anchor" case.}
\format{
A data frame with 3 variable that records the position of the 578 bars on Snow's map.

 \describe{
    \item{\code{case}}{numerical case ID}
    \item{\code{x}}{x-coordinate}
    \item{\code{y}}{y-coordinate}
    \item{\code{lon}}{longitude}
    \item{\code{lat}}{latitude}
}
}
\usage{
fatalities.unstacked
}
\description{
The "anchor" case is the case at the base of a stack.
}
\note{
\code{\link{unstackFatalities}} documents the code for these data. For details, see \code{vignette}("unstacking.fatalities").
}
\seealso{
\code{\link{caseLocator}}

\code{\link{streetNameLocator}}

\code{\link{streetNumberLocator}}
}
\keyword{datasets}
