% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatalityLocator.R
\name{fatalityLocator}
\alias{fatalityLocator}
\title{Locate fatality by numerical case ID.}
\usage{
fatalityLocator(case, zoom = FALSE, radius = 2, stacked = TRUE)
}
\arguments{
\item{case}{Whole number between 1 and 578. Need not be of integer type.}

\item{zoom}{Logical.}

\item{radius}{Numeric. Controls the degree of zoom. Default value is 2.}

\item{stacked}{Logical. TRUE uses \code{fatalities} ("stacked"); FALSE uses \code{fatalities.address} ("unstacked").}
}
\value{
A base R graphics plot.
}
\description{
Plots John Snow's map of the 1854 London cholera outbreak and highlights the
selected case.
}
\examples{
fatalityLocator(290)
fatalityLocator(290, zoom = TRUE)
fatalityLocator(290, stacked = FALSE)
fatalityLocator(290, zoom = TRUE, stacked = FALSE)
}
\seealso{
\code{fatalities}

\code{fatalities.address}

\url{http://r.789695.n4.nabble.com/no-visible-binding-for-global-variable-for-data-sets-in-a-package-td4696053.html}
}

