% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unstack.R
\name{fixFatalities}
\alias{fixFatalities}
\title{Fix apparent coding error in Dodson and Tobler's digitization of Snow's cholera data.}
\usage{
fixFatalities()
}
\value{
An R dataframe.
}
\description{
Fixes two apparent coding errors using 3 misplaced cases in Dodson and Tobler's data.
}
