% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{frame.data}
\alias{frame.data}
\title{Map frame data c("x", "y") and c("lon", "lat").}
\format{
A data frame with 106 observations (points) and 8 variables.
 \describe{
    \item{\code{street}}{street number}
    \item{\code{n}}{street street component number}
    \item{\code{x}}{native x-coordinate}
    \item{\code{y}}{native y-coordinate}
    \item{\code{id}}{segment numeric ID}
    \item{\code{name}}{street name}
    \item{\code{lon}}{longitude}
    \item{\code{lat}}{latitude}
 }
}
\usage{
frame.data
}
\description{
Map frame data c("x", "y") and c("lon", "lat").
}
\keyword{datasets}
