% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{frame.sample}
\alias{frame.sample}
\title{Sample of map framework points (segment endpoints).}
\format{
A vector of length 106 of cholera::roads$id
 \describe{
    \item{\code{frame.sample}}{cholera::roads$id }
 }
}
\usage{
frame.sample
}
\description{
Sample of map framework points (segment endpoints).
}
\keyword{datasets}
