% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesicMeters.R
\name{geodesicMeters}
\alias{geodesicMeters}
\title{Decompose geodesic distances into horizontal and vertical components.}
\usage{
geodesicMeters(dat = cholera::pumps, case.address = FALSE)
}
\arguments{
\item{dat}{Object. Data.}

\item{case.address}{Logical. Use fatalities.address$anchor}
}
\description{
Compute geodesic distance from origin to pump and decompose result into
}
