% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoLines.R
\name{isoLines}
\alias{isoLines}
\title{Plot isochrone and isodistance regions (prototype)}
\usage{
isoLines(pump.select = 7, post = 50, post.type = "distance",
  palette = "Spectral")
}
\arguments{
\item{pump.select}{Numeric.}

\item{post}{Numeric.}

\item{post.type}{Character. "distance or "time".}

\item{palette}{Character. RColorBrewer palette.}
}
\description{
Plot isochrone and isodistance regions (prototype)
}
