% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoVertices.R
\name{isoVertices}
\alias{isoVertices}
\title{isochrone and isodistance vertices (prototype)}
\usage{
isoVertices(post = 50, post.type = "distance", multi.core = FALSE)
}
\arguments{
\item{post}{Numeric.}

\item{post.type}{Character. "distance or "time".}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. On Windows, only \code{multi.core = FALSE} is available.}
}
\description{
isochrone and isodistance vertices (prototype)
}
