% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongAuxilliaryFunctions.R
\name{kmeansRectanlge}
\alias{kmeansRectanlge}
\title{Compute rectangle vertices (prototype).}
\usage{
kmeansRectanlge(x)
}
\arguments{
\item{x}{Object. Points/pixel count.}
}
\description{
Compute rectangle vertices (prototype).
}
