% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landmark.squares}
\alias{landmark.squares}
\title{Centers of city squares.}
\format{
A data frame with 2 observations and 6 variables that records the position of landmark square labels.
 \describe{
    \item{\code{case}}{numeric case ID}
    \item{\code{x}}{x-coordinate}
    \item{\code{y}}{y-coordinate}
    \item{\code{lon}}{longitude}
    \item{\code{lat}}{latitude}
    \item{\code{name}}{square name}
 }
}
\usage{
landmark.squares
}
\description{
Centers of city squares.
}
\keyword{datasets}
