% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongLandmarks.R
\name{landmarksPDF}
\alias{landmarksPDF}
\title{Create PDFs of landmark addreses.}
\usage{
landmarksPDF(path, pch = 15, cex = 0.2)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/"}

\item{pch}{Integer. R pch.}

\item{cex}{Numeric.}
}
\description{
For QGIS geo-referencing.
}
