% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latitudeLongitude.R
\name{latitudeLongitude}
\alias{latitudeLongitude}
\title{Compute latitude and longitude from geo TIFF (in progress).}
\usage{
latitudeLongitude(geotiff, dataset = "roads", cutpoint = 6e-06,
  multi.core = TRUE)
}
\arguments{
\item{geotiff}{Object. GeoTIFF file.}

\item{dataset}{Object. data set or layer.}

\item{cutpoint}{Numeric. Cutpoint for agglomerative hierarchical cluster analysis.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. See \code{vignette("Parallelization")} for details.}
}
\description{
Georeferenced TIFF file from QGIS3.10.
}
\note{
This function documents the code that generates "roads2" latitude and longitude estimates.
}
