% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latitudeLongitudeFatality.R
\name{latitudeLongitudeFatality}
\alias{latitudeLongitudeFatality}
\title{Compute latitude and longitude for fatalities cases (prototype).}
\usage{
latitudeLongitudeFatality(path, multi.core = TRUE)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/"}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. See \code{vignette("Parallelization")} for details.}
}
\value{
An R data frame.
}
\description{
Non-anchor points: in stack, not at base of stack.
}
