% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latlong.fatalities.address}
\alias{latlong.fatalities.address}
\title{"Unstacked" amended cholera data with address as unit of observation.}
\format{
A data frame with 6 variables for 321 addresses
 \describe{
    \item{\code{anchor}}{numerical case ID of address}
    \item{\code{x}}{x-coordinate}
    \item{\code{y}}{y-coordinate}
    \item{\code{case.count}}{number of fatalities at address}
    \item{\code{lon}}{longitude}
    \item{\code{lat}}{latitude}
}
}
\usage{
latlong.fatalities.address
}
\description{
"Unstacked" amended cholera data with address as unit of observation.
}
\keyword{datasets}
