% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latlong.frame}
\alias{latlong.frame}
\title{"Map Frame" data with and georeferenced longitude and latitude.}
\format{
A data frame with 106 observations and 8 variables. The data describe the roads on Snow's map.
\describe{
  \item{\code{street}}{street segment number, which range between 1 and 528}
  \item{\code{n}}{number of points in this street line segment}
  \item{\code{x}}{x-coordinate}
  \item{\code{y}}{y-coordinate}
  \item{\code{id}}{unique numeric ID}
  \item{\code{name}}{road name}
  \item{\code{lon}}{longitude}
  \item{\code{lat}}{latitude}
}
}
\usage{
latlong.frame
}
\description{
"Map Frame" data with and georeferenced longitude and latitude.
}
\keyword{datasets}
