% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latlong.nearest.pump}
\alias{latlong.nearest.pump}
\title{Walking distance to Nearest Pump (latlong prototype).}
\format{
A data frames with 4 variables.
\describe{
  \item{\code{case}}{case ID}
  \item{\code{pump}}{pump ID}
  \item{\code{distance}}{walking distance in meters}
  \item{\code{time}}{walking time in seconds based on 5 km/hr walking speed}
}
}
\usage{
latlong.nearest.pump
}
\description{
Walking distance to Nearest Pump (latlong prototype).
}
\keyword{datasets}
