% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latlong.ortho.addr}
\alias{latlong.ortho.addr}
\title{Orthogonal projection of observed address (latlong) cases onto road network.}
\format{
A data frame with 7 variables that records the position of the orthogonal projection of the 321 cases onto the network of roads.
 \describe{
    \item{\code{road.segment}}{"address" road segment}
    \item{\code{x.proj}}{x-coordinate}
    \item{\code{y.proj}}{y-coordinate}
    \item{\code{ortho.dist}}{orthogonal distance to home road segment}
    \item{\code{case}}{numeric case ID}
    \item{\code{lon}}{longitude}
    \item{\code{lat}}{latitude}
 }
}
\usage{
latlong.ortho.addr
}
\description{
Orthogonal projection of observed address (latlong) cases onto road network.
}
\note{
\code{\link{unstackFatalities}} documents the code for these data.
}
\keyword{datasets}
