% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latlong.ortho.proj.pump}
\alias{latlong.ortho.proj.pump}
\title{Orthogonal projection of georeferenced 13 original pumps.}
\format{
A data frame with 5 variables that records the position of the orthogonal projection of the 13 original pumps onto the network of roads.
 \describe{
    \item{\code{road.segment}}{"address" road segment}
    \item{\code{lon.proj}}{longitude}
    \item{\code{lat.proj}}{latitude}
    \item{\code{ortho.dist}}{orthogonal distance to home road segment}
    \item{\code{pump}}{numeric ID}
 }
}
\usage{
latlong.ortho.proj.pump
}
\description{
Orthogonal projection of georeferenced 13 original pumps.
}
\note{
\code{\link{pumpData}} documents the code for these data.
}
\keyword{datasets}
