% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latlong.ortho.proj}
\alias{latlong.ortho.proj}
\title{Orthogonal projection of georeferenced cases onto georeferenced road network.}
\format{
A data frame with 5 variables that records the position of the orthogonal projection of the 578 cases onto the network of roads.
 \describe{
    \item{\code{road.segment}}{"address" road segment}
    \item{\code{lon.proj}}{longitude}
    \item{\code{lat.proj}}{latitude}
    \item{\code{ortho.dist}}{orthogonal distance to home road segment}
    \item{\code{case}}{numeric case ID}
 }
}
\usage{
latlong.ortho.proj
}
\description{
Orthogonal projection of georeferenced cases onto georeferenced road network.
}
\note{
\code{\link{unstackFatalities}} documents the code for these data.
}
\keyword{datasets}
