% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latlong.ortho.pump}
\alias{latlong.ortho.pump}
\title{Orthogonal projection of 13 original pumps (latlong).}
\format{
A data frame with 7 variables that records the position of the orthogonal projection of the 13 original pumps onto the network of roads.
 \describe{
    \item{\code{road.segment}}{"address" road segment}
    \item{\code{x.proj}}{x-coordinate}
    \item{\code{y.proj}}{y-coordinate}
    \item{\code{d}}{distance to home road segment}
    \item{\code{type}}{orthogonal or proximate (Euclidean) distance}
    \item{\code{id}}{numeric ID}
    \item{\code{lon}}{longitude}
    \item{\code{lat}}{latitude}
 }
}
\usage{
latlong.ortho.pump
}
\description{
Orthogonal projection of 13 original pumps (latlong).
}
\note{
\code{\link{pumpData}} documents the code for these data.
}
\keyword{datasets}
