% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latlong.pumps}
\alias{latlong.pumps}
\title{Dodson and Tobler's pump data with street name.}
\format{
A data frame with 13 observations and 6 variables that describe the pumps on Snow's map.
\describe{
  \item{\code{id}}{pump number between 1 and 13}
  \item{\code{street}}{nearest street}
  \item{\code{x}}{x-coordinate}
  \item{\code{y}}{y-coordinate}
  \item{\code{lon}}{longitude}
  \item{\code{lat}}{latitude}
}
}
\usage{
latlong.pumps
}
\description{
Dodson and Tobler's pump data with street name.
}
\keyword{datasets}
