% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latlong.road.segments}
\alias{latlong.road.segments}
\title{Georeferenced Dodson and Tobler's road segments data.}
\format{
A data frame with 657 observations and 8 variables. The data describe the straight line segments used to recreate the roads on Snow's map.
\describe{
  \item{\code{street}}{numeric street ID, which range between 1 and 528}
  \item{\code{id}}{character segment ID}
  \item{\code{name}}{road name}
  \item{\code{lon1}}{longitude of first endpoint}
  \item{\code{lat1}}{latitude of first endpoint}
  \item{\code{lon2}}{longitude of second endpoint}
  \item{\code{lat2}}{latitude of second endpoint}
  \item{\code{distance}}{distance (meters)}
}
}
\usage{
latlong.road.segments
}
\description{
This data set transforms Dodson and Tobler's street data to give each straight line segment of a "road" a unique ID.
}
\keyword{datasets}
