% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latlong.sim.ortho.proj}
\alias{latlong.sim.ortho.proj}
\title{Road "address" of simulated (i.e., "expected") cases (latlong).}
\format{
A data frame with 8 variables that records the "address" of 19,993 regularly spaced simulated Cartesian/geodesic cases regularly spaced across map.
\describe{
    \item{\code{case}}{numeric case ID}
    \item{\code{road.segment}}{"address" road segment}
    \item{\code{x.proj}}{x-coordinate}
    \item{\code{y.proj}}{y-coordinate}
    \item{\code{dist}}{Euclidean or orthogonal distance to home road segment}
    \item{\code{type}}{type of projection: Euclidean ("eucl") or orthogonal ("ortho")}
    \item{\code{lon}}{longitude}
    \item{\code{lat}}{latitude}
 }
}
\usage{
latlong.sim.ortho.proj
}
\description{
Road "address" of simulated (i.e., "expected") cases (latlong).
}
\keyword{datasets}
