% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongAddress.R
\name{latlongAddress}
\alias{latlongAddress}
\title{Compute latitude and longitude of case "addresses" (prototype).}
\usage{
latlongAddress(path, multi.core = TRUE)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/"}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. See \code{vignette("Parallelization")} for details.}
}
\value{
An R data frame.
}
\description{
Compute latitude and longitude of case "addresses" (prototype).
}
\note{
This documents the computation of the latlong version of the fatalities.address data frame.
}
