% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongCoordinates.R
\name{latlongCoordinates}
\alias{latlongCoordinates}
\title{Compute Georeferenced Latitude and Longitude (prototype).}
\usage{
latlongCoordinates(tif, cutpoint = 1e-04)
}
\arguments{
\item{tif}{Character. Georeferenced QGIS TIFF file.}

\item{cutpoint}{Numeric. Cutpoint for hierarchical cluster analysis.}
}
\description{
Compute Georeferenced Latitude and Longitude (prototype).
}
