% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongCoordinatesC.R
\name{latlongCoordinatesC}
\alias{latlongCoordinatesC}
\title{Compute Georeferenced Latitude and Longitude (prototype).}
\usage{
latlongCoordinatesC(tif, k, path)
}
\arguments{
\item{tif}{Character. Georeferenced QGIS TIFF file.}

\item{k}{Numeric. Number of clusters, k, to identify.}

\item{path}{Character. e.g., "~/Documents/Data/".}
}
\description{
Compute Georeferenced Latitude and Longitude (prototype).
}
