% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongEmbed.R
\name{latlongEmbed}
\alias{latlongEmbed}
\title{Embed anchors and pumps into road segments (prototype).}
\usage{
latlongEmbed(path, tif, vestry = FALSE)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/"}

\item{tif}{Character. Georeferenced QGIS TIFF file.}

\item{vestry}{Logical.}
}
\description{
Embed anchors and pumps into road segments (prototype).
}
