% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongFrame.R
\name{latlongFramePDF}
\alias{latlongFramePDF}
\title{Create PDFs of frame points.}
\usage{
latlongFramePDF(path, pch = 46, cex = 1)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/"}

\item{pch}{Integer. R pch.}

\item{cex}{Numeric.}
}
\description{
For QGIS geo-referencing.
}
