% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongLandmarks.R
\name{latlongLandmarks}
\alias{latlongLandmarks}
\title{Compute Georeferenced Latitude and Longitude (prototype).}
\usage{
latlongLandmarks(path)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/"}
}
\description{
Compute Georeferenced Latitude and Longitude (prototype).
}
