% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongNeighborhoodDataB.R
\name{latlongNeighborhoodDataB}
\alias{latlongNeighborhoodDataB}
\title{Compute network graph of roads, cases and pumps.}
\usage{
latlongNeighborhoodDataB(vestry = FALSE)
}
\arguments{
\item{vestry}{Logical.}
}
\description{
Assembles cases, pumps and road into a network graph.
}
