% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongOrthoAddress.R
\name{latlongOrthoAddress}
\alias{latlongOrthoAddress}
\title{Compute latitude and longitude for orthogonal case projection (address).}
\usage{
latlongOrthoAddress(path, vestry = FALSE, multi.core = TRUE)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/".}

\item{vestry}{Logical.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. See \code{vignette("Parallelization")} for details.}
}
\value{
An R data frame.
}
\description{
Compute latitude and longitude for orthogonal case projection (address).
}
