% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongOrthoAddressB.R
\name{latlongOrthoAddressB}
\alias{latlongOrthoAddressB}
\title{Compute latitude and longitude for orthogonal case projection (address).}
\usage{
latlongOrthoAddressB(multi.core = TRUE)
}
\arguments{
\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. See \code{vignette("Parallelization")} for details.}
}
\value{
An R data frame.
}
\description{
Compute latitude and longitude for orthogonal case projection (address).
}
