% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongOrthoProj.R
\name{latlongOrthoProj}
\alias{latlongOrthoProj}
\title{Compute orthogonal projection of fatality addresses (prototype).}
\usage{
latlongOrthoProj(path, multi.core = TRUE, radius = 5e-04)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/"}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. See \code{vignette("Parallelization")} for details.}

\item{radius}{Numeric.}
}
\value{
An R data frame.
}
\description{
Projection from fatality address to nearest road segment.
}
