% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongOrthoProjPumps.R
\name{latlongOrthoProjPumps}
\alias{latlongOrthoProjPumps}
\title{Compute orthogonal projection of pumps (prototype).}
\usage{
latlongOrthoProjPumps(path, vestry = FALSE, radius = 0.001, multi.core = TRUE)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/"}

\item{vestry}{Logical.}

\item{radius}{Numeric.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. See \code{vignette("Parallelization")} for details.}
}
\description{
Projection from fatality address to nearest road segment.
}
