% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongOrthoPump.R
\name{latlongOrthoPump}
\alias{latlongOrthoPump}
\title{Compute orthogonal projection of pumps (prototype).}
\usage{
latlongOrthoPump(path, vestry = FALSE)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/"}

\item{vestry}{Logical.}
}
\description{
Projection from fatality address to nearest road segment.
}
