% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongOrthoPumpB.R
\name{latlongOrthoPumpB}
\alias{latlongOrthoPumpB}
\title{Compute pump coordinates.}
\usage{
latlongOrthoPumpB(vestry = FALSE, multi.core = TRUE)
}
\arguments{
\item{vestry}{Logical. \code{TRUE} uses the 14 pumps from the Vestry report. \code{FALSE} uses the 13 in the original map.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. With Numeric, you specify the number logical cores (rounds with \code{as.integer()}). See \code{vignette("Parallelization")} for details.}
}
\description{
Returns either the set of x-y coordinates for the pumps themselves or for their orthogonally projected "addresses" on the network of roads.
}
