% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthoRectangle.R
\name{latlongOrthoRect}
\alias{latlongOrthoRect}
\title{Compute latitude and longitude for georeferenced orthogonal rectangle.}
\usage{
latlongOrthoRect(path, tif, k = 4, multi.core = TRUE)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/".}

\item{tif}{Character. Georeferenced QGIS TIFF file.}

\item{k}{Numeric. Number of clusters, k, to identify.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. See \code{vignette("Parallelization")} for details.}
}
\description{
Uses georeferenced TIF from QGIS.
}
