% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongPumps.R
\name{latlongPumps}
\alias{latlongPumps}
\title{Compute Georeferenced Latitude and Longitude (prototype).}
\usage{
latlongPumps(tif, cutpoint = 0.001, vestry = FALSE)
}
\arguments{
\item{tif}{Character. Georeferenced QGIS TIFF file.}

\item{cutpoint}{Numeric. Cutpoint for hierarchical cluster analysis.}

\item{vestry}{Logical.}
}
\description{
Compute Georeferenced Latitude and Longitude (prototype).
}
