% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongRoads.R
\name{latlongRoads}
\alias{latlongRoads}
\title{Compute latitude and longitude for unique road segment endpoints (prototype).}
\usage{
latlongRoads(path, multi.core = TRUE)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/"}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. See \code{vignette("Parallelization")} for details.}
}
\value{
An R data frame.
}
\description{
Compute latitude and longitude for unique road segment endpoints (prototype).
}
\note{
This documents the computation of the lat-long version of the roads data frame.
}
