% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongRoadSegments.R
\name{latlongRoadSegments}
\alias{latlongRoadSegments}
\title{Reshape 'roads' data frame into 'road.segments' data frame.}
\usage{
latlongRoadSegments(path)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/"}
}
\value{
An R data frame.
}
\description{
Used to integrate pumps and cases into road network when computing walking neighborhoods.
}
\note{
This function documents the code that generates \code{\link[cholera]{road.segments}}.
}
