% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongSegmentLocator.R
\name{latlongSegmentLocator}
\alias{latlongSegmentLocator}
\title{Locate road segment by ID}
\usage{
latlongSegmentLocator(segment.id = "216-1", zoom = TRUE,
  cases = "address", token = "id", add.title = TRUE,
  add.subtitle = TRUE, add.pump = TRUE, vestry = FALSE,
  highlight = TRUE, distance.unit = "meter", time.unit = "minute",
  walking.speed = 5)
}
\arguments{
\item{segment.id}{Character vector. Note that \code{streetNameLocator}() tries to correct for case and to remove extra spaces.}

\item{zoom}{Logical.}

\item{cases}{Character. Plot cases: \code{NULL}, "address" or "fatality".}

\item{token}{Character. "id" or "point".}

\item{add.title}{Logical. Include title.}

\item{add.subtitle}{Logical. Include subtitle with road information.}

\item{add.pump}{Logical. Include nearby pumps.}

\item{vestry}{Logical. \code{TRUE} uses the 14 pumps from the Vestry report. \code{FALSE} uses the 13 in the original map.}

\item{highlight}{Logical. Highlight selected road and its cases.}

\item{distance.unit}{Character. Unit of distance: "meter", "yard" or "native". "native" returns the map's native scale. See \code{vignette("roads")} for information on conversion.}

\item{time.unit}{Character. "hour", "minute", or "second".}

\item{walking.speed}{Numeric. Walking speed in km/hr.}
}
\value{
A base R graphics plot.
}
\description{
Highlight road segment and its cases.
}
