% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongVoronoi.R
\name{latlongVoronoi}
\alias{latlongVoronoi}
\title{Compute Georeferenced Latitude and Longitude of vertices of Voronoi polygons.}
\usage{
latlongVoronoi(path, vestry = FALSE, multi.core = TRUE)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/"}

\item{vestry}{Logical. \code{TRUE} uses the 14 pumps from the Vestry Report. \code{FALSE} uses the 13 in the original map.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. See \code{vignette("Parallelization")} for details.}
}
\description{
Compute Georeferenced Latitude and Longitude of vertices of Voronoi polygons.
}
