% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongWalkingPathB.R
\name{latlongWalkingPathB}
\alias{latlongWalkingPathB}
\title{Plot walking path to nearest pump (prototype).}
\usage{
latlongWalkingPathB(case = 1, zoom = TRUE, vestry = FALSE)
}
\arguments{
\item{case}{Numeric.}

\item{zoom}{Logical or Numeric. A numeric value >= 0 that controls the degree of zoom.}

\item{vestry}{Logical. \code{TRUE} uses the 14 pumps from the map in the Vestry Report. \code{FALSE} uses the 13 pumps from the original map.}
}
\description{
Plot walking path to nearest pump (prototype).
}
