% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latitudeLongitudeFrame.R
\name{mapFrameSamples}
\alias{mapFrameSamples}
\title{Sample for map frame segment endpoints.}
\usage{
mapFrameSamples(cutpoint = 0.1)
}
\arguments{
\item{cutpoint}{Numeric.}
}
\description{
Sample for map frame segment endpoints.
}
