% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowMap.R
\name{mapRange}
\alias{mapRange}
\title{Compute xlim and ylim of Snow's map.}
\usage{
mapRange(latlong = FALSE)
}
\arguments{
\item{latlong}{Logical. Use estimated longitude and latitude.}
}
\description{
Compute xlim and ylim of Snow's map.
}
