% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meter.to.yard}
\alias{meter.to.yard}
\title{Meter to yard conversion factor.}
\format{
A vector of length one.
 \describe{
    \item{\code{meter.to.yard}}{conversion factor = 1.093613}
 }
}
\usage{
meter.to.yard
}
\description{
Meter to yard conversion factor.
}
\keyword{datasets}
