% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milePosts.R
\name{milePosts}
\alias{milePosts}
\title{Compute the coordinates of walking path mileposts.}
\usage{
milePosts(pump.select, milepost.interval = 50)
}
\arguments{
\item{pump.select}{Numeric.}

\item{milepost.interval}{Numeric. Meters.}
}
\value{
An R data frame:
}
\description{
In-progress prottotype.
}
