% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{nearestPump}
\alias{nearestPump}
\title{Compute walking distance from anchor to nearest pump (or among selected pumps).}
\usage{
nearestPump(pump.select = NULL, vestry = FALSE)
}
\arguments{
\item{pump.select}{Numeric. Default is NULL: all pumps are used. Otherwise, selection by a vector of numeric IDs: 1 to 13 for \code{pumps}; 1 to 14 for \code{pumps.vestry}. Negative selection allowed.}

\item{vestry}{Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 in the original map.}
}
\value{
An R data frame.
}
\description{
Compute walking distance from anchor to nearest pump (or among selected pumps).
}
