% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neighborhood.distances}
\alias{neighborhood.distances}
\title{Walking distance for cases in pump neighborhoods.}
\format{A data frame with 4 variables that records the walking distance between each of the 578 cases and its nearest pump.
 \describe{
    \item{\code{case}}{observed case: 1 through 578.}
    \item{\code{anchor}}{anchor case or "address".}
    \item{\code{pump}}{numerical pump ID}
    \item{\code{pump.name}}{pump name (street)}
    \item{\code{distance}}{distance to pump in native map units.}
 }}
\usage{
neighborhood.distances
}
\description{
Walking distance for cases in pump neighborhoods.
}
\section{Notes}{
 \code{\link{neighborhoodDistances}} documents the code for these data.
}

\keyword{datasets}
