% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhood.r
\name{neighborhood}
\alias{neighborhood}
\title{Observed walking path neighborhood data.}
\usage{
neighborhood(selection = NULL, vestry = FALSE)
}
\arguments{
\item{selection}{Numeric. Default is NULL; all pumps are used. Otherwise, selection by a vector of numeric IDs: 1 to 13 for \code{pumps}; 1 to 14 for \code{pumps.vestry}}

\item{vestry}{Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 in the original map.}
}
\description{
Neighborhoods are based on the shortest paths between a fatality's address and its nearest pump.
}

