% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neighborhood.segments}
\alias{neighborhood.segments}
\title{List of data frames of road segments in the original 13 pump neighborhoods.}
\format{A list of 13 data frames. Each data frame has 7 variables that describe the straight line segments in each pump neighborhood.
\describe{
  \item{\code{street}}{numeric street ID, which range between 1 and 528}
  \item{\code{id}}{characger segement ID}
  \item{\code{name}}{road name}
  \item{\code{x1}}{x-coordinate of first endpoint}
  \item{\code{y1}}{y-coordinate of first endpoint}
  \item{\code{x2}}{x-coordinate of second endpoint}
  \item{\code{y2}}{y-coordinate of second endpoint}
}}
\usage{
neighborhood.segments
}
\description{
List of data frames of road segments in the original 13 pump neighborhoods.
}
\keyword{datasets}

