% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neighborhood.segments.sp}
\alias{neighborhood.segments.sp}
\title{List of 13 data frames describing the road segments of "regular" simulated cases grouped by pump neighborhood.}
\format{Each data frame has 8 variables that describe the straight line segments of road segments.
\describe{
  \item{\code{street}}{numeric street ID, which range between 1 and 528}
  \item{\code{id}}{characger segement ID}
  \item{\code{name}}{road name}
  \item{\code{x1}}{x-coordinate of first endpoint}
  \item{\code{y1}}{y-coordinate of first endpoint}
  \item{\code{x2}}{x-coordinate of second endpoint}
  \item{\code{y2}}{y-coordinate of second endpoint}
  \item{\code{trimmed}}{logical indicates whether road segment is trimmed}
}}
\usage{
neighborhood.segments.sp
}
\description{
List of 13 data frames describing the road segments of "regular" simulated cases grouped by pump neighborhood.
}
\keyword{datasets}

