% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoodDistances.R
\name{neighborhoodDistances}
\alias{neighborhoodDistances}
\title{Compute distances within pump neighborhoods.}
\usage{
neighborhoodDistances(metric = "walking", unit = NULL, multi.core = FALSE)
}
\arguments{
\item{metric}{Character. Type of distance: "walking" or "Euclidean".}

\item{unit}{Character. Unit of measurement: "meter" or "yard". Default is NULL, which returns the map's native scale. Meaningful only when "weighted" is TRUE.}

\item{multi.core}{Logical or Numeric. TRUE uses parallel::detectCores(). FALSE uses one, single core. With Numeric, you specify the number logical cores (rounds with as.integer()). On Windows, only "multi.core = FALSE" is available.}
}
\description{
Compute walking or Eucldean distances between the 578 observed cases and their nearest pump.
}
\note{
Documents "neighborhood.distances".
}
