% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoodExpected.R
\name{neighborhoodExpectedCensus}
\alias{neighborhoodExpectedCensus}
\title{Summary statistics for expected walking neighborhoods.}
\usage{
neighborhoodExpectedCensus(selection = NULL, vestry = FALSE,
  weighted = TRUE, multi.core = FALSE)
}
\arguments{
\item{selection}{Numeric. Default is NULL; all pumps are used. Otherwise, selection by a vector of numeric IDs: from 1 to 13 for \code{pumps}; from 1 to 14 for \code{pumps.vestry}}

\item{vestry}{Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 pumps from the original map.}

\item{weighted}{Logical. TRUE uses shortest path weighted by road distance.}

\item{multi.core}{Logical. TRUE uses parallel::mclapply and parallel::detectCores(). Note that this option is not currently available on Windows. See parallel::mclapply() for details. FALSE uses 1 logical core and can be run in GUI and on Windows.}
}
\value{
A data frame.
}
\description{
Neighborhoods are based on the shortest paths between a fatality's address and its nearest pump.
}
\details{
Currently computationally intensive (On an Intel Core i7, streets = TRUE appox. 90 seocnds; streets = FALSE appox. 75 seconds). For better performance, run in parallel with multiple cores in the terminal or in batch mode (appox. 45 and 30 seconds). Note that this is currently unavailable on Windows. See \code{vignette}("walking.neighborhoods") for details.
}
\examples{
# neighborhoodExpectedCensus()
# neighborhoodExpectedCensus(streets = FALSE)
}

