% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoodObserved.R
\name{neighborhoodObservedCensus}
\alias{neighborhoodObservedCensus}
\title{Summary statistics for observed walking neighborhoods.}
\usage{
neighborhoodObservedCensus(selection = NULL, vestry = FALSE,
  weighted = TRUE, statistic = "address")
}
\arguments{
\item{selection}{Numeric. Default is NULL; all pumps are used. Otherwise, selection by a vector of numeric IDs: 1 to 13 for \code{pumps}; 1 to 14 for \code{pumps.vestry}}

\item{vestry}{Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 in the original map.}

\item{weighted}{Logical. TRUE uses shortest path weighted by road distance.}

\item{statistic}{Character. "address" uses the count of address at pumps' locations. "fatality" plots the total count of fatality at pumps' locations.}
}
\value{
A data frame.
}
\description{
Neighborhoods are based on the shortest paths between a fatality's address and its nearest pump.
}
\examples{
# neighborhoodObservedCensus()
# neighborhoodObservedCensus(6:7)
}
\seealso{
\code{vignette}("walking.neighborhoods")
}

