% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi.R
\name{neighborhoodVoronoi}
\alias{neighborhoodVoronoi}
\title{Voronoi neighborhood object.}
\usage{
neighborhoodVoronoi(selection = NULL, vestry = FALSE,
  polygon.vertices = FALSE)
}
\arguments{
\item{selection}{Numeric. Default is NULL: all pumps are used. Ortherwise, selection by a vector of numeric IDs: 1 to 13 for \code{pumps}; 1 to 14 for \code{pumps.vestry}.}

\item{vestry}{Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 in the original map.}

\item{polygon.vertices}{Logical. TRUE returns a list of x-y coordinates of the vertices of Voronoi cells. Useful for sp::point.in.polygon() and used in choleraSummary().}
}
\value{
A list of data and parameters of computed Voronoi neighborhoods. "voronoi" class.
}
\description{
Data for Voronoi tessellation of John Snow's 1854 London cholera data.
}
\examples{
neighborhoodVoronoi()
neighborhoodVoronoi(vestry = TRUE)
neighborhoodVoronoi(selection = 6:7)
neighborhoodVoronoi(selection = -6)
neighborhoodVoronoi(selection = -6, polygon.vertices = TRUE)

# coordinate for vertices also available in object
dat <- neighborhoodVoronoi(selection = -6)
dat$coordinates
}
\seealso{
\code{addVoronoi()}
\code{choleraSummary()}
\code{plot.voronoi()}
}

