% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoodVoronoi.R
\name{neighborhoodVoronoiCensus}
\alias{neighborhoodVoronoiCensus}
\title{Summary statistics for Voronoi neighborhoods.}
\usage{
neighborhoodVoronoiCensus(selection = NULL, vestry = FALSE,
  statistic = "address")
}
\arguments{
\item{selection}{Numeric. Default is NULL: all pumps are used. Ortherwise, selection by a vector of numeric IDs: 1 to 13 for \code{pumps}; 1 to 14 for \code{pumps.vestry}.}

\item{vestry}{Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 in the original map.}

\item{statistic}{Character. "address" plots the total count of addresses at pumps' locations. "fatality" plots the total count of fatalities at pumps' locations.}
}
\value{
A data frame.
}
\description{
Summary statistics for Voronoi neighborhoods.
}
\examples{
neighborhoodVoronoiCensus()
neighborhoodVoronoiCensus(selection = 6:7)
neighborhoodVoronoiCensus(selection = -6)
neighborhoodVoronoiCensus(statistic = "address")
}

