% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkPeriphery.R
\name{networkPeriphery}
\alias{networkPeriphery}
\title{Compute the periphery of walkin path network.}
\usage{
networkPeriphery(pump.select)
}
\arguments{
\item{pump.select}{Numeric.}
}
\value{
An R data frame:
}
\description{
In-progress prottotype.
}
