% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeData.R
\name{nodeData}
\alias{nodeData}
\title{Embed cases and pumps into road network graph.}
\usage{
nodeData(vestry = FALSE)
}
\arguments{
\item{vestry}{Logical. Use Vestry Report pump data.}
}
\value{
An R list of nodes, edges and graph.
}
\description{
Embed cases and pumps into road network graph.
}
