% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nominalNeighborhoodData.R
\name{nominalNeighborhoodData}
\alias{nominalNeighborhoodData}
\title{Compute network graph of roads, cases and pumps (prototype).}
\usage{
nominalNeighborhoodData(vestry = FALSE, case.set = "observed",
  embed.addr = TRUE, embed.landmarks = TRUE, multi.core = TRUE)
}
\arguments{
\item{vestry}{Logical. Use Vestry Report pump data.}

\item{case.set}{Character. "observed", "expected", or "snow". "snow" captures John Snow's annotation of the Broad Street pump neighborhood printed in the Vestry report version of the map.}

\item{embed.addr}{Logical. Embed cases into road network.}

\item{embed.landmarks}{Logical. Embed landmarks into road network.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. See \code{vignette("Parallelization")} for details.}
}
\value{
An R list of nodes, edges and an 'igraph' network graph.
}
\description{
Assembles cases, pumps and road into a network graph.
}
