% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observedPathsPlot.R
\name{observedPathsPlot}
\alias{observedPathsPlot}
\title{Plot observed shortest walking path.}
\usage{
observedPathsPlot(weighted = TRUE, vestry = FALSE, selection = NULL)
}
\arguments{
\item{weighted}{Logical. Shortest path weighted by road distance.}

\item{vestry}{Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 in the original map.}

\item{selection}{Numeric. Default is NULL and all pumps are used. Ortherwise, selection by a vector of numeric IDs: 1 to 13 for \code{pumps}; 1 to 14 for \code{pumps.vestry}.}
}
\value{
A base R graphics plot.
}
\description{
Plots the shortest walking path between observed addressses and their nearest pump.
}
\examples{
observedPathsPlot()
observedPathsPlot(selection = -7) # exclude pump 7
observedPathsPlot(selection = 6)  # only consider pump 6
}
\seealso{
\code{fatalities.address}
}

