% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionAddresses.R
\name{openTriads}
\alias{openTriads}
\title{Rotate, stack and partition open triads.}
\usage{
openTriads(subgraphs, census, census.ct)
}
\arguments{
\item{subgraphs}{Object. 'igraph' list of subgraphs}

\item{census}{Object. List of vertices of subgraphs.}

\item{census.ct}{Object. Count of vertices in subgraphs.}
}
\value{
An R data frame.
}
\description{
Rotate, stack and partition open triads.
}
