% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientationPlot.R
\name{orientationPlot}
\alias{orientationPlot}
\title{Orientation plot.}
\usage{
orientationPlot(pump.id = 6, pump.select = NULL, cases = "orthogonal",
  convex.hull = TRUE, metric = "euclidean", pca.line = FALSE,
  vestry = FALSE)
}
\arguments{
\item{pump.id}{Numeric. Select pump focus.}

\item{pump.select}{Numeric. Pump neighborhoods.}

\item{cases}{Character. "address" or "orthogonal".}

\item{convex.hull}{Logical. Add convex hull.}

\item{metric}{Character. "eucldidean" or "walking".}

\item{pca.line}{Logical. Add PCA (TSS) line through data.}

\item{vestry}{Logical. \code{TRUE} uses the 14 pumps from the Vestry Report. \code{FALSE} uses the 13 in the original map.}
}
\description{
Orientation plot.
}
