% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ortho.proj.landmarks}
\alias{ortho.proj.landmarks}
\title{Orthogonal projection of landmarks onto road network.}
\format{A data frame with 5 variables that records the position of the orthogonal projection of landmarks onto the network of roads.
 \describe{
    \item{\code{road.segment}}{"address" road segment}
    \item{\code{x.proj}}{x-coordinate}
    \item{\code{y.proj}}{y-coordinate}
    \item{\code{ortho.dist}}{orthogonal distance to home road segment}
    \item{\code{case}}{numeric case ID}
 }}
\usage{
ortho.proj.landmarks
}
\description{
Orthogonal projection of landmarks onto road network.
}
\note{
\code{\link{landmarks}} documents the code for these data.
}
\keyword{datasets}
