% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ortho.proj.pump}
\alias{ortho.proj.pump}
\title{Orthogonal projection of 13 original pumps.}
\format{
A data frame with 6 variables that records the position of the orthogonal projection of the 13 original pumps onto the network of roads.
 \describe{
    \item{\code{road.segment}}{"address" road segment}
    \item{\code{x.proj}}{x-coordinate}
    \item{\code{y.proj}}{y-coordinate}
    \item{\code{ortho.dist}}{orthogonal distance to home road segment}
    \item{\code{node}}{node ID}
    \item{\code{pump.id}}{numeric ID}
 }
}
\usage{
ortho.proj.pump
}
\description{
Orthogonal projection of 13 original pumps.
}
\note{
\code{\link{pumpData}} documents the code for these data.
}
\keyword{datasets}
