% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ortho.proj.pump.vestry}
\alias{ortho.proj.pump.vestry}
\title{Orthogonal projection of the 14 pumps from the Vestry Report.}
\format{A data frame with 5 variables that records the position of the orthogonal projection of the 14 pumps onto the network of roads.
 \describe{
    \item{\code{road.segment}}{"address" road segment}
    \item{\code{x.proj}}{x-coordinate}
    \item{\code{y.proj}}{y-coordinate}
    \item{\code{ortho.dist}}{orthogonal distance to home road segment}
    \item{\code{pump.id}}{numeric ID}
 }}
\usage{
ortho.proj.pump.vestry
}
\description{
Orthogonal projection of the 14 pumps from the Vestry Report.
}
\section{Notes}{
 \code{\link{pumpData}} documents the code for these data.
}

\keyword{datasets}
