% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ortho.proj}
\alias{ortho.proj}
\title{Orthogonal projection of observed cases onto road network.}
\format{
A data frame with 5 variables that records the position of the orthogonal projection of the 578 cases onto the network of roads.
 \describe{
    \item{\code{case}}{numeric case ID}
    \item{\code{road.segment}}{"address" road segment}
    \item{\code{x.proj}}{x-coordinate}
    \item{\code{y.proj}}{y-coordinate}
    \item{\code{dist}}{distance to home road segment}
    \item{\code{type}}{type of distance: orthogonal to road or Euclidean to endpoint}
 }
}
\usage{
ortho.proj
}
\description{
Orthogonal projection of observed cases onto road network.
}
\note{
\code{\link{unstackFatalities}} documents the code for these data.
}
\keyword{datasets}
