% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ortho.rect}
\alias{ortho.rect}
\title{Orthogonal rectangle.}
\format{
A data frame with 3 variables and 5 observations.
 \describe{
    \item{\code{id}}{vertex ID}
    \item{\code{x}}{longitude}
    \item{\code{y}}{latitude}
 }
}
\usage{
ortho.rect
}
\description{
Extent of map data to filter out frame shadow using sp::point.in.polygon().
}
\keyword{datasets}
