% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongOrthoProj.R
\name{orthoAddressPDF}
\alias{orthoAddressPDF}
\title{Create PDFs of orthogonal projection addresses.}
\usage{
orthoAddressPDF(path, pch = 46, cex = 1)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/"}

\item{pch}{Integer. R pch.}

\item{cex}{Numeric.}
}
\description{
For QGIS geo-referencing.
}
