% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthoRectangle.R
\name{orthogonalRectanlge}
\alias{orthogonalRectanlge}
\title{Compute orthogonal rectangle for range of map data (prototype).}
\usage{
orthogonalRectanlge(pad = 0.01)
}
\arguments{
\item{pad}{Numeric. Padding adjustment percent [0, 1].}
}
\description{
To be georeferenced for use as polygon filter.
}
