% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthoRectangle.R
\name{orthogonalRectanlgePDF}
\alias{orthogonalRectanlgePDF}
\title{PDF orthogonal rectangle  (prototype).}
\usage{
orthogonalRectanlgePDF(path)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/"}
}
\description{
PDF orthogonal rectangle  (prototype).
}
