% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonalSegment.R
\name{orthogonalSegment}
\alias{orthogonalSegment}
\title{Find a case's closest segment (orthogonal distance).}
\usage{
orthogonalSegment(case = 145, multi.core = FALSE)
}
\arguments{
\item{case}{Integer. Numerical case fatality ID.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores.}
}
\description{
Find a case's closest segment (orthogonal distance).
}
