% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifierAudit.R
\name{orthoProjector}
\alias{orthoProjector}
\title{Visually check orthogonal projection of case fatalities to road segments (address).}
\usage{
orthoProjector(case = 483, rd.seg = "326-2")
}
\arguments{
\item{case}{Numeric. case ID from \code{fatalities}.}

\item{rd.seg}{Character. Road segment ID.}
}
\description{
Visually check orthogonal projection of case fatalities to road segments (address).
}
