% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthoProjFatalities.R
\name{orthoProjFatalities}
\alias{orthoProjFatalities}
\title{Orthogonal projection of case fatalities to road segments (address).}
\usage{
orthoProjFatalities(multi.core = TRUE)
}
\arguments{
\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores.}
}
\description{
Orthogonal projection of case fatalities to road segments (address).
}
\note{
Documents ortho.proj data frame (unstack6c.R).
}
