% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthoProjPumpPDF.R
\name{orthoProjPumpPDF}
\alias{orthoProjPumpPDF}
\title{Create PDFs of pump orthogonal projections (prototype).}
\usage{
orthoProjPumpPDF(path, vestry = FALSE, pch = 46)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/".}

\item{vestry}{Logical. Use vestry report pumps.}

\item{pch}{Numeric or Character.}
}
\description{
For georeferencing in QGIS.
}
