% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxfordWeather.R
\name{oxfordWeather}
\alias{oxfordWeather}
\title{Weather data recorded in Oxford (Met Office UK).}
\usage{
oxfordWeather()
}
\value{
An R data frame.
}
\description{
Add and use last day of month as unit of observation to oxford.weather.
}
\note{
December 1860 observation is dropped due to missing "tmin" value.
}
\examples{
plot(oxfordWeather())
}
