% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionAddresses.R
\name{partitionAddressesPDF}
\alias{partitionAddressesPDF}
\title{Create PDFs of fatality address partitions (prototype).}
\usage{
partitionAddressesPDF(path, pch = 46)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/".}

\item{pch}{Numeric or Character.}
}
\description{
For georeferencing in QGIS.
}
