% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionAddresses.R
\name{partitionEvenString}
\alias{partitionEvenString}
\title{Rotate, stack and partition even n-tuple strings (n >= 4).}
\usage{
partitionEvenString(g)
}
\arguments{
\item{g}{Object. 'igraph' list of graphs.}
}
\value{
An R data frame.
}
\description{
Separate even and odd vertices to avoid point overlap.
}
