% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionFatalities.R
\name{partitionFatalities}
\alias{partitionFatalities}
\title{Partition non-address fatalities (prototype).}
\usage{
partitionFatalities(inter.point.dist = 0.15)
}
\arguments{
\item{inter.point.dist}{Numeric. Ceiling for overlapping points.}
}
\value{
An R list.
}
\description{
Based on set of subgraphs defined by selected inter-point distance.
}
