% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionFatalities.R
\name{partitionFatalitiesPDF}
\alias{partitionFatalitiesPDF}
\title{Create PDFs of non-address fatalities (prototype).}
\usage{
partitionFatalitiesPDF(path, pch = 46)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/".}

\item{pch}{Numeric or Character.}
}
\description{
For georeferencing in QGIS.
}
