% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongFrame.R
\name{partitionFrame}
\alias{partitionFrame}
\title{Partition overlapping frame points.}
\usage{
partitionFrame(inter.point.dist = 0.15)
}
\arguments{
\item{inter.point.dist}{Numeric. Ceiling for overlapping points.}
}
\value{
An R list.
}
\description{
Partition overlapping frame points.
}
