% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionAddresses.R
\name{partitionOddString}
\alias{partitionOddString}
\title{Rotate, stack and partition odd n-tuple strings (n >= 5).}
\usage{
partitionOddString(subgraphs, group = c("6", "7"))
}
\arguments{
\item{subgraphs}{Object. 'igraph' list of graphs.}

\item{group}{Character. Vector of group IDs.}
}
\value{
An R data frame.
}
\description{
Rotate, stack and partition odd n-tuple strings (n >= 5).
}
