% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionOrthoAddresses.R
\name{partitionOrthoAddressesPDF}
\alias{partitionOrthoAddressesPDF}
\title{Create PDFs of fatality address orthogonal projection partitions (prototype).}
\usage{
partitionOrthoAddressesPDF(path, pch = 46)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/".}

\item{pch}{Numeric or Character.}
}
\description{
For georeferencing in QGIS.
}
