% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionRoadEndpoints.R
\name{partitionRoadEndpoints}
\alias{partitionRoadEndpoints}
\title{Partition road endpoints to avoid over-printing of points (prototype).}
\usage{
partitionRoadEndpoints(path, multi.core = TRUE)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/".}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. See \code{vignette("Parallelization")} for details.}
}
\description{
Partition road endpoints to avoid over-printing of points (prototype).
}
