% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionRoads.R
\name{partitionRoads}
\alias{partitionRoads}
\title{Partiion overlapping road line segment endpoints.}
\usage{
partitionRoads(inter.point.dist = 0.15)
}
\arguments{
\item{inter.point.dist}{Numeric. Ceiling for overlapping points.}
}
\value{
An R list.
}
\description{
Based on set of subgraphs defined by selected inter-point distance.
}
