% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionRoads.R
\name{partitionRoadsPDF}
\alias{partitionRoadsPDF}
\title{Create PDFs of road endpoints partition (prototype).}
\usage{
partitionRoadsPDF(path)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/".}
}
\description{
For georeferencing in QGIS.
}
