% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionAddress.R
\name{partitionStringGraph}
\alias{partitionStringGraph}
\title{Partition string-of-pearl fatality address graphs.}
\usage{
partitionStringGraph(g)
}
\arguments{
\item{g}{Object. 'igraph' list of graphs.}
}
\value{
An R data frame.
}
\description{
Separate even and odd vertices to avoid point overlap.
}
