% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongRoads.R
\name{pdfPartitionRoadEndpoints}
\alias{pdfPartitionRoadEndpoints}
\title{Create PDFs of road endpoints partition (prototype).}
\usage{
pdfPartitionRoadEndpoints(path)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/".}
}
\description{
For georeferencing in QGIS .
}
