% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plague.pit}
\alias{plague.pit}
\title{Plague pit coordinates.}
\format{A data frame with 13 observations and 2 variables.
\describe{
  \item{\code{x}}{x-coordinate}
  \item{\code{y}}{y-coordinate}
}}
\usage{
plague.pit
}
\description{
Coordinates for polygon() or sp::Polygon(). In progress.
}
\keyword{datasets}
