% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plague.pit.segments}
\alias{plague.pit.segments}
\title{Plague pit coordinates.}
\format{
A data frame with 12 observations and 5 variables.
\describe{
  \item{\code{id}}{integers 1 through 12}
  \item{\code{x1}}{x-coordinate}
  \item{\code{y1}}{y-coordinate}
  \item{\code{x2}}{x-coordinate}
  \item{\code{y2}}{y-coordinate}
  \item{\code{lon1}}{longitude}
  \item{\code{lat1}}{latitude}
  \item{\code{lon2}}{longitude}
  \item{\code{lat2}}{latitude}
}
}
\usage{
plague.pit.segments
}
\description{
Coordinates for GeoPackage data. In progress.
}
\keyword{datasets}
