% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifierAudit.R
\name{plot.classifier_audit}
\alias{plot.classifier_audit}
\title{Plot test result for classifierAudit().}
\usage{
\method{plot}{classifier_audit}(x, radius = 0.1, ...)
}
\arguments{
\item{x}{An object of class "classifier_audit" created by classifierAudit().}

\item{radius}{Numeric. Controls the degree of zoom.}

\item{...}{Additional parameters.}
}
\value{
A base R graphic.
}
\description{
Plot test result for classifierAudit().
}
\examples{
plot(classifierAudit(case = 483, segment = "326-2"))
}
