% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{plot.euclidean}
\alias{plot.euclidean}
\title{Plot method for neighborhoodEuclidean().}
\usage{
\method{plot}{euclidean}(x, type = "star", ...)
}
\arguments{
\item{x}{An object of class "euclidean" created by \code{neighborhoodEuclidean()}.}

\item{type}{Character. "star", "area.points" or "area.polygons". "area" flavors only valid when \code{case.set = "expected"}.}

\item{...}{Additional plotting parameters.}
}
\value{
A base R plot.
}
\description{
Plot method for neighborhoodEuclidean().
}
\examples{
\dontrun{

plot(neighborhoodEuclidean())
plot(neighborhoodEuclidean(-6))
plot(neighborhoodEuclidean(pump.select = 6:7))
}
}
