% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclideanDistance.R
\name{plot.euclidean_distance}
\alias{plot.euclidean_distance}
\title{Plot method for euclideanDistance().}
\usage{
\method{plot}{euclidean_distance}(x, ...)
}
\arguments{
\item{x}{An object of class "euclidean_distance" created by euclideanDistance().}

\item{...}{Additional parameters.}
}
\value{
An R data frame.
}
\description{
plot output.
}
