% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclideanLatlong.R
\name{plot.euclideanLatlong}
\alias{plot.euclideanLatlong}
\title{Plot method for euclideanLatlong()}
\usage{
\method{plot}{euclideanLatlong}(x, type = "star", ...)
}
\arguments{
\item{x}{Object.}

\item{type}{Character. "star", "area.points" or "area.polygons". "area" flavors only valid when \code{case.set = "expected"}.}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for euclideanLatlong()
}
