% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoVertices.R
\name{plot.iso}
\alias{plot.iso}
\title{Plot method for isoVertices().}
\usage{
\method{plot}{iso}(x, palette = "Spectral", alpha.level = 1/3, ...)
}
\arguments{
\item{x}{An object of class "iso" created by \code{isoVertices()}.}

\item{palette}{Character. RColorBrewer palette.}

\item{alpha.level}{Numeric. Alpha level transparency}

\item{...}{Additional arguments.}
}
\value{
A vector with observed counts.
}
\description{
Plot method for isoVertices().
}
