% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoVertices.R
\name{plot.iso_vertices}
\alias{plot.iso_vertices}
\title{Plot method for isoVertices().}
\usage{
\method{plot}{iso_vertices}(x, selected.post = "all", palette = "plasma",
  alpha.level = 1/2, add = FALSE, ...)
}
\arguments{
\item{x}{An object of class "iso" created by \code{isoVertices()}.}

\item{selected.post}{Character or Numeric. Select milepost polygon. "all" or number.}

\item{palette}{Character.}

\item{alpha.level}{Numeric. Alpha level transparency}

\item{add}{Logical. Add to existing plot or separate plot.}

\item{...}{Additional arguments.}
}
\value{
A vector with observed counts.
}
\description{
Plot method for isoVertices().
}
