% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongEuclideanPath.R
\name{plot.latlong_euclidean_path}
\alias{plot.latlong_euclidean_path}
\title{Plot the path of the Euclidean distance between cases and/or pumps.}
\usage{
\method{plot}{latlong_euclidean_path}(x, zoom = TRUE, mileposts = TRUE,
  milepost.unit = "distance", milepost.interval = NULL, ...)
}
\arguments{
\item{x}{An object of class "latlong_euclidean_path" created by latlongEuclideanPath().}

\item{zoom}{Logical.}

\item{mileposts}{Logical. Plot mile/time posts.}

\item{milepost.unit}{Character. "distance" or "time".}

\item{milepost.interval}{Numeric. Mile post interval unit of distance (yard or meter) or unit of time (seconds).}

\item{...}{Additional plotting parameters.}
}
\description{
Plot the path of the Euclidean distance between cases and/or pumps.
}
