% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongNeighborhoodData.R
\name{plot.latlong_neighborhood_data}
\alias{plot.latlong_neighborhood_data}
\title{Plot method for latlongNeighborhoodData().}
\usage{
\method{plot}{latlong_neighborhood_data}(x, ...)
}
\arguments{
\item{x}{An 'igraph' object of class "latlong_neighborhood_data" created by \code{latlongNeighborhoodData()}.}

\item{...}{Additional plotting parameters.}
}
\value{
An igraph base graphics plot.
}
\description{
Visualize underlying road network (with or without cases and pumps).
}
