% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongWalkingPath.R
\name{plot.latlong_walking_path}
\alias{plot.latlong_walking_path}
\title{Plot the walking path between selected cases and/or pumps.}
\usage{
\method{plot}{latlong_walking_path}(x, zoom = TRUE, mile.posts = TRUE,
  post.unit = 50, ...)
}
\arguments{
\item{x}{An object of class "latlong_walking_path" created by latlongWalkingPath().}

\item{zoom}{Logical or Numeric. A numeric value >= 0 that controls the degree of zoom.}

\item{mile.posts}{Logical. Plot mile posts.}

\item{post.unit}{Numeric. Mile post interval (meters).}

\item{...}{Additional plotting parameters.}
}
\value{
A base R plot.
}
\description{
Plot the walking path between selected cases and/or pumps.
}
