% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongNeighborhoodEuclidean.R
\name{plot.latlongEuclidean}
\alias{plot.latlongEuclidean}
\title{Plot method for latlongNeighborhoodEuclidean()}
\usage{
\method{plot}{latlongEuclidean}(x, type = "star", ...)
}
\arguments{
\item{x}{Object.}

\item{type}{Character. "star", "area.points" or "area.polygons". "area" flavors only valid when \code{case.set = "expected"}.}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for latlongNeighborhoodEuclidean()
}
