% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongNeighborhoodEuclidean.R
\name{plot.latlongNeighborhoodEuclidean}
\alias{plot.latlongNeighborhoodEuclidean}
\title{Plot method for latlongNeighborhoodEuclidean()}
\usage{
\method{plot}{latlongNeighborhoodEuclidean}(x, euclidean.paths = TRUE, ...)
}
\arguments{
\item{x}{Object.}

\item{euclidean.paths}{Logical.}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for latlongNeighborhoodEuclidean()
}
