% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongNeighborhoodVoronoiB.R
\name{plot.latlongNeighborhoodVoronoiB}
\alias{plot.latlongNeighborhoodVoronoiB}
\title{Plot method for latlongNeighborhoodVoronoiB()}
\usage{
\method{plot}{latlongNeighborhoodVoronoiB}(x, add.cases = TRUE,
  add.pumps = TRUE, euclidean.paths = FALSE, ...)
}
\arguments{
\item{x}{Object. Currently separate classification check.}

\item{add.cases}{Logical.}

\item{add.pumps}{Logical.}

\item{euclidean.paths}{Logical. Currently separate classification check.}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for latlongNeighborhoodVoronoiB()
}
