% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoodData.R
\name{plot.neighborhood_data}
\alias{plot.neighborhood_data}
\title{Plot method for neighborhoodData().}
\usage{
\method{plot}{neighborhood_data}(x, ...)
}
\arguments{
\item{x}{An 'igraph' object of class "neighborhood_data" created by \code{neighborhoodData()}.}

\item{...}{Additional plotting parameters.}
}
\value{
A base R plot of road network (with or without cases and pumps) using igraph::plot.igraph().
}
\description{
Plot method for neighborhoodData().
}
\examples{
\dontrun{

plot(neighborhoodData())
plot(neighborhoodData(embed = FALSE))
}
}
