% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoodSnow.R
\name{plot.neighborhood_snow}
\alias{plot.neighborhood_snow}
\title{Plot method for neighborhoodSnow().}
\usage{
\method{plot}{neighborhood_snow}(x, type = "area.points", ...)
}
\arguments{
\item{x}{An object of class "neighborhood_snow" created by \code{neighborhoodSnow()}.}

\item{type}{Character. "roads", "area.points" or "area.polygons".}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for neighborhoodSnow().
}
