% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxfordWeather.R
\name{plot.oxfordWeather}
\alias{plot.oxfordWeather}
\title{Plot method for oxfordWeather().}
\usage{
\method{plot}{oxfordWeather}(x, statistic = "temperature", ...)
}
\arguments{
\item{x}{object.}

\item{statistic}{Character.}

\item{...}{Additional plotting parameters.}
}
\value{
A base R plot.
}
\description{
Plot method for oxfordWeather().
}
\examples{
plot(oxfordWeather())
plot(oxfordWeather(), "rain")
}
