% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointOverlap.R
\name{plot.pointOverlap}
\alias{plot.pointOverlap}
\title{Plot method for pointOverlap().}
\usage{
\method{plot}{pointOverlap}(x, data = "ntuple", type = "map", ...)
}
\arguments{
\item{x}{Object.}

\item{data}{Character. "ntuple" or "dyad".}

\item{type}{Character. "map" or "graph".}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for pointOverlap().
}
