% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointOverlapRoad.R
\name{plot.pointOverlapRoad}
\alias{plot.pointOverlapRoad}
\title{Plot method for pointOverlapRoad().}
\usage{
\method{plot}{pointOverlapRoad}(x, data = "ntuple", type = "map", ...)
}
\arguments{
\item{x}{Object.}

\item{data}{Character. "ntuple" or "dyad".}

\item{type}{Character. "map" or "graph".}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for pointOverlapRoad().
}
