% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/povertyLondon.R
\name{plot.povertyLondon}
\alias{plot.povertyLondon}
\title{Plot method for povertyLondon().}
\usage{
\method{plot}{povertyLondon}(x, district = c("City", "Westminster", "Marylebone", "St. Giles"), ...)
}
\arguments{
\item{x}{object.}

\item{district}{Character. Selected district(s).}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for povertyLondon().
}
