% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profilePlot.R
\name{plot.profile}
\alias{plot.profile}
\title{Plot method for profilePerspective().}
\usage{
\method{plot}{profile}(x, ...)
}
\arguments{
\item{x}{An object of class "profile" created by profilePerspective().}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for profilePerspective().
}
