% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile2D.R
\name{plot.profile2D}
\alias{plot.profile2D}
\title{Plot method for profile2D().}
\usage{
\method{plot}{profile2D}(x, type = "base", ...)
}
\arguments{
\item{x}{An object of class "profile2D" created by \code{profile2D()}.}

\item{type}{Character. Type of graphic: "base" or "ggplot2".}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for profile2D().
}
