% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile3D.R
\name{plot.profile3D}
\alias{plot.profile3D}
\title{Plot method for profile3D().}
\usage{
\method{plot}{profile3D}(x, pump.subset = NULL,
  drop.neg.selection = FALSE, ...)
}
\arguments{
\item{x}{An object of class "profile3D" created by \code{profile3D()}.}

\item{pump.subset}{Numeric. Vector of numeric pump IDs to subset from the neighborhoods defined by \code{pump.select}. Negative selection possible. \code{NULL} selects all pumps in \code{pump.select}.}

\item{drop.neg.selection}{Logical. Drop negative selection.}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for profile3D().
}
