% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile2D.R
\name{plot.profile_perspective}
\alias{plot.profile_perspective}
\title{Plot method for profilePerspective().}
\usage{
\method{plot}{profile_perspective}(x, ...)
}
\arguments{
\item{x}{An object of class "profile" created by \code{profilePerspective()}.}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for profilePerspective().
}
