% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sohoGraph.R
\name{plot.soho_graph}
\alias{plot.soho_graph}
\title{Plot method for sohoGraph().}
\usage{
\method{plot}{soho_graph}(x, ...)
}
\arguments{
\item{x}{An 'igraph' object of class "neighborhood_data" created by \code{sohoGraph()}.}

\item{...}{Additional plotting parameters.}
}
\value{
A base R plot.
}
\description{
Visualize underlying road network (with or without cases and pumps).
}
