% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeries.R
\name{plot.time.series}
\alias{plot.time.series}
\title{Plot aggregate time series data from Vestry report.}
\usage{
\method{plot}{time.series}(x, statistic = "fatal.attacks",
  pump.handle = TRUE, ...)
}
\arguments{
\item{x}{An object of class "time.series": \code{\link[cholera]{time.series.snow}} or \code{\link[cholera]{time.series.vestry}}.}

\item{statistic}{Character. Fatality measure: either "fatal.attacks" (default) or "deaths".}

\item{pump.handle}{Logical. Indicate date of removal of Broad Street pump handle.}

\item{...}{Additional plotting parameters.}
}
\description{
Plot aggregate fatality data and indicates the date of the removal of the handle of the Broad Street pump.
}
\examples{
# plot(time.series.snow)
}
\seealso{
\code{\link{timeSeries}}
}
