% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeries.R
\name{plot.time_series}
\alias{plot.time_series}
\title{Plot aggregate time series data from Vestry report.}
\usage{
\method{plot}{time_series}(x, all.data = FALSE,
  statistic = "fatal.attacks", vestry = FALSE, multi.plot = FALSE,
  main = "Removal of the Broad Street Pump Handle", points = TRUE,
  pump.handle = TRUE, weekend = FALSE, xlab = "Date", ylab = NULL, ...)
}
\arguments{
\item{x}{Data frame of four fatality datasets.}

\item{all.data}{Logical. Use all 4 datasets.}

\item{statistic}{Character. Fatality measure:  "fatal.attacks" or "deaths".}

\item{vestry}{Logical. TRUE = Vestry Report; FALSE = John Snow.}

\item{multi.plot}{Logical. Multiple data sets in single plot. Meaningful only when `all.data = TRUE`.}

\item{main}{Character. Title of graph.}

\item{points}{Logical. Plot points.}

\item{pump.handle}{Logical. Plot date of removal of Broad Street pump handle.}

\item{weekend}{Logical. Highlight weekend.}

\item{xlab}{Character. x-axis label.}

\item{ylab}{Character. y-axis label.}

\item{...}{Additional plotting parameters.}
}
\description{
Plot aggregate fatality data and indicates the date of the removal of the handle of the Broad Street pump.
}
