% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoiLatlong.R
\name{plot.voronoi_latlong}
\alias{plot.voronoi_latlong}
\title{Plot method for voronoiLatlong()}
\usage{
\method{plot}{voronoi_latlong}(x, add.pumps = TRUE,
  delaunay.voronoi = "voronoi", euclidean.paths = FALSE, ...)
}
\arguments{
\item{x}{Object.}

\item{add.pumps}{Logical.}

\item{delaunay.voronoi}{Character "delaunay", "voronoi", or "both".}

\item{euclidean.paths}{Logical.}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for voronoiLatlong()
}
