% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoiNominal.R
\name{plot.voronoi_nominal}
\alias{plot.voronoi_nominal}
\title{Plot Voronoi neighborhoods.}
\usage{
\method{plot}{voronoi_nominal}(x, delaunay.voronoi = "voronoi",
  euclidean.paths = FALSE, ...)
}
\arguments{
\item{x}{An object of class "voronoi" created by \code{voronoiNominal()}.}

\item{delaunay.voronoi}{Character "delaunay", "voronoi", or "both".}

\item{euclidean.paths}{Logical. Plot all Euclidean paths (star graph).}

\item{...}{Additional plotting parameters.}
}
\value{
A base R graph.
}
\description{
Plot Voronoi neighborhoods.
}
\seealso{
\code{voronoiNominal()}

\code{addVoronoi()}
}
