% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkingNominal.R
\name{plot.walking}
\alias{plot.walking}
\title{Plot method for walkingNominal().}
\usage{
\method{plot}{walking}(x, type = "roads", tsp.method = "repetitive_nn",
  ...)
}
\arguments{
\item{x}{An object of class "walking" created by \code{walkingNominal()}.}

\item{type}{Character. "roads", "area.points" or "area.polygons". "area" flavors only valid when \code{case.set = "expected"}.}

\item{tsp.method}{Character. Traveling salesperson problem algorithm.}

\item{...}{Additional plotting parameters.}
}
\value{
A base R plot.
}
\description{
Plot method for walkingNominal().
}
\note{
When plotting area graphs with simulated data (i.e., \code{case.set = "expected"}), there may be discrepancies between observed cases and expected neighborhoods, particularly between neighborhoods. type = "roads" inspired by Shiode et. al. (2015).
}
