% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkingDistance.R
\name{plot.walking_distance}
\alias{plot.walking_distance}
\title{Plot method for walkingDistance().}
\usage{
\method{plot}{walking_distance}(x, ...)
}
\arguments{
\item{x}{An object of class "walking_distance" created by walkingDistance().}

\item{...}{Additional parameters.}
}
\value{
An R data frame.
}
\description{
plot output.
}
