% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkingPath.R
\name{plot.walking_path}
\alias{plot.walking_path}
\title{Plot the walking path between cases and/or pumps.}
\usage{
\method{plot}{walking_path}(x, zoom = TRUE, radius = 0.5,
  unit.posts = "distance", unit.interval = NULL, ...)
}
\arguments{
\item{x}{An object of class "walking_path" created by walkingPath().}

\item{zoom}{Logical.}

\item{radius}{Numeric. Controls the degree of zoom.}

\item{unit.posts}{Character. "distance" for mileposts; "time" for timeposts.}

\item{unit.interval}{Numeric. Interval between posts: mileposts default is 50 meters; timepost default is 60 seconds.}

\item{...}{Additional plotting parameters.}
}
\value{
A base R plot.
}
\description{
Plot the walking path between cases and/or pumps.
}
\section{Note}{
 "unit.posts" and "unit.interval" are represented as arrows pointing away from the destination (i.e., the reverse path).
}

\examples{
plot(walkingPath(15))
plot(walkingPath(15), unit.posts = "time")
}
