% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkingPathB.R
\name{plot.walking_path_B}
\alias{plot.walking_path_B}
\title{Plot the walking path between selected cases and/or pumps.}
\usage{
\method{plot}{walking_path_B}(x, zoom = TRUE, long.title = TRUE,
  mileposts = TRUE, milepost.unit = "distance", milepost.interval = NULL,
  alpha.level = 1, ...)
}
\arguments{
\item{x}{An object of class "latlong_walking_path" created by latlongWalkingPath().}

\item{zoom}{Logical or Numeric. A numeric value >= 0 controls the degree of zoom. A value of 1 equals zoom = TRUE.}

\item{long.title}{Logical. Tile with names.}

\item{mileposts}{Logical. Plot mile/time posts.}

\item{milepost.unit}{Character. "distance" or "time".}

\item{milepost.interval}{Numeric. Mile post interval unit of distance (yard or meter) or unit of time (seconds).}

\item{alpha.level}{Numeric. Alpha level transparency for path: a value in [0, 1].}

\item{...}{Additional plotting parameters.}
}
\value{
A base R plot.
}
\description{
Plot the walking path between selected cases and/or pumps.
}
