% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{plot.walkingB}
\alias{plot.walkingB}
\title{Plot method for neighborhoodPaths().}
\usage{
\method{plot}{walkingB}(x, ...)
}
\arguments{
\item{x}{An object of class "walkingB" created by neighborhoodPaths().}

\item{...}{Additional plotting parameters.}
}
\value{
A base R plot.
}
\description{
Plot method for neighborhoodPaths().
}
\examples{
# plot(neighborhoodPaths())
}
