% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkingLatlong.R
\name{plot.walkingLatlong}
\alias{plot.walkingLatlong}
\title{Plot method for walkingLatlong().}
\usage{
\method{plot}{walkingLatlong}(x, type = "roads", ...)
}
\arguments{
\item{x}{An object of class "latlong_walking" created by \code{walkingLatlongwalkingLatlong()}.}

\item{type}{Character. "area.points", "area.polygons" or "roads". For walkingLatlong(case.set = "expected").}

\item{...}{Additional plotting parameters.}
}
\value{
A base R plot.
}
\description{
Plot method for walkingLatlong().
}
