% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winterTemperatures.R
\name{plot.winterTemperatures}
\alias{plot.winterTemperatures}
\title{Plot method for winterTemperatures().}
\usage{
\method{plot}{winterTemperatures}(x, end.date = NULL, ...)
}
\arguments{
\item{x}{object.}

\item{end.date}{Date.}

\item{...}{Additional plotting parameters.}
}
\value{
A base R plot.
}
\description{
Plot method for winterTemperatures().
}
