% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeSeries.R
\name{plotTimeSeries}
\alias{plotTimeSeries}
\title{Plot time series data.}
\usage{
plotTimeSeries(statistic = "fatal.attacks", pump.handle = TRUE,
  vestry = FALSE, ...)
}
\arguments{
\item{statistic}{Character. "deaths" or "fatal.attacks".}

\item{pump.handle}{Logical. Annotate date of removal of Broad Street pump handle.}

\item{vestry}{Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 in the original map.}

\item{...}{Additional plotting parameters.}
}
\value{
A base R graphics plot.
}
\description{
Vestry report data.
}
\examples{
plotTimeSeries()
}
\seealso{
\code{
\link{time.series.snow}},
\code{\link{time.series.vestry}},
\code{vignette}("time.series")
}
