% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongAuxilliaryFunctions.R
\name{pointIndex}
\alias{pointIndex}
\title{Index of subsets.}
\usage{
pointIndex(max.ct = 321, bin.size = 50)
}
\arguments{
\item{max.ct}{Integer. Upper count of observations.}

\item{bin.size}{Integer. bin size size of subgroups.}
}
\description{
Index of subsets.
}
