% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointOverlap.R
\name{pointOverlap}
\alias{pointOverlap}
\title{Census of overlapping fatalties}
\usage{
pointOverlap(path, dist.between.pts = 0.15, output = "n-tuple")
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/"}

\item{dist.between.pts}{Numeric. Ceiling for overlapping points.}

\item{output}{Character. "dyad" (data frame) or "n-tuple" (PDF)}
}
\description{
"Resolution" of georeferenced points with pch = 46 (i.e., pch = ".").
}
