% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointOverlapRoad.R
\name{pointOverlapRoad}
\alias{pointOverlapRoad}
\title{Census of overlapping road line segment endpoints.}
\usage{
pointOverlapRoad(inter.point.dist = 0.15)
}
\arguments{
\item{inter.point.dist}{Numeric. Ceiling for overlapping points.}
}
\description{
"Resolution" of georeferenced points with pch = 46 (i.e., pch = ".").
}
