% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongPumps.R
\name{pointsFromGeoTIFF}
\alias{pointsFromGeoTIFF}
\title{Extract points from GeoTiff (prototype).}
\usage{
pointsFromGeoTIFF(x)
}
\arguments{
\item{x}{Object. GeoTIFF.}
}
\description{
Extract points from GeoTiff (prototype).
}
