% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearlString.R
\name{polygonAudit}
\alias{polygonAudit}
\title{Plot neighborhood polygon.}
\usage{
polygonAudit(x, i = 1)
}
\arguments{
\item{x}{Object. Neighborhood data.}

\item{i}{Numeric. Neighborhood ID.}
}
\description{
Plot neighborhood polygon.
}
