% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unstack.R
\name{print.classifier_audit}
\alias{print.classifier_audit}
\title{Print method for classifierAudit().}
\usage{
\method{print}{classifier_audit}(x, ...)
}
\arguments{
\item{x}{An object of class "classifier_audit" created by classifierAudit().}

\item{...}{Additional parameters.}
}
\value{
An R data frame.
}
\description{
Return summary results.
}
\examples{
classifierAudit(case = 483, segment = "326-2")
print(classifierAudit(case = 483, segment = "326-2"))
}
