% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclideanPathB.R
\name{print.euclidean_path_B}
\alias{print.euclidean_path_B}
\title{Print method for walkingPathB().}
\usage{
\method{print}{euclidean_path_B}(x, ...)
}
\arguments{
\item{x}{An object of class "euclidean_path_B" created by euclideanPathB().}

\item{...}{Additional parameters.}
}
\value{
An R data frame.
}
\description{
Summary output.
}
