% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclideanPathB.R
\name{print.euclidean_pathB}
\alias{print.euclidean_pathB}
\title{Summary of euclideanPath().}
\usage{
\method{print}{euclidean_pathB}(x, ...)
}
\arguments{
\item{x}{An object of class "euclidean_path" created by euclideanPath().}

\item{...}{Additional parameters.}
}
\value{
An R data frame.
}
\description{
Summary of euclideanPath().
}
\examples{
euclideanPath(1)
print(euclideanPath(1))
}
