% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoVertices.R
\name{print.iso}
\alias{print.iso}
\title{Print method for isoVertices().}
\usage{
\method{print}{iso}(x, ...)
}
\arguments{
\item{x}{An object of class "iso" created by \code{isoVertices()}.}

\item{...}{Additional arguments.}
}
\value{
A vector with observed counts.
}
\description{
Print method for isoVertices().
}
