% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongWalkingPath.R
\name{print.latlong_walking_path}
\alias{print.latlong_walking_path}
\title{Print method for latlongWalkingPath().}
\usage{
\method{print}{latlong_walking_path}(x, ...)
}
\arguments{
\item{x}{An object of class "latlong_walking_path" created by latlongWalkingPath().}

\item{...}{Additional parameters.}
}
\value{
An R data frame.
}
\description{
Summary output.
}
