% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongNeighborhoodVoronoi.R
\name{print.latlongNeighborhoodVoronoi}
\alias{print.latlongNeighborhoodVoronoi}
\title{Print method for latlongNeighborhoodVoronoi().}
\usage{
\method{print}{latlongNeighborhoodVoronoi}(x, ...)
}
\arguments{
\item{x}{An object of class "latlongNeighborhoodVoronoi" created by \code{latlongNeighborhoodVoronoi()}.}

\item{...}{Additional arguments.}
}
\value{
A list of argument values.
}
\description{
Parameter values for latlongNeighborhoodVoronoi().
}
