% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongNeighborhoodVoronoiB.R
\name{print.latlongNeighborhoodVoronoiB}
\alias{print.latlongNeighborhoodVoronoiB}
\title{Print method for latlongNeighborhoodVoronoiB().}
\usage{
\method{print}{latlongNeighborhoodVoronoiB}(x, ...)
}
\arguments{
\item{x}{An object of class "latlongNeighborhoodVoronoiB" created by \code{latlongNeighborhoodVoronoiB()}.}

\item{...}{Additional arguments.}
}
\value{
A list of argument values.
}
\description{
Parameter values for latlongNeighborhoodVoronoiB().
}
