% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile2D.R
\name{print.profile2D}
\alias{print.profile2D}
\title{Print method for profile2D().}
\usage{
\method{print}{profile2D}(x, ...)
}
\arguments{
\item{x}{An object of class "profile2D" created by \code{profile2D()}.}

\item{...}{Additional plotting parameters.}
}
\description{
Count of cases.
}
