% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile3D.R
\name{print.profile3D}
\alias{print.profile3D}
\title{Print method for profile3D().}
\usage{
\method{print}{profile3D}(x, ...)
}
\arguments{
\item{x}{An object of class "profile3D" created by \code{profile3D()}.}

\item{...}{Additional parameters.}
}
\description{
Print method for profile3D().
}
