% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeries.R
\name{print.time_series}
\alias{print.time_series}
\title{Print summary data for timeSeries().}
\usage{
\method{print}{time_series}(x, ...)
}
\arguments{
\item{x}{An object of class "time_series" created by timeSeries().}

\item{...}{Additional parameters.}
}
\value{
An R data frame.
}
\description{
Return summary results.
}
\examples{
timeSeries()
print(timeSeries())
}
