% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoiNominal.R
\name{print.voronoi}
\alias{print.voronoi}
\title{Print method for voronoiNominal().}
\usage{
\method{print}{voronoi}(x, ...)
}
\arguments{
\item{x}{An object of class "voronoi" created by \code{voronoiNominal()}.}

\item{...}{Additional arguments.}
}
\value{
A list of argument values.
}
\description{
Parameter values for voronoiNominal().
}
