% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoiLatlong.R
\name{print.voronoiLatlong}
\alias{print.voronoiLatlong}
\title{Print method for voronoiLatlong().}
\usage{
\method{print}{voronoiLatlong}(x, ...)
}
\arguments{
\item{x}{An object of class "voronoiLatlong" created by \code{voronoiLatlong()}.}

\item{...}{Additional arguments.}
}
\value{
A list of argument values.
}
\description{
Parameter values for voronoiLatlong().
}
