% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walking.R
\name{print.walking}
\alias{print.walking}
\title{Print method for neighborhoodWalking().}
\usage{
\method{print}{walking}(x, ...)
}
\arguments{
\item{x}{An object of class "walking" created by neighborhoodWalking().}

\item{...}{Additional parameters.}
}
\value{
An R vector.
}
\description{
Return count of paths (anchor cases) by pump neighborhood.
}
\section{Note that printing an object with case.set = "expected" is computationally intensive () since you are computing the paths of almost 5000 cases.}{
NA
}

\examples{
# neighborhoodWalking()
# print(neighborhoodWalking())
}
