% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkingDistance.R
\name{print.walking_distance}
\alias{print.walking_distance}
\title{Print method for walkingDistance().}
\usage{
\method{print}{walking_distance}(x, ...)
}
\arguments{
\item{x}{An object of class "walking_distance" created by walkingDistance().}

\item{...}{Additional parameters.}
}
\value{
An R data frame.
}
\description{
Return summary results.
}
\examples{
walkingDistance(1)
print(walkingDistance(1))
}
