% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkingPath.R
\name{print.walking_path}
\alias{print.walking_path}
\title{Print method for walkingPath().}
\usage{
\method{print}{walking_path}(x, ...)
}
\arguments{
\item{x}{An object of class "walking_path" created by walkingPath().}

\item{...}{Additional parameters.}
}
\value{
An R data frame.
}
\description{
Summary output.
}
\examples{
\donttest{

walkingPath()
print(walkingPath())
}
}
