% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkingPathB.R
\name{print.walking_path_B}
\alias{print.walking_path_B}
\title{Print method for walkingPathB().}
\usage{
\method{print}{walking_path_B}(x, ...)
}
\arguments{
\item{x}{An object of class "latlong_walking_path" created by latlongWalkingPath().}

\item{...}{Additional parameters.}
}
\value{
An R data frame.
}
\description{
Summary output.
}
