% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{print.walkingB}
\alias{print.walkingB}
\title{Print method for neighborhoodPaths().}
\usage{
\method{print}{walkingB}(x, ...)
}
\arguments{
\item{x}{An object of class "walkingB" created by neighborhoodPaths().}

\item{...}{Additional parameters.}
}
\value{
An R vector.
}
\description{
Return count of paths (anchor cases) by pump neighborhood.
}
\examples{
# neighborhoodPaths()
# print(neighborhoodPaths())
}
