% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile2D.R
\name{profile2D}
\alias{profile2D}
\title{2D Profile Plot.}
\usage{
profile2D(pump = 7, angle = 0, vestry = FALSE, multi.core = FALSE,
  type = "base", drop = TRUE)
}
\arguments{
\item{pump}{Numeric. Select pump as focal point.}

\item{angle}{Numeric. Angle of perspective axis in degrees.}

\item{vestry}{Logical. \code{TRUE} uses the 14 pumps from the Vestry Report. \code{FALSE} uses the 13 in the original map.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. On Windows, only \code{multi.core = FALSE} is available.}

\item{type}{Character. "base" or "ggplot2".}

\item{drop}{Logical. Drop negative selection.}
}
\description{
2D Profile Plot.
}
