% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profilePlot.R
\name{profilePlot}
\alias{profilePlot}
\title{Profile Plot.}
\usage{
profilePlot(pump = 7, angle = 0, vestry = FALSE,
  multi.core = FALSE, type = "threejs", drop = TRUE)
}
\arguments{
\item{pump}{Numeric. Selected pump focal point.}

\item{angle}{Numeric. Angle of perspective axis in degrees.}

\item{vestry}{Logical. \code{TRUE} uses the 14 pumps from the Vestry Report. \code{FALSE} uses the 13 in the original map.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. On Windows, only \code{multi.core = FALSE} is available.}

\item{type}{Character. "base", "ggplot2", or "threejs".}

\item{drop}{Logical. Drop negative selection.}
}
\description{
Profile Plot.
}
