% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pump.cases}
\alias{pump.cases}
\title{List of the 321 observed "addresses" of cholera fatalities grouped by pump neighborhood.}
\format{A list with 13 vectors.
\describe{
 \item{\code{pump.cases}}{numerical ID}
}}
\usage{
pump.cases
}
\description{
List of the 321 observed "addresses" of cholera fatalities grouped by pump neighborhood.
}
\section{Notes}{
 \code{\link{neighborhoodWalking}} documents the code for these data.

For details, see \code{vignette}("pump.neighborhoods").
}

\examples{
neighborhoodWalking()$pump.case
}
\keyword{datasets}
