% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pump.cases.sp}
\alias{pump.cases.sp}
\title{List of 4972 regular "simulated" cases grouped by pump neighborhood.}
\format{A list with 13 vectors.
\describe{
 \item{\code{pump.cases.sp}}{numerical ID}
}}
\usage{
pump.cases.sp
}
\description{
List of 4972 regular "simulated" cases grouped by pump neighborhood.
}
\seealso{
\code{\link[cholera]{neighborhoodWalking}}
}
\keyword{datasets}
