% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpCase.R
\name{pumpCase}
\alias{pumpCase}
\title{Numeric case IDs by pump neighborhood.}
\usage{
pumpCase(obj)
}
\arguments{
\item{obj}{An object created by neighborhoodVoronoi() or neighborhoodWalking().}
}
\value{
A list of pump neighborhoods with the numeric ID of cases.
}
\description{
Numeric case IDs by pump neighborhood.
}
\examples{
# pumpCase(neighborhoodVoronoi())
# pumpCase(neighborhoodWalking())
}
\seealso{
\code{\link{neighborhoodVoronoi}}, \code{\link{neighborhoodWalking}},
}
