% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpCases.R
\name{pumpCases}
\alias{pumpCases}
\title{Numerical case IDs by Voronoi neighborhood.}
\usage{
pumpCases(obj)
}
\arguments{
\item{obj}{An object of class "voronoi" created by neighborhoodVoronoi().}
}
\value{
A list of pump neighborhoods with the numeric ID of observed cases.
}
\description{
Numerical case IDs by Voronoi neighborhood.
}
\examples{
dat <- neighborhoodVoronoi()
pumpCases(dat)
}
