% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpFatalities.R
\name{pumpFatalities}
\alias{pumpFatalities}
\title{Compute fatalities by pump.}
\usage{
pumpFatalities(pump.select = NULL, metric = "walking", vestry = FALSE)
}
\arguments{
\item{pump.select}{Numeric. Pump candidates to consider. Default is \code{NULL}: all pumps are used. Otherwise, selection by a vector of numeric IDs: 1 to 13 for \code{pumps}; 1 to 14 for \code{pumps.vestry}. Negative selection allowed.}

\item{metric}{Character. "eucldidean" or "walking".}

\item{vestry}{Logical. \code{TRUE} uses the 14 pumps from the Vestry Report. \code{FALSE} uses the 13 in the original map.}
}
\description{
Compute fatalities by pump.
}
\examples{
\dontrun{
pumpFatalities(pump.select = -7)
pumpFatalities(metric = "euclidean")
pumpFatalities(vestry = TRUE)
}
}
