% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pumps}
\alias{pumps}
\title{Dodson and Tobler's pump data with street name.}
\format{A data frame with 13 observations and 4 variables that describe the pumps on Snow's map.
\describe{
  \item{\code{id}}{pump number between 1 and 13}
  \item{\code{street}}{nearest street}
  \item{\code{x}}{x-coordinate}
  \item{\code{y}}{y-coordinate}
}}
\usage{
pumps
}
\description{
Adds and amends road locations for water pumps from John Snow's map to Dodson and Tobler's street data. The latter are available at Michael Friendly's HistData::Snow.streets.
}
\note{
\code{\link{pumpData}} documents the code for these data.
}
\seealso{
\code{\link{pumpLocator}}
}
\keyword{datasets}
