% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pumps.vestry}
\alias{pumps.vestry}
\title{Vestry report pump data.}
\format{A data frame with 14 observations and 4 variables that describe the pumps on the Vestry Report version of Snow's map. Amends Broad Street pump location.
\describe{
  \item{\code{id}}{pump number between 1 and 14}
  \item{\code{street}}{nearest street}
  \item{\code{x}}{x-coordinate}
  \item{\code{y}}{y-coordinate}
}}
\usage{
pumps.vestry
}
\description{
From the version Snow includes in the Vestry Report, this data set adds the Hanover Square pump and moves the Broad Street pump.
}
\keyword{datasets}

