% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pumps.vestry2}
\alias{pumps.vestry2}
\title{Vestry report pump data.}
\format{A data frame with 14 observations and 4 variables.
\describe{
  \item{\code{id}}{pump number between 1 and 14}
  \item{\code{street}}{nearest street}
  \item{\code{x}}{x-coordinate}
  \item{\code{y}}{y-coordinate}
  \item{\code{longitude}}{x-coordinate}
  \item{\code{latitude}}{y-coordinate}
}}
\usage{
pumps.vestry2
}
\description{
These data include the fourteenth pump, at Hanover Square, and the "corrected" location of the Broad Street pump that Snow includes in the second version of his map in the Vestry report.
}
\note{
\code{\link{pumpData}} documents the code for these data.
}
\seealso{
\code{\link{pumpLocator}}
}
\keyword{datasets}
