% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongPumps.R
\name{pumpsPDF}
\alias{pumpsPDF}
\title{Create PDFs of pumps.}
\usage{
pumpsPDF(path, vestry = FALSE, pch = 46, cex = 1)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/"}

\item{vestry}{Logical.}

\item{pch}{Integer. R pch.}

\item{cex}{Numeric.}
}
\description{
For QGIS geo-referencing.
}
