% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpSummary.R
\name{pumpSummary}
\alias{pumpSummary}
\title{Summary statistics for Voronoi neighborhoods.}
\usage{
pumpSummary(obj, statistic)
}
\arguments{
\item{obj}{An object of class "voronoi" created by neighborhoodVoronoi().}

\item{statistic}{Character. "address" summarizes the number of addresses in pump neighbohoods. "fatality", the default, summarizes the number of fatalities in pump neighborhoods.}
}
\value{
A data frame with observed and expected counts, observed percentage, and the Pearson residual, (observed - expected) / sqrt(expected).
}
\description{
Summary statistics for Voronoi neighborhoods.
}
\examples{
dat <- neighborhoodVoronoi()
pumpSummary(dat)

pumpSummary(dat, statistic = "address")
}
\seealso{
\code{addVoronoi()}
\code{plot.voronoi()}
}

