% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongVoronoi.R
\name{pumpsVoronoiPolygons}
\alias{pumpsVoronoiPolygons}
\title{Compute Vertices of Voronoi Polygons.}
\usage{
pumpsVoronoiPolygons(vestry = FALSE)
}
\arguments{
\item{vestry}{Logical.}
}
\description{
Compute Vertices of Voronoi Polygons.
}
\note{
This documents the computation of the voronoi.polygons and voronoi.polygons.vestry lists.
}
