% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongAuxilliaryFunctions.R
\name{radians}
\alias{radians}
\title{Compute radians between observed point and centroid of 'roads'.}
\usage{
radians(points.data)
}
\arguments{
\item{points.data}{Object. Data frame of centroid and point.}
}
\description{
Compute radians between observed point and centroid of 'roads'.
}
