% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rd.sample}
\alias{rd.sample}
\title{Sample of road intersections (segment endpoints).}
\format{
A list with 2 variables that list randomly re-arranges unique road intersections (segment endpoints).
 \describe{
    \item{\code{one}}{endpoints with 1 intersection}
    \item{\code{three}}{endpoints with 3 intersections}
 }
}
\usage{
rd.sample
}
\description{
Sample of road intersections (segment endpoints).
}
\keyword{datasets}
