% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rectangle.filter}
\alias{rectangle.filter}
\title{Rectangular filter data.}
\format{
A data frame with 2 variables and 4 observations.
 \describe{
    \item{\code{x}}{longitude}
    \item{\code{y}}{latitude}
 }
}
\usage{
rectangle.filter
}
\description{
Coordinates to filter out frame shadow using sp::point.in.polygon().
}
\keyword{datasets}
