% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongAuxilliaryFunctions.R
\name{referenceRadian}
\alias{referenceRadian}
\title{Estimate of georeferencing rotation (radians).}
\usage{
referenceRadian(id1 = 66, id2 = 171)
}
\arguments{
\item{id1}{Numeric. Road segment endpoint ID. Margaret Street.}

\item{id2}{Numeric. Road segment endpoint ID. Phoenix Yard.}
}
\description{
QGIS georeferencing realigns map: left side approximately parallel to y-axis.
}
\note{
The two default points are the first two observations on the top left.
}
