% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{road.gcp.points}
\alias{road.gcp.points}
\title{Ground control points for QGIS 3.10 (experimental).}
\format{
A data frame with 8 observations and 8 variables.
\describe{
  \item{\code{mapX}}{}
  \item{\code{mapY}}{}
  \item{\code{pixelX }}{}
  \item{\code{pixelY}}{}
  \item{\code{enable}}{}
  \item{\code{dX}}{}
  \item{\code{dY}}{}
  \item{\code{residual}}{}
}
}
\usage{
road.gcp.points
}
\description{
Ground control points for QGIS 3.10 (experimental).
}
\keyword{datasets}
