% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roadHighlight.R
\name{roadHighlight}
\alias{roadHighlight}
\title{Highlight road by street name.}
\usage{
roadHighlight(road.name, col = "red", lwd = 1)
}
\arguments{
\item{road.name}{Character vector. Note that \code{streetNameLocator}() tries to correct for case and to remove extra spaces.}

\item{col}{Character. Highlight color.}

\item{lwd}{Numeric. Line width.}
}
\value{
A base R graphics segment(s).
}
\description{
Highlight road by street name.
}
