% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimPaths.R
\name{roadLength}
\alias{roadLength}
\title{Compute total road length in pump neighborhood}
\usage{
roadLength(trimmed.data)
}
\arguments{
\item{trimmed.data}{List of trimmed road segments created by trimPaths().}
}
\value{
A vector of total length of roads in pump neighbohood.
}
\description{
Compute total road length in pump neighborhood
}
\examples{
# dat <- trimPaths(6:7)
# roadLength(dat)
}

