% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roadPointDistance.R
\name{roadPointDistance}
\alias{roadPointDistance}
\title{Compute distance between endpoints of road segments.}
\usage{
roadPointDistance(a = 85, b = 86, meters = FALSE)
}
\arguments{
\item{a}{Numeric. Road point ID.}

\item{b}{Numeric. Road point ID.}

\item{meters}{Logical. Compute metric (meters) or nominal distance.}
}
\description{
Compute distance between endpoints of road segments.
}
