% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roadSegmentFix.R
\name{roadSegmentFix}
\alias{roadSegmentFix}
\title{Bar orientation classification errors.}
\usage{
roadSegmentFix()
}
\description{
Bar orientation classification errors.
}
\note{
Bars lie parallel to the road where that fatality is observed. This can lead to (classification) errors when using orthogonal projection to assign a street address: the closest road is not always the right road. This R list manually assigns those problematic bars to their "correct" road segment.
}
