% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roadSegments2.R
\name{roadSegments2}
\alias{roadSegments2}
\title{Reshape 'roads2' data frame into 'road.segments2' data frame.}
\usage{
roadSegments2()
}
\value{
An R data frame.
}
\description{
Used to integrate pumps and cases into road network when computing walking neighborhoods.
}
\note{
This function documents the code that generates \code{\link[cholera]{road.segments2}}.
}
