% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roadsLatLong.R
\name{roadsLatLong}
\alias{roadsLatLong}
\title{Compute latitude and longitude from geo TIFF (prototype).}
\usage{
roadsLatLong(geotiff, cutpoint = 6e-06, multi.core = TRUE)
}
\arguments{
\item{geotiff}{Object. GeoTIFF file.}

\item{cutpoint}{Numeric. Cutpoint for agglomerative hierarchical cluster analysis.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. See \code{vignette("Parallelization")} for details.}
}
\description{
Georeferenced TIFF file from QGIS3.10.
}
\note{
This function documents the code that generates "roads2" latitude and longitude estimates.
}
