% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roadsPlot.R
\name{roadsPlot}
\alias{roadsPlot}
\title{Extract points from GeoTiff (prototype).}
\usage{
roadsPlot(quadrant = 1, add.roads = TRUE)
}
\arguments{
\item{quadrant}{Numeric. 1:4.}

\item{add.roads}{Logical.}
}
\description{
Extract points from GeoTiff (prototype).
}
