% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongAuxilliaryFunctions.R
\name{rotatePoint}
\alias{rotatePoint}
\title{Rotate points (prototype).}
\usage{
rotatePoint(id = 1, dataset = "roads")
}
\arguments{
\item{id}{Numeric. Road segment endpoint ID.}

\item{dataset}{Character. "roads", "fatalities", "fatalities.address", "pumps", "pumps.vestry", or "ortho.proj".}
}
\description{
Rotate points (prototype).
}
