% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongAuxilliaryFunctions.R
\name{segmentDistance}
\alias{segmentDistance}
\title{Compute estimated length of 'road.segments' in meters.}
\usage{
segmentDistance(dat, latlong = FALSE)
}
\arguments{
\item{dat}{Object. R data frame of road segments}

\item{latlong}{Logical. Data use longitude/latitude or native map coordinates.}
}
\value{
An R data frame.
}
\description{
Compute estimated length of 'road.segments' in meters.
}
