% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentHighlight.R
\name{segmentHighlight}
\alias{segmentHighlight}
\title{Highlight segment by ID.}
\usage{
segmentHighlight(id, highlight = TRUE, col = "red", rotate.label = FALSE,
  latlong = FALSE)
}
\arguments{
\item{id}{Character. Segment ID: a concatenation of a street's numeric ID, a whole number between 1 and 528, and a second number to identify the segment.}

\item{highlight}{Logical. Color segment.}

\item{col}{Character. Highlight color.}

\item{rotate.label}{Logical. Rotate segment ID label.}

\item{latlong}{Logical. Use estimated longitude and latitude.}
}
\value{
A base R graphics segment(s).
}
\description{
Highlight segment by ID.
}
\examples{
streetNameLocator("Soho Square", zoom = TRUE, highlight = FALSE)
ids <- road.segments[road.segments$name == "Soho Square", "id"]
invisible(lapply(ids, function(x) segmentHighlight(x, highlight = FALSE)))
}
