% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongAuxilliaryFunctions.R
\name{segmentTheta}
\alias{segmentTheta}
\title{Angle between road segments.}
\usage{
segmentTheta(id1 = "297-1", id2 = "290-1")
}
\arguments{
\item{id1}{Character. First segment ID.}

\item{id2}{Character. Second segment ID.}
}
\description{
For latitude longitude audit.
}
