% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectPump.R
\name{selectPump}
\alias{selectPump}
\title{Select desired pump neighborhoods.}
\usage{
selectPump(pump.select = NULL, metric = "walking", vestry = FALSE)
}
\arguments{
\item{pump.select}{Numeric. Vector of numeric pump IDs to define pump neighborhoods (i.e., the "population"). Negative selection possible. \code{NULL} selects all pumps.}

\item{metric}{Character. "eucldidean" or "walking".}

\item{vestry}{Logical. \code{TRUE} uses the 14 pumps from the Vestry Report. \code{FALSE} uses the 13 in the original map.}
}
\description{
Allows negative selection.
}
