% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim.ortho.proj}
\alias{sim.ortho.proj}
\title{Road "address" of simulated (i.e., "expected") cases.}
\format{A data frame with 6 variables that records the "address" of 20007 simulate cases along the network of roads.
\describe{
    \item{\code{road.segment}}{"address" road segment}
    \item{\code{x.proj}}{x-coordinate}
    \item{\code{y.proj}}{y-coordinate}
    \item{\code{dist}}{Euclidean or orthogonal distance to home road segment}
    \item{\code{type}}{type of projection: Euclidean ("eucl") or orthogonal ("ortho")}
    \item{\code{case}}{numeric case ID}
 }}
\usage{
sim.ortho.proj
}
\description{
Road "address" of simulated (i.e., "expected") cases.
}
\note{
\code{\link{simulateFatalities}} documents the code for these data.
}
\keyword{datasets}
