% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim.pump.case}
\alias{sim.pump.case}
\title{List of 4972 regular "simulated" cases grouped by pump neighborhood.}
\format{A list with 13 vectors.
\describe{
 \item{\code{sim.pump.case}}{numerical ID}
}}
\usage{
sim.pump.case
}
\description{
List of 4972 regular "simulated" cases grouped by pump neighborhood.
}
\section{Notes}{
 \code{\link{neighborhoodWalking}} documents the code for these data.

For details, see \code{vignette}("pump.neighborhoods").
}

\examples{
neighborhoodWalking()$sim.pump.case
}
\keyword{datasets}
