% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim.walking.distance}
\alias{sim.walking.distance}
\title{Walking distance to pump.}
\format{A list of 12 data frames with 5 variables.
\describe{
  \item{\code{case}}{case ID}
  \item{\code{pump}}{pump ID}
  \item{\code{pump.name}}{pump name}
  \item{\code{distance}}{walking distance in meters}
  \item{\code{time}}{walking time in seconds based on 5 km/hr walking speed}
}}
\usage{
sim.walking.distance
}
\description{
Walking distance to pump.
}
\keyword{datasets}
