% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatedCases.R
\name{simulatedCases}
\alias{simulatedCases}
\title{Generate simulated cases and their orthogonal projections.}
\usage{
simulatedCases(compute = FALSE, multi.core = FALSE, simulated.obs = 5000L)
}
\arguments{
\item{compute}{Logical. TRUE computes data. FALSE uses pre-computed data. For replication of data used in the package, FALSE is the default.}

\item{multi.core}{Logical or Numeric. TRUE uses parallel::detectCores(). FALSE uses one, single core. With Numeric, you specify the number logical cores (rounds with as.integer()). On Windows, only "multi.core = FALSE" is available.}

\item{simulated.obs}{Numeric. Number of sample cases.}
}
\value{
An R list with ortho.proj.sp and regular.cases
}
\description{
Places regularly spaced "simulated" cases across the face of the map. These cases are then orthogonally projected onto the road network, which are used to generate "expected" cases, paths and neighborhoods. The function relies on sp::spsample() and sp::Polygon().
}
\section{Notes}{
 This function is computationally intensive. Documents code used to generate the pre-computed data in the package.
}

\seealso{
\code{ortho.proj.sp} and \code{regular.cases}
}
