% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateWalkingDistance.R
\name{simulateWalkingDistance}
\alias{simulateWalkingDistance}
\title{Compute walking distance for simulated cases.}
\usage{
simulateWalkingDistance(pump.select = 7, multi.core = FALSE,
  dev.mode = FALSE, compute = FALSE)
}
\arguments{
\item{pump.select}{Numeric.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores.}

\item{dev.mode}{Logical. Development mode uses parallel::parLapply().}

\item{compute}{Logical.}
}
\description{
Compute walking distance for simulated cases.
}
\note{
This function is computationally intensive. See \code{vignette("Parallelization")} for details. This functions document the code that generates \code{\link{sim.walking.distance}}.
}
