% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simWalkingDistance.R
\name{simWalkingDistance}
\alias{simWalkingDistance}
\title{Compute walking distance for simulated cases.}
\usage{
simWalkingDistance(pump.select = 7)
}
\arguments{
\item{pump.select}{Numeric.}
}
\description{
Compute walking distance for simulated cases.
}
\note{
This function documents sim.walking.distance data frame.
}
