% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snow.neighborhood}
\alias{snow.neighborhood}
\title{Snow neighborhood cases.}
\format{A vector with 384 observations.
\describe{
  \item{\code{snow.neighborhood}}{numeric case ID}
}}
\usage{
snow.neighborhood
}
\description{
Numeric IDs of cases that fall within Snow's Broad Street pump neighborhood. Taken from from Dodson and Tobler.
}
\keyword{datasets}

