% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snow.time.series}
\alias{snow.time.series}
\title{Snow Time Series Data.}
\format{A data frame with 43 observations and 3 variables.
\describe{
  \item{\code{date}}{calendar date}
  \item{\code{fatalities}}{fatal attacks}
  \item{\code{deaths}}{daily deaths}
}}
\usage{
snow.time.series
}
\description{
Time series of Fatalities from Snow's contribution to the Vestry Report (p. 117). Useful for studying the effect of removing the handle on the Broad Street pump on 08 September 1854.
}
\keyword{datasets}

