% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snow.trimmed.segments}
\alias{snow.trimmed.segments}
\title{"Trimmed" Snow neighborhood road segments.}
\format{A data frame with 6 variables and 384 observations.
\describe{
  \item{\code{road.segment}}{road segment ID}
  \item{\code{x1}}{first x-coordinate}
  \item{\code{y1}}{first y-coordinate}
  \item{\code{x2}}{second x-coordinate}
  \item{\code{y2}}{second y-coordinate}
  \item{\code{trimmed}}{trimmed road segment}
}}
\usage{
snow.trimmed.segments
}
\description{
The road segments (i.e., endpoints) that describe Snow's Broad Street pump neighborhood (Vestry Report 1855b). Note that "outer" segments are trimmed: only the portion of the segment transversed by the walking path of the most "distant" case on that segment is included.
}
\keyword{datasets}

