% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowColors.R
\name{snowColors}
\alias{snowColors}
\title{Create a set of colors for pump neighborhoods.}
\usage{
snowColors(vestry = FALSE, palette = "RColorBrewer")
}
\arguments{
\item{vestry}{Logical. \code{TRUE} uses the 14 pumps in the Vestry Report. \code{FALSE} uses the original 13.}

\item{palette}{Character. "RColorBrewer" or "viridis".}
}
\value{
A character vector of colors.
}
\description{
Uses \code{RColorBrewer::brewer.pal()}.
}
\note{
Built with 'RColorBrewer' package.
}
