% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowMap.R
\name{snowMap}
\alias{snowMap}
\title{Plot John Snow's cholera map.}
\usage{
snowMap(add.landmarks = TRUE, vestry = FALSE, stacked = TRUE)
}
\arguments{
\item{add.landmarks}{Logical. Add landmarks.}

\item{vestry}{Logical. Use the 14 pumps from the Vestry Report.}

\item{stacked}{Logical. Use stacked fatalities.}
}
\value{
A base R graphics plot.
}
\description{
Plots Dodson and Tober's digitization of Snow's map and landmarks
}
\examples{
snowMap()
snowMap(vestry = TRUE, stacked = FALSE)
}
\seealso{
\code{\link[cholera]{addLandmarks}}
}

