% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowNeighborhoodB.R
\name{snowNeighborhoodB}
\alias{snowNeighborhoodB}
\title{Plotting data for Snow's graphical annotation of the Broad Street pump neighborhood.}
\usage{
snowNeighborhoodB(latlong = TRUE)
}
\arguments{
\item{latlong}{Logical. Use estimated longitude and latitude.}
}
\value{
An R list of edge IDs and simulated case IDs.
}
\description{
Computes "missing" and split road segments data, and area plot data.
}
