% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sohoGraph.R
\name{sohoGraph}
\alias{sohoGraph}
\title{Compute network graph of roads and embed cases, landmarks and pumps (prototype).}
\usage{
sohoGraph(vestry = FALSE, case.set = "observed", embed.anchors = TRUE,
  embed.landmarks = TRUE, embed.pumps = TRUE, latlong = FALSE,
  drop.isolates = FALSE, ellipsoid = "WGS", multi.core = FALSE)
}
\arguments{
\item{vestry}{Logical. Use Vestry Report pump data.}

\item{case.set}{Character. "observed", "expected", or "snow". "snow" captures John Snow's annotation of the Broad Street pump neighborhood printed in the Vestry report version of the map.}

\item{embed.anchors}{Logical. Embed all 321 or selected anchor cases into graph network.}

\item{embed.landmarks}{Logical or Numeric. Embed all or selected landmarks into road network.}

\item{embed.pumps}{Logical or Numeric. Embed all or selected pumps into road network.}

\item{latlong}{Logical. Use estimated longitude and latitude.}

\item{drop.isolates}{Logical. Exclude Adam and Eve Court (and Pump #2) and Falconberg Court and Mews.}

\item{ellipsoid}{Character. "WGS" for WGS-84 or "BNG" for British National Grid (i.e., Airy 1830).}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. See \code{vignette("Parallelization")} for details. Useful with `case.set = "expected"`.}
}
\value{
An R list of nodes, edges and an 'igraph' network graph.
}
\description{
Compute network graph of roads and embed cases, landmarks and pumps (prototype).
}
