% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latitudeLongitudeFatality.R
\name{stackCaseCount}
\alias{stackCaseCount}
\title{Count of non-anchor cases by anchor.}
\usage{
stackCaseCount()
}
\description{
Cases at address, excluding anchor; in the stack, excluding the base case.
}
