% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi.R
\name{starGraph}
\alias{starGraph}
\title{Plot all Euclidean paths.}
\usage{
starGraph(x)
}
\arguments{
\item{x}{An object of class "voronoi" created by neighborhoodVoronoi().}
}
\value{
A base R graph.
}
\description{
Plot all Euclidean paths.
}
