% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starPlot.R
\name{starPlot}
\alias{starPlot}
\title{Plot all Euclidean paths from pumps to cases.}
\usage{
starPlot(pump.select = NULL, vestry = FALSE, multi.core = FALSE)
}
\arguments{
\item{pump.select}{Numeric. Vector of numeric pump IDs to define pump neighborhoods (i.e., the "population"). Negative selection possible. NULL selects all pumps.}

\item{vestry}{Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 in the original map.}

\item{multi.core}{Logical or Numeric. TRUE uses parallel::detectCores(). FALSE uses one, single core. You can also specify the number logical cores. On Window, only "multi.core = FALSE" is available.}
}
\value{
A base R graph.
}
\description{
Plots a star graph from pump to its cases.
}
