% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitMeter.R
\name{stLength}
\alias{stLength}
\title{Compute total length of roads.}
\usage{
stLength(dat, input.unit = "nominal")
}
\arguments{
\item{dat}{Object. data frame.}

\item{input.unit}{Character. "nominal" or "latlong".}
}
\description{
Compute total length of roads.
}
