% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latitudeLongitudeFatality.R
\name{stratifiedFatalities}
\alias{stratifiedFatalities}
\title{Case by strata and strata numeric ID.}
\usage{
stratifiedFatalities()
}
\description{
Cases at address, excluding anchor; in the stack, excluding the base fatality.
}
