% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetHighlight.R
\name{streetHighlight}
\alias{streetHighlight}
\title{Highlight road by name.}
\usage{
streetHighlight(road.name, col = "red", lwd = 3, latlong = FALSE)
}
\arguments{
\item{road.name}{Character vector. The function tries to correct for case and remove extra spaces (includes "Map Frame").}

\item{col}{Character. Highlight color.}

\item{lwd}{Numeric. Line width.}

\item{latlong}{Logical. Use estimated longitude and latitude.}
}
\value{
A base R graphics segment(s).
}
\description{
Highlight road by name.
}
\examples{
snowMap()
streetHighlight("Broad Street")
}
