% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetNames.R
\name{streetNames}
\alias{streetNames}
\title{Street names (alphabetized).}
\usage{
streetNames()
}
\value{
An R character vector.
}
\description{
Unique road names from Snow's cholera map.
}
\note{
See vignette("roads"), and roads and road.segment data frames.
}
