% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetNameViewer.R
\name{streetNameViewer}
\alias{streetNameViewer}
\title{Locate road by name.}
\usage{
streetNameViewer(road.name, zoom = FALSE, radius = 1)
}
\arguments{
\item{road.name}{Character vector. Note that \code{streetNameViewer}() tries to correct for case and removes extra spaces.}

\item{zoom}{Logical.}

\item{radius}{Numeric. Controls the degree of zoom. For "radius" <= 5, the anchor case number is plotted.}
}
\value{
A base R graphics plot.
}
\description{
Plots John Snow's map of the 1854 London cholera outbreak and highlights the
selected road. See the list of road names in \code{vignette}("road.names").
}
\examples{
streetNameViewer("Oxford Street")
streetNameViewer("oxford street")
streetNameViewer("Cambridge Street", zoom = TRUE)
streetNameViewer("Cambridge Street", zoom = TRUE, radius = 0)
}
\seealso{
\code{roads}

\code{vignette}("road.names")

\code{\link[cholera]{streetNumberViewer}}
}

