% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetNumberLocator.R
\name{streetNumberLocator}
\alias{streetNumberLocator}
\title{Locate road by its numerical ID.}
\usage{
streetNumberLocator(road.number, zoom = FALSE, radius = 1)
}
\arguments{
\item{road.number}{Numeric or integer. A whole number between 1 and 528.}

\item{zoom}{Logical.}

\item{radius}{Numeric. Controls the degree of zoom. For values <= 5, the anchor case number is plotted.}
}
\value{
A base R graphics plot.
}
\description{
Plots John Snow's map of the 1854 London cholera outbreak and highlights the
selected road. See cholera::roads for numerical IDs and \code{vignette}("road.names") for details.
}
\examples{
streetNumberLocator(243)
streetNumberLocator(243, zoom = TRUE)
streetNumberLocator(243, zoom = TRUE, radius = 0)
}
\seealso{
\code{\link{roads}}, \code{\link{road.segments}}, \code{\link{streetNameLocator}}, \code{vignette("road.names")}
}
