% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetNumberViewer.R
\name{streetNumberViewer}
\alias{streetNumberViewer}
\title{Locate road by its numerical ID.}
\usage{
streetNumberViewer(road.number, zoom = FALSE, radius = 1)
}
\arguments{
\item{road.number}{Numeric or integer. A whole number between 1 and 528.}

\item{zoom}{Logical.}

\item{radius}{Numeric. Controls the degree of zoom. For values <= 5, the anchor case number is plotted.}
}
\value{
A base R graphics plot.
}
\description{
Plots John Snow's map of the 1854 London cholera outbreak and highlights the
selected road. See \code{vignette}("road.names") for details.
}
\examples{
streetNumberViewer(243)
streetNumberViewer(243, zoom = TRUE)
streetNumberViewer(243, zoom = TRUE, radius = 0)
}
\seealso{
\code{roads}

\code{\link[cholera]{streetNameViewer}}
}

