% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetNameViewer.R
\name{streetViewer}
\alias{streetViewer}
\title{Plot road by name.}
\usage{
streetViewer(road.name, zoom = TRUE, radius = 1)
}
\arguments{
\item{road.name}{Character vector. To match names in the data set, \code{streetNameViewer}() corrects for case and removes extraneous spaces.}

\item{zoom}{Logical.}

\item{radius}{Numeric. Controls the degree of zoom. Default value is 1. For values <= 5, the anchor case number is plotted rather than a point for each case.}
}
\value{
A base R graphics plot.
}
\description{
Diagnostic plot with appropriate defaults.
}

