% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latitudeLongitudeFrame.R
\name{subsetFramePDF}
\alias{subsetFramePDF}
\title{Create PDFs of map frame points.}
\usage{
subsetFramePDF(path, pch = 15, cex = 0.2)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/".}

\item{pch}{Integer. R pch.}

\item{cex}{Numeric.}
}
\description{
For QGIS geo-referencing.
}
