% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongAuxilliaryFunctions.R
\name{subsetPDF}
\alias{subsetPDF}
\title{Create subsetted PDFs (prototype).}
\usage{
subsetPDF(path, dataset = "fatalities.address")
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/"}

\item{dataset}{Object. 'cholera' dataset.}
}
\description{
Reduce over-printing of points.
}
