% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latitudeLongitudeRoads.R
\name{subsetRoadsPDF}
\alias{subsetRoadsPDF}
\title{Create PDFs road segment endpoints stratified by number of intersections.}
\usage{
subsetRoadsPDF(path)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/"}
}
\description{
For QGIS geo-referencing.
}
