% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclideanNominal.R
\name{summary.euclidean}
\alias{summary.euclidean}
\title{Summary method for neighborhoodEuclidean().}
\usage{
\method{summary}{euclidean}(object, ...)
}
\arguments{
\item{object}{Object. An object of class "euclidean" created by \code{neighborhoodEuclidean()}.}

\item{...}{Additional parameters.}
}
\value{
A vector of counts by neighborhood.
}
\description{
Return computed counts for Euclidean neighborhoods.
}
\examples{
\dontrun{
summary(neighborhoodEuclidean())
}
}
