% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclideanDistance.R
\name{summary.euclidean_distance}
\alias{summary.euclidean_distance}
\title{Summary method for euclideanDistance().}
\usage{
\method{summary}{euclidean_distance}(object, ...)
}
\arguments{
\item{object}{An object of class "euclidean_distance" created by euclideanDistance().}

\item{...}{Additional parameters.}
}
\value{
An R data frame.
}
\description{
Summary method for euclideanDistance().
}
\examples{
summary(euclideanDistance(1))
}
