% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclideanLatlong.R
\name{summary.euclideanLatlong}
\alias{summary.euclideanLatlong}
\title{Summary method for neighborhoodEuclidean(latlong = TRUE).}
\usage{
\method{summary}{euclideanLatlong}(object, ...)
}
\arguments{
\item{object}{Object. An object of class "euclidean" created by \code{neighborhoodEuclidean(latlong = TRUE)}.}

\item{...}{Additional parameters.}
}
\value{
A vector of counts by neighborhood.
}
\description{
Return computed counts for Euclidean neighborhoods.
}
\examples{
\dontrun{
summary(neighborhoodEuclidean(latlong = TRUE))
}
}
