% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow.R
\name{summary.snow}
\alias{summary.snow}
\title{Summary method for neighborhoodSnow().}
\usage{
\method{summary}{snow}(object, ...)
}
\arguments{
\item{object}{Object. An object of class "snow" created by \code{neighborhoodSnow()}.}

\item{...}{Additional parameters.}
}
\value{
A data frame of counts inside and outside of Snow's neighborhood.
}
\description{
Return computed counts for Snow neighborhood.
}
\examples{
\dontrun{
summary(neighborhoodSnow())
}
}
