% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoiNominal.R
\name{summary.voronoi_nominal}
\alias{summary.voronoi_nominal}
\title{Summary method for voronoiNominal().}
\usage{
\method{summary}{voronoi_nominal}(object, ...)
}
\arguments{
\item{object}{Object. An object of class "voronoi" created by \code{voronoiNominal()}.}

\item{...}{Additional arguments.}
}
\value{
A vector of counts by neighborhood.
}
\description{
Return computed counts for Voronoi neighborhoods.
}
\seealso{
\code{addVoronoi()}
\code{plot.voronoi()}
}
