% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkingNominal.R
\name{summary.walking}
\alias{summary.walking}
\title{Summary method for walkingNominal().}
\usage{
\method{summary}{walking}(object, ...)
}
\arguments{
\item{object}{Object. An object of class "walking" created by \code{walkingNominal()}.}

\item{...}{Additional parameters.}
}
\value{
An R vector.
}
\description{
Return computed counts for walking neighborhoods.
}
