% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanakaContourPlot.R
\name{tanakaContourPlot}
\alias{tanakaContourPlot}
\title{Tanaka contour plot.}
\usage{
tanakaContourPlot(add = FALSE, alpha.level = 1)
}
\arguments{
\item{add}{Logical. Add to exisiting plot.}

\item{alpha.level}{Numeric.}
}
\description{
Soho elevation data.
}
