% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionAddresses.R
\name{thresholdAddressGraph}
\alias{thresholdAddressGraph}
\title{Network graph of addresses with selected inter-point distance.}
\usage{
thresholdAddressGraph(inter.point.dist = 0.15)
}
\arguments{
\item{inter.point.dist}{Numeric. Ceiling for overlapping points.}
}
\value{
An 'igraph' object.
}
\description{
Network graph of addresses with selected inter-point distance.
}
