% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongFrame.R
\name{thresholdFrameGraph}
\alias{thresholdFrameGraph}
\title{Network graph of frame points with selected inter-point distance.}
\usage{
thresholdFrameGraph(inter.point.dist = 0.15)
}
\arguments{
\item{inter.point.dist}{Numeric. Ceiling for overlapping points.}
}
\value{
An 'igraph' object.
}
\description{
Network graph of frame points with selected inter-point distance.
}
