% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{time.series.snow}
\alias{time.series.snow}
\title{Snow time series data.}
\format{A data frame with 43 observations and 3 variables.
\describe{
  \item{\code{date}}{calendar date}
  \item{\code{fatal.attacks}}{fatal attacks}
  \item{\code{deaths}}{daily deaths}
}}
\usage{
time.series.snow
}
\description{
Time series of recorded fatalities from Snow's contribution to the Vestry Report (p. 117). Useful for studying the effect of removing the handle from the Broad Street pump on 08 September 1854.
}
\seealso{
\code{
\code{\link{plotTimeSeries}},
\link{time.series.vestry}},
\code{vignette}("time.series")
}
\keyword{datasets}
