% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{time.series.vestry}
\alias{time.series.vestry}
\title{Vestry Report Time Series Data.}
\format{An R list of class "time.series" with two elements: "source", which is "vestry", and "data", a data frame with 61 observations and 3 variables.
\describe{
  \item{\code{date}}{calendar date}
  \item{\code{fatal.attacks}}{fatal attacks}
  \item{\code{deaths}}{deaths}
}}
\usage{
time.series.vestry
}
\description{
Time series of recorded fatalities from Vestry Report Appendix B. Useful for studying the effect of removing the handle from the Broad Street pump on 08 September 1854.
}
\seealso{
\code{\link{plot.time.series}},
\link{time.series.snow},
\code{vignette}("time.series")
}
\keyword{datasets}
