% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timePosts.R
\name{timePosts}
\alias{timePosts}
\title{Compute the coordinates of walking path timeposts.}
\usage{
timePosts(pump.select, timepost.interval = 60, multi.core = FALSE)
}
\arguments{
\item{pump.select}{Numeric.}

\item{timepost.interval}{Numeric. Seconds.}

\item{multi.core}{Logical or Numeric. TRUE uses parallel::detectCores(). FALSE uses one, single core. You can also specify the number logical cores. On Window, only "multi.core = FALSE" is available.}
}
\value{
An R list.
}
\description{
In-progress prototype.
}
