% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimPaths.R
\name{trimPaths}
\alias{trimPaths}
\title{Trim computed walking paths}
\usage{
trimPaths(pump.select = NULL, vestry = FALSE, obs = TRUE,
  multi.core = FALSE, save.file = FALSE)
}
\arguments{
\item{pump.select}{Default is NULL: all pumps are used. Ortherwise, selection by a vector of numeric IDs: 1 to 13 for \code{pumps}; 1 to 14 for \code{pumps.vestry}.}

\item{vestry}{Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 in the original map.}

\item{obs}{Logical. TRUE uses paths of observed cases. FALSE uses paths of simulated cases.}

\item{multi.core}{Logical. TRUE uses parallel::detectCores(). FALSE uses one, single core.}

\item{save.file}{Logical. TRUE save output to working directory.}
}
\value{
A list of data frames.
}
\description{
To reduce duplication when plotting, this function computes the unique road segments of all walking paths. Note that only "anchor cases" are used.
}
\examples{
# trimPaths()
# trimPaths(6:7)
# trimPaths(-6)
# trimPaths(vestry = TRUE)
}

