% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimPlot.R
\name{trimPlot}
\alias{trimPlot}
\title{Plots the 13 expected walking neighborhoods.}
\usage{
trimPlot(streets = TRUE, add.landmarks = TRUE, color = snowColors())
}
\arguments{
\item{streets}{Logical. TRUE plots neighborhoods by street. FALSE plots orthogonal neighborhoods (area).}

\item{add.landmarks}{Logical. Include landmarks.}

\item{color}{Character. Uses snowColor().}
}
\value{
A base R graphics plot.
}
\description{
Neighborhoods are based on the shortest paths between a fatality's address and its nearest pump.
}
\examples{
trimPlot()
trimPlot(streets = FALSE)
}

