% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{voronoi.polygons}
\alias{voronoi.polygons}
\title{Coordinates of Voronoi polygon vertices.}
\format{
A list of 13 data frames frames with 2 variables.
\describe{
  \item{\code{x}}{x-coordinate}
  \item{\code{y}}{y-coordinate}
}
}
\usage{
voronoi.polygons
}
\description{
Coordinates of Voronoi polygon vertices.
}
\keyword{datasets}
