% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongVoronoi.R
\name{voronoiPDF}
\alias{voronoiPDF}
\title{Create PDFs of Voronoi Polygon Vertices.}
\usage{
voronoiPDF(path, vestry = FALSE, pch = 46, cex = 1)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/"}

\item{vestry}{Logical.}

\item{pch}{Integer. R pch.}

\item{cex}{Numeric.}
}
\description{
For QGIS georeferencing.
}
