% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoiPlot.R
\name{voronoiPlot}
\alias{voronoiPlot}
\title{Plot Voronoi diagram.}
\usage{
voronoiPlot(select = NULL, vestry = FALSE, output = NULL)
}
\arguments{
\item{select}{Numeric. Default is NULL and all pumps are used. Ortherwise, selection by a vector of numeric IDs: 1 to 13 for \code{pumps}; 1 to 14 for \code{pumps.vestry}.}

\item{vestry}{Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 in the original map.}

\item{output}{Character. NULL plots address points. "addresses" plots the total count of addresses at pumps' locations. "fatalities" plots the total count of fatalities at pumps' locations.}
}
\value{
A base R graphics plot.
}
\description{
Plot  Voronoi diagrams of John Snow's 1854 London cholera data.
}
\examples{
voronoiPlot()
voronoiPlot(output = "addresses")
voronoiPlot(vestry = TRUE)
voronoiPlot(select = 6:7)
voronoiPlot(select = -6)
voronoiPlot(select = -6, output = "fatalities")
}
\seealso{
\code{addVoronoi()}
}

