% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoiPolygons.R
\name{voronoiPolygons}
\alias{voronoiPolygons}
\title{Compute vertices of Voronoi tiles.}
\usage{
voronoiPolygons(sites, rw.data = NULL, rw = NULL)
}
\arguments{
\item{sites}{Object. Data frame of sites to construct Voronoi diagram with variables "x" and "y".}

\item{rw.data}{Object. Data frame of secondary source of data for rectangular window or bounding box: observations, cases, addresses, etc. with variables "x" and "y".}

\item{rw}{Numeric. Vector of corners of the rectangular window or bounding box: xmin, xmax, ymin, ymax. For deldir::deldir().}
}
\value{
A list of data frames.
}
\description{
For use with polygon functions.
}
\examples{
polygon.vertices <- voronoiPolygons(cholera::pumps)
cholera::snowMap()
invisible(lapply(polygon.vertices, polygon))

polygon.vertices <- voronoiPolygons(cholera::pumps, cholera::roads)
cholera::snowMap()
invisible(lapply(polygon.vertices, polygon))
}
