% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkingDistance2.R
\name{walkingDistance2}
\alias{walkingDistance2}
\title{Compute the walking distance between cases and/or pumps.}
\usage{
walkingDistance2(origin, destination = NULL, type = "case-pump",
  weighted = TRUE, vestry = FALSE, unit = NULL)
}
\arguments{
\item{origin}{Numeric or Integer. Numeric ID of case or pump.}

\item{destination}{Numeric or Integer. Numeric ID(s) of case(s) or pump(s). Exclusion is possible via negative selection (e.g., -7). Default is NULL: this returns closest pump or "anchor" case.}

\item{type}{Character "case-pump", "cases" or "pumps".}

\item{weighted}{Logical. TRUE computes shortest path in terms of road length. FALSE computes shortest path in terms of nodes.}

\item{vestry}{Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 in the original map.}

\item{unit}{Character. Unit of measurement: "meter" or "yard". Default is NULL, which returns the map's native scale. Meaningful only when "weighted" is TRUE. See \code{vignette("roads")} for information on unit distances.}
}
\value{
A base R data frame.
}
\description{
(beta v.1)
}
\note{
The function uses a case's "address" or "anchor" case to compute distance. Because Adam and Eve Court is disconnected from the larger road network (an isolate), only cases on that road can reach pump 2. All others will return Inf.
}
\examples{
walkingDistance2(1)
walkingDistance2(1, 6) # distance from case 1 to pump 6.
walkingDistance2(1, -7) # exclude pump 7 from consideration.
walkingDistance2(1, 6, type = "pumps") # distance from pump 1 to pump 2.
walkingDistance2(1, 6, type = "cases") # distance from case 1 to case 2.

## Pairwise Euclidean distance (meters) between pumps. ##
# pairs <- combn(cholera::pumps$id, 2, simplify = FALSE)
#
# vapply(pairs, function(x) {
#   walkingDistance2(x[1], x[2], type = "pumps", unit = "meter")$distance
# }, numeric(1L))
}
\seealso{
\code{\link{fatalities}}, \code{\link{simulateFatalities}}, \code{vignette("pump.neighborhoods")}
}
