% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkingNeighborhood.R
\name{walkingNeighborhood}
\alias{walkingNeighborhood}
\title{Plots shortest path walking neighborhoods.}
\usage{
walkingNeighborhood(selection = NULL, vestry = FALSE, streets = TRUE,
  weighted = TRUE, add.obs = FALSE, add.landmarks = TRUE, arrow = FALSE,
  alpha.level = NULL, obs.lwd = 2, sim.lwd = 2, multi.core = FALSE)
}
\arguments{
\item{selection}{Numeric. Default is NULL; all pumps are used. Otherwise, selection by a vector of numeric IDs: 1 to 13 for \code{pumps}; 1 to 14 for \code{pumps.vestry}}

\item{vestry}{Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 in the original map.}

\item{streets}{TRUE plots neighborhoods by street. FALSE plots orthogonal neighborhoods.}

\item{weighted}{Logical. TRUE uses shortest path weighted by road distance.}

\item{add.obs}{Logical. Include observed addresses.}

\item{add.landmarks}{Logical. Include landmarks.}

\item{arrow}{Logical. Use arrows.}

\item{alpha.level, }{Numeric. If desired, set alpha level.}

\item{obs.lwd}{Numeric. Set line width for observed paths.}

\item{sim.lwd}{Numeric. Set line width for expected paths}

\item{multi.core}{Logical. TRUE uses parallel::mclapply and parallel::detectCores(). Note that you must run this in the terminal or in batch mode (i.e., not in interactive mode in the GUI). See parallel::mclapply() for details. FALSE uses 1 logical core and can be run in GUI.}
}
\value{
A base R graphics plot.
}
\description{
Plots walking neighborhoods based on the weighted shortest paths between a fatality's address and its nearest pump.
}
\details{
Currently computationally intensive (streets = TRUE appox. 90 seocnds; streets = FALSE appox. 75 seconds on Intel Core i7). For better performance, run in batch mode in parallel with multiple cores (appox. 45 and 30 seconds). See \code{vignette}("walking.neighborhoods") for details.
}
\examples{
# walkingNeighborhoodPlot()
# walkingNeighborhoodPlot(streets = FALSE)
# walkingNeighborhoodPlot(streets = FALSE, weighted = FALSE)
}
\seealso{
\code{addLandmarks()}
}

