% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkingNeighborhoodPlot.R
\name{walkingNeighborhoodPlot}
\alias{walkingNeighborhoodPlot}
\title{Plots walking neighborhoods.}
\usage{
walkingNeighborhoodPlot(selection = NULL, vestry = FALSE, streets = TRUE,
  weighted = TRUE, add.obs = FALSE, add.landmarks = TRUE, cores = 1L)
}
\arguments{
\item{selection}{Numeric. Default is NULL; all pumps are used. Otherwise, selection by a vector of numeric IDs: 1 to 13 for \code{pumps}; 1 to 14 for \code{pumps.vestry}}

\item{vestry}{Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 in the original map.}

\item{streets}{TRUE plots neighborhoods by street. FALSE plots orthogonal neighborhoods.}

\item{weighted}{Logical. Shortest path weighted by road distance.}

\item{add.obs}{Logical. Include observed addresses.}

\item{add.landmarks}{Logical. Include landmarks.}

\item{cores}{Numeric or "all". Number of cores to use; default is 1. If you select more than 1 core or use "all", which uses parallel::detectCores(), you should run this function in batch, not interactive mode (i.e., not in GUI).}
}
\value{
A base R graphics plot.
}
\description{
Plots walking neighborhoods based on the weighted shortest paths between a fatalities and the nearest pump. Beta version.
}
\details{
Currently computationally intensive. For now, it is better to run in batch mode in parallel with multiple cores. See \code{vignette}("walking.neighborhoods") for details
}
\seealso{
\code{addLandmarks()}
}

