% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkingPathsPlot.R
\name{walkingPathPlot}
\alias{walkingPathPlot}
\title{Plot shortest walking path.}
\usage{
walkingPathPlot(x, zoom = TRUE, radius = 0.5, weighted = TRUE,
  vestry = FALSE, selection = NULL, use.arrows = TRUE)
}
\arguments{
\item{x}{Numeric or Integer. Whole number between 1 and 578 for fatality.}

\item{zoom}{Logical.}

\item{radius}{Numeric. Controls the degree of zoom.}

\item{weighted}{Logical. Shortest path weighted by road distance.}

\item{vestry}{Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 in the original map.}

\item{selection}{Numeric. Default is NULL and all pumps are used. Ortherwise, selection by a vector of numeric IDs: 1 to 13 for \code{pumps}; 1 to 14 for \code{pumps.vestry}.}

\item{use.arrows}{Logical. TRUE uses arrows. FALSE uses segments}
}
\value{
A base R graphics plot.
}
\description{
Plots the weighted shortest walking path between a fatality and its nearest pump.
}
\examples{
walkingPathPlot(1)
walkingPathPlot(1, selection = -7)
}
\seealso{
\code{fatalities}

\code{fatalities.address}
}

