% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winterTemperatures.R
\name{winterTemperatures}
\alias{winterTemperatures}
\title{Average Winter Temperatures.}
\usage{
winterTemperatures()
}
\description{
Gareth Stedman Jones Appendix 2, Table 12, p.384.
}
\examples{
plot(winterTemperatures(), "1859-6-1")
}
