% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withinRadius.R
\name{withinRadius}
\alias{withinRadius}
\title{Test whether point "b" is within a given radius of point "a".}
\usage{
withinRadius(a, b, radius = 2)
}
\arguments{
\item{a}{Numeric. Data frame of x-y coordinates.}

\item{b}{Numeric. Data frame of x-y coordinates.}

\item{radius}{Numeric.}
}
\description{
Test whether point "b" is within a given radius of point "a".
}
