% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCases.R
\name{addCases}
\alias{addCases}
\title{Add observed cases.}
\usage{
addCases(pump.subset = NULL, pump.select = NULL, type = "address",
  vestry = FALSE, weighted = TRUE, case.color = NULL,
  multi.core = FALSE, ...)
}
\arguments{
\item{pump.subset}{Numeric. Vector of pumps to select (subset) from neighborhoods defined by "pump.select". Negative selection possible. NULL selects all pumps in "pump.select".}

\item{pump.select}{Numeric. Numeric vector of pumps to define pump neighborhoods (i.e. the "population"). Negative selection possible. NULL selects all pumps.}

\item{type}{Character. Type of case: "address" (base of stack) or "fatalities" (entire stack).}

\item{vestry}{Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 in the original map.}

\item{weighted}{Logical. TRUE computes shortest path weighted by road length. FALSE computes shortest path in terms of the number of nodes.}

\item{case.color}{Character. Use a single color for all paths. NULL uses neighborhood colors defined by snowColors().}

\item{multi.core}{Logical or Numeric. TRUE uses parallel::detectCores(). FALSE uses one, single core. You can also specify the number logical cores. On Window, only "multi.core = FALSE" is available.}

\item{...}{Additional plotting parameters.}
}
\description{
Add cases, as "address" (cholera::fatalties.address) or stacked "fatalties" (cholera::fatalties), as points to an existing plot.
}
\examples{
snowMap()
addCases(pump.susbet = c(6, 10))

snowMap()
addCases(pump.select = c(6, 10))
}
