% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEuclideanPath.R
\name{addEuclideanPath}
\alias{addEuclideanPath}
\title{Add the path for the Euclidean distance between cases and/or pumps.}
\usage{
addEuclideanPath(origin, destination = NULL, type = "case-pump",
  observed = TRUE, vestry = FALSE, unit = "meter", time.unit = "second",
  walking.speed = 5, unit.posts = "distance", unit.interval = NULL,
  alpha.level = 1)
}
\arguments{
\item{origin}{Numeric or Integer. Numeric ID of case or pump.}

\item{destination}{Numeric or Integer. Numeric ID(s) of case(s) or pump(s). Exclusion is possible via negative selection (e.g., -7). Default is NULL: this returns closest pump or "anchor" case.}

\item{type}{Character "case-pump", "cases" or "pumps".}

\item{observed}{Logical. Use observed or "simulated" expected data.}

\item{vestry}{Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 pumps from the original map.}

\item{unit}{Character. Unit of distance: "meter", "yard" or "native". "native" returns the map's native scale. See \code{vignette("roads")} for information on unit distances.}

\item{time.unit}{Character. "hour", "minute", or "second".}

\item{walking.speed}{Numeric. Default is 5 km/hr.}

\item{unit.posts}{Character. "distance" for mileposts; "time" for timeposts.}

\item{unit.interval}{Numeric. Sets interval between posts: for "distance",}

\item{alpha.level}{Numeric. Alpha level transparency for path: a value in [0, 1].}
}
\value{
An R list with 3 data frames: x-y coordinates for the origin and destination, and a summary of results.
}
\description{
Add the path for the Euclidean distance between cases and/or pumps.
}
\note{
The function uses a case's "address" (i.e., "anchor" case of a stack) to compute distance. Time is computed using distanceTime().
}
