% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowMap.R
\name{addFrame}
\alias{addFrame}
\title{Add map border to plot.}
\usage{
addFrame(latlong = FALSE, col = "black", ...)
}
\arguments{
\item{latlong}{Logical. Use estimated longitude and latitude.}

\item{col}{Character. Color}

\item{...}{Additional plotting parameters.}
}
\description{
Add map border to plot.
}
