% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLandmarks.R
\name{addKernelDensity}
\alias{addKernelDensity}
\title{Add 2D kernel density contours.}
\usage{
addKernelDensity(bandwidth = 0.5, color = "black", line.type = "solid",
  data = NULL)
}
\arguments{
\item{bandwidth}{Numeric. bandwidth for kernel density esitmation. Default is 0.5}

\item{color}{Character. Color of cotour lines. Default is "black".}

\item{line.type}{Character. Line type for contour. Default is "solid"}

\item{data}{Character. Default is NULL uses "fatalities.unstacked". "address" uses "fatalities.address". "stacked" uses "fatalities".}
}
\value{
Add contours to a graphics plot.
}
\description{
Uses KernSmooth::bkde2D().
}
\examples{
snowMap(add.landmarks = FALSE)
addKernelDensity()
}

