% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLandmarks.R
\name{addKernelDensity}
\alias{addKernelDensity}
\title{Add 2D kernel density contours.}
\usage{
addKernelDensity(bandwidth = 0.5, color = "black", line.type = "solid",
  data = NULL)
}
\arguments{
\item{bandwidth}{Numeric. bandwidth for kernel density esitmation.}

\item{color}{Character. Color of cotour lines.}

\item{line.type}{Character. Line type for contour.}

\item{data}{Character. NULL, uses \code{fatalities.unstacked}. "address" uses \code{fatalities.address}. "stacked" uses \code{fatalities}.}
}
\value{
Add contours to a graphics plot.
}
\description{
Uses KernSmooth::bkde2D().
}
\examples{
snowMap(add.landmarks = FALSE)
addKernelDensity()
}

