% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLandmarks.R
\name{addLandmarks}
\alias{addLandmarks}
\title{Add landmarks to plot.}
\usage{
addLandmarks(text.size = 0.5, text.col = "black",
  highlight.perimeter = TRUE, latlong = FALSE)
}
\arguments{
\item{text.size}{Numeric. cex for text labels.}

\item{text.col}{Character. col for text labels.}

\item{highlight.perimeter}{Logical. Highlight Lion Brewery and Model Housing.}

\item{latlong}{Logical. Use estimated longitude and latitude.}
}
\value{
Base R points and text.
}
\description{
Add landmarks to plot.
}
\note{
The location of 18 Sackville Street and 28 Dean Street are approximate. Falconberg Court & Mews form an isolate: they are not part of the network of roads and are technically unreachable. Adam and Eve Court and its pump also form an isolate.
}
\examples{
snowMap(add.landmarks = FALSE)
addLandmarks()
}
