% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMilePosts.R
\name{addMilePosts}
\alias{addMilePosts}
\title{Add distance or time based "mileposts" to an observed walking neighborhood plot.}
\usage{
addMilePosts(pump.subset = NULL, pump.select = NULL, vestry = FALSE,
  unit = "distance", interval = NULL, walking.speed = 5,
  type = "arrows", multi.core = FALSE, dev.mode = FALSE)
}
\arguments{
\item{pump.subset}{Numeric. Vector of numeric pump IDs to subset from the neighborhoods defined by \code{pump.select}. Negative selection possible. \code{NULL} uses all pumps in \code{pump.select}.}

\item{pump.select}{Numeric. Numeric vector of pumps to define possible pump neighborhoods (i.e. the "population"). Negative selection is possible. NULL selects all "observed" pumps (i.e., pumps with at least one case).}

\item{vestry}{Logical. \code{TRUE} uses the 14 pumps from the Vestry Report. FALSE uses the 13 from the original map.}

\item{unit}{Character. Milepost unit of measurement: "distance" or "time".}

\item{interval}{Numeric. Interval between mileposts: 50 meters for "distance";  60 seconds for "time".}

\item{walking.speed}{Numeric. Walking speed in km/hr.}

\item{type}{Character. "arrows" or "points".}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. See \code{vignette("Parallelization")} for details.}

\item{dev.mode}{Logical. Development mode uses parallel::parLapply().}
}
\value{
R base graphics arrows or points.
}
\description{
Add distance or time based "mileposts" to an observed walking neighborhood plot.
}
