% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPlaguePit.R
\name{addPlaguePit}
\alias{addPlaguePit}
\title{Add plague pit (Marshall Street).}
\usage{
addPlaguePit(color = "black", line.type = "solid", ...)
}
\arguments{
\item{color}{Character. Color of polygon.}

\item{line.type}{Character. Polygon line type.}

\item{...}{Additional plotting parameters.}
}
\value{
Adds a polygon to a graphics plot.
}
\description{
Draws a polygon that approximates the plague pit located around Marshall Street. From Vestry Report map.
}
\examples{
snowMap(add.landmarks = FALSE)
addPlaguePit()
}
\seealso{
\code{\link{snowMap}},
\code{\link{addIndexCase}},
\code{\link{addKernelDensity}},
\code{\link{addLandmarks}},
\code{\link{addSnow}},
\code{\link{addVoronoi}},
\code{\link{addWhitehead}}
}
