% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPump.R
\name{addPump}
\alias{addPump}
\title{Add water pump by numerical ID.}
\usage{
addPump(id, vestry = FALSE, col = NULL, pch = 24, label = TRUE,
  pos = 1, ...)
}
\arguments{
\item{id}{Numeric or Integer. With "vestry = TRUE", a whole number between 1 and 14. With "vestry = FALSE", a whole number between 1 and 13. See cholera::pumps.vestry and cholera::pumps for IDs and details about specific pumps.}

\item{vestry}{Logical. TRUE for the 14 pumps from Vestry Report. FALSE for the original 13 pumps.}

\item{col}{Character. Color of point.}

\item{pch}{Numeric. Shape of point character.}

\item{label}{Logical. TRUE adds text label.}

\item{pos}{Numeric. Position of label.}

\item{...}{Additional plotting parameters.}
}
\description{
Add water pump to existing plot.
}
