% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPump.R
\name{addPump}
\alias{addPump}
\title{Add selected pump(s) to plot.}
\usage{
addPump(pump.select = NULL, vestry = FALSE, col = NULL, pch = 24,
  label = TRUE, pos = 1, cex = 1, latlong = FALSE)
}
\arguments{
\item{pump.select}{Numeric or Integer. Vector of water pump numerical ID(s). With \code{vestry = TRUE}, whole number(s) between 1 and 14. With \code{vestry = FALSE}, whole number(s) between 1 and 13. See \code{pumps.vestry} and \code{pumps} for IDs and details about specific pumps. \code{NULL} plots all pumps. Negative selection allowed.}

\item{vestry}{Logical. \code{TRUE} for the 14 pumps from Vestry Report. \code{FALSE} for the original 13 pumps.}

\item{col}{Character. Color of pump points.}

\item{pch}{Numeric. Shape of point character.}

\item{label}{Logical. TRUE adds text label.}

\item{pos}{Numeric. Position of label.}

\item{cex}{Numeric. point cex.}

\item{latlong}{Logical. Use c("lon". "lat") or c("x", "y").}
}
\description{
Add selected pump(s) to plot.
}
