% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowMap.R
\name{addRoads}
\alias{addRoads}
\title{Add all streets and roads to plot.}
\usage{
addRoads(latlong = FALSE, col = "gray")
}
\arguments{
\item{latlong}{Logical. Use estimated longitude and latitude.}

\item{col}{Character. Color}
}
\description{
Add all streets and roads to plot.
}
