% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowMap.R
\name{addRoads}
\alias{addRoads}
\title{Add roads to plot.}
\usage{
addRoads(col = "gray", ...)
}
\arguments{
\item{col}{Character. Color}

\item{...}{Additional plotting parameters.}
}
\description{
Add roads to plot.
}
