% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSnow.R
\name{addSnow}
\alias{addSnow}
\title{Adds Snow's Broad Street pump walking neighborhood.}
\usage{
addSnow(type = "street", color = "dodgerblue", alpha.level = 0.25,
  line.width = 2, ...)
}
\arguments{
\item{type}{Character. "area", "boundary" or "street".}

\item{color}{Character. Neighborhood color.}

\item{alpha.level}{Numeric. A value in [0, 1] to set alpha level transparency for area plot.}

\item{line.width}{Numeric. lwd for "street" and "boundary"}

\item{...}{Additional plotting parameters.}
}
\description{
Reproduces Snow's graphical annotation of the Vestry Report map.
}
\examples{
# plot(neighborhoodVoronoi())
# addSnow()
}
\seealso{
\code{\link{snowMap}},
\code{\link{addIndexCase}},
\code{\link{addKernelDensity}},
\code{\link{addLandmarks}},
\code{\link{addPlaguePit}},
\code{\link{addVoronoi}},
\code{\link{addWhitehead}}
}
