% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSnow.R
\name{addSnow}
\alias{addSnow}
\title{Adds Snow's graphical annotation of the Broad Street pump walking neighborhood.}
\usage{
addSnow(type = "area", color = "dodgerblue", alpha.level = 0.25,
  line.width = 2)
}
\arguments{
\item{type}{Character. Type of annotation plot: "area", "perimeter" or "street".}

\item{color}{Character. Neighborhood color.}

\item{alpha.level}{Numeric. Alpha level transparency: a value in [0, 1].}

\item{line.width}{Numeric. Line width for \code{type = "street"} and \code{type = "perimeter"}.}
}
\description{
Adds Snow's graphical annotation of the Broad Street pump walking neighborhood.
}
\examples{
\donttest{

plot(neighborhoodVoronoi())
addSnow()
}
}
