% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSnow.R
\name{addSnow}
\alias{addSnow}
\title{Adds Snow's graphical annotation of the Broad Street pump walking neighborhood.}
\usage{
addSnow(type = "street", color = "dodgerblue", alpha.level = 0.25,
  line.width = 2, ...)
}
\arguments{
\item{type}{Character. Type of annotation plot: "area", "boundary" or "street".}

\item{color}{Character. Neighborhood color.}

\item{alpha.level}{Numeric. Alpha level transparency: a value in [0, 1].}

\item{line.width}{Numeric. Line width for "street" and "boundary".}

\item{...}{Additional plotting parameters.}
}
\description{
Uses alphahull::ashape().
}
\examples{
# plot(neighborhoodVoronoi())
# addSnow()
}
\seealso{
\code{\link{snowMap}},
\code{\link{addIndexCase}},
\code{\link{addKernelDensity}},
\code{\link{addLandmarks}},
\code{\link{addPlaguePit}},
\code{\link{addVoronoi}},
\code{\link{addWhitehead}}
}
