% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLandmarks.R
\name{addVoronoi}
\alias{addVoronoi}
\title{Add Voronoi cells.}
\usage{
addVoronoi(select = NULL, vestry = FALSE, col = "black", lty = "solid")
}
\arguments{
\item{select}{Numeric. Vector of selected or excluded pumps are used. For "pumps", there are 13 possibilities; for "vestry", there are 14. Default is NULL: all pumps used.}

\item{vestry}{Logical. FALSE for original 13 pumps. TRUE for 14 pumps in Vestry Report.}

\item{col}{Character. Color of circle. Default is "black".}

\item{lty}{Character. Type of line for circle. Default is "solid".}
}
\description{
Compute and draws Voronoi cells using deldir::deldir().
}
\examples{
snowMap(add.landmarks = FALSE)
addVoronoi()
}

