% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVoronoi.R
\name{addVoronoi}
\alias{addVoronoi}
\title{Add Voronoi cells.}
\usage{
addVoronoi(select = NULL, vestry = FALSE, col = "black", lty = "solid",
  ...)
}
\arguments{
\item{select}{Numeric. Default is NULL and all pumps are used. Ortherwise, selection by a vector of numeric IDs: 1 to 13 for \code{pumps}; 1 to 14 for \code{pumps.vestry}.}

\item{vestry}{Logical. FALSE for original 13 pumps. TRUE for 14 pumps in Vestry Report.}

\item{col}{Character. Color of borders.}

\item{lty}{Character. Type of line for borders.}

\item{...}{Additional plotting parameters.}
}
\description{
Computes and draws Voronoi cells using deldir::deldir().
}
\examples{
snowMap(add.landmarks = FALSE)
addVoronoi()
}
