% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caseDistance.R
\name{caseDistance}
\alias{caseDistance}
\title{Compute distance between case fatalities.}
\usage{
caseDistance(a = 19, b = 263, meters = FALSE)
}
\arguments{
\item{a}{Numeric. Case ID.}

\item{b}{Numeric. Case ID.}

\item{meters}{Logical. Compute metric (meters) or nominal distance.}
}
\description{
Compute distance between case fatalities.
}
