% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unstack.R
\name{classifierAudit}
\alias{classifierAudit}
\title{Test classification of case.}
\usage{
classifierAudit(case = 483, segment = "326-2")
}
\arguments{
\item{case}{Numeric or Integer. Numeric ID of (anchor) case.}

\item{segment}{Character. Segment ID from cholera::road.segments}
}
\value{
Logical TRUE or FALSE
}
\description{
Test assignment of case to segment using orthogonal bisection of segment as criterion.
}
\note{
segment "326-2 is part of Little Windmill Street.
}
