% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unstack.R
\name{classifierAudit}
\alias{classifierAudit}
\title{Test whether case is orthogonal to segment.}
\usage{
classifierAudit(case = 483, segment = "326-2")
}
\arguments{
\item{case}{Numeric or Integer. Numeric ID of case.}

\item{segment}{Character. Segment ID. See cholera::road.segments}
}
\value{
Logical TRUE or FALSE
}
\description{
Test whether case is orthogonal to segment.
}
\note{
segment "326-2" is part of Little Windmill Street. This is a diagnostics; it is not a guarantee of correct classification.
}
