% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifierAudit.R
\name{classifierAudit}
\alias{classifierAudit}
\title{Test if case is orthogonal to segment.}
\usage{
classifierAudit(case = 483, segment = "326-2", observed = TRUE,
  coordinates = FALSE)
}
\arguments{
\item{case}{Numeric or Integer. Numeric ID of observed case.}

\item{segment}{Character. Segment ID. See \code{road.segments}.}

\item{observed}{Logical. \code{FALSE} observed case; \code{TRUE} simulated case (\code{regular.cases}).}

\item{coordinates}{Logical. Orthogonal projection coordinates.}
}
\value{
Logical \code{TRUE} or \code{FALSE}
}
\description{
Diagnostic to check classification of case to a road segment.
}
\note{
This function is a diagnostic. It is not a guarantee of correct classification.
}
\examples{
classifierAudit(case = 483, segment = "326-2")
plot(classifierAudit(case = 483, segment = "326-2"))
}
