% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unstack.R
\name{classifierAuditPlot}
\alias{classifierAuditPlot}
\title{Plot test classification of case.}
\usage{
classifierAuditPlot(case = 483, segment = "326-2", radius = 0.1)
}
\arguments{
\item{case}{Numeric or Integer. Numeric ID of (anchor) case.}

\item{segment}{Character. Segment ID from cholera::road.segments}

\item{radius}{Numeric. Controls the degree of zoom.}
}
\value{
A base R graphic.
}
\description{
Test assignment of case to segment using orthogonal bisection of segment as criterion.
}
\note{
segment "326-2 is part of Little Windmill Street.
}
