% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceTime.R
\name{distanceTime}
\alias{distanceTime}
\title{Convert distance to elapsed time.}
\usage{
distanceTime(x, unit = "second", meter.unit = cholera::unitMeter(1, unit =
  "meter"), speed = 5)
}
\arguments{
\item{x}{Numeric. Nominal map distance.}

\item{unit}{Character. Unit of measurement: "hour", "minute" or "second".}

\item{meter.unit}{Numeric. 1 meter is approximately 54 map units. Set by cholera::unitMeter().}

\item{speed}{Numeric. Walking speed in km/hr.}
}
\value{
An R vector.
}
\description{
Convert distance to elapsed time.
}
