% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclideanDistance.R
\name{euclideanDistance}
\alias{euclideanDistance}
\title{Compute Euclidean distance to nearest pump or selected pump.}
\usage{
euclideanDistance(case, pump.select = NULL, vestry = FALSE, unit = NULL)
}
\arguments{
\item{case}{Numeric or Integer. Case must be a whole number between 1 and 578. To compute distance, the function uses the coordinates of the case's "address" (i.e., its "anchor case").}

\item{pump.select}{Numeric or Integer vector. For "vestry = FALSE", 1 >= |pump.select| <= 13. For "vestry = TRUE", 1 >= |pump.select| <= 14. Negative values are excluded from consideration. Default is NULL, which returns the closest pump.}

\item{vestry}{Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 in the original map.}

\item{unit}{Character. Unit of measurement: "meter" or "yard". Default is NULL, which returns the map's native scale. See \code{vignette("roads")} for information on unit distances.}
}
\value{
A base R data frame.
}
\description{
Compute Euclidean distance to nearest pump or selected pump.
}
\examples{
euclideanDistance(1)
euclideanDistance(1, 2)
euclideanDistance(1, -7)
}
\seealso{
\code{\link{fatalities}}, \code{\link{fatalities.address}}, \code{\link{fatalities.unstacked}}, \code{vignette("pump.neighborhoods")}
}
