% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclideanDistance.R
\name{euclideanDistance}
\alias{euclideanDistance}
\title{Compute Euclidean distance between a case and a pump.}
\usage{
euclideanDistance(case, pump.select, vestry = FALSE, unit = NULL)
}
\arguments{
\item{case}{Numeric or Integer. Case must be a whole number between 1 and 578. To compute distance, the function uses the coordinates of the case's "address" (i.e., the "anchor case").}

\item{pump.select}{Numeric or Integer. 1 to 13 for "vestry = FALSE"; 1 to 14 for "vestry = TRUE".}

\item{vestry}{Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 in the original map.}

\item{unit}{Character. Unit of measurement: "meter" or "yard". Default is NULL, which returns the map's native scale.}
}
\value{
A base R data frame.
}
\description{
Compute Euclidean distance between a case and a pump.
}
\examples{
euclideanDistance(1, 7)
}
\seealso{
\code{\link{fatalities}}, \code{\link{fatalities.address}}, \code{\link{fatalities.unstacked}}, \code{vignette("pump.neighborhoods")}
}
