% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclideanDistance.R
\name{euclideanDistance}
\alias{euclideanDistance}
\title{Compute the Euclidean distance between cases and/or pumps. (Beta)}
\usage{
euclideanDistance(origin = 1, destination = NULL, type = "case-pump",
  observed = TRUE, case.location = "address", vestry = FALSE,
  unit = "meter", time.unit = "second", walking.speed = 5)
}
\arguments{
\item{origin}{Numeric or Character. Numeric ID of case or pump. Character landmark name.}

\item{destination}{Numeric or Character. Numeric ID(s) of case(s) or pump(s). Exclusion is possible via negative selection (e.g., -7). Default is \code{NULL}, which returns closest pump or "anchor" case. Character landmark name.}

\item{type}{Character "case-pump", "cases" or "pumps".}

\item{observed}{Logical. Use observed or "simulated" expected data.}

\item{case.location}{Character. For \code{observed = FALSE}: "address" or "nominal". "nominal" is the x-y coordinate of \code{regular.cases}.}

\item{vestry}{Logical. \code{TRUE} uses the 14 pumps from the Vestry Report. \code{FALSE} uses the 13 pumps from the original map.}

\item{unit}{Character. Unit of distance: "meter", "yard" or "native". "native" returns the map's native scale. See \code{vignette("roads")} for information on unit distances.}

\item{time.unit}{Character. "hour", "minute", or "second".}

\item{walking.speed}{Numeric. Walking speed in km/hr.}
}
\value{
An R data frame.
}
\description{
Compute the Euclidean distance between cases and/or pumps. (Beta)
}
\note{
The function uses a case's "address" (i.e., "anchor" case of a stack) to compute distance. Time is computed using \code{distanceTime()}.
}
\examples{
# path from case 1 to nearest pump.
euclideanDistance(1)

# path from case 1 to pump 6.
euclideanDistance(1, 6)

# exclude pump 7 from consideration.
euclideanDistance(1, -7)

# path from case 1 to case 6.
euclideanDistance(1, 6, type = "cases")

# path from pump 1 to pump 6.
euclideanDistance(1, 6, type = "pumps")

# distance for cases 1 through 3 nearest pump.
lapply(1:3, euclideanDistance)
}
