% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclideanPath.R
\name{euclideanPath}
\alias{euclideanPath}
\title{Compute Euclidean path coordinates from observed case/landmark to nearest/selected pump.}
\usage{
euclideanPath(origin = 1, destination = NULL, type = "case-pump",
  vestry = FALSE, latlong = FALSE, case.set = "observed",
  location = "nominal", distance.unit = "meter", time.unit = "second",
  walking.speed = 5, include.landmarks = TRUE, anchor.anchor = TRUE)
}
\arguments{
\item{origin}{Numeric. Vector of origin(s) (numeric ID or character name landmark/pump ).}

\item{destination}{Numeric. Vector of destination(s) (numeric or landmark/pump name).}

\item{type}{Character. Path case to pump. FALSE is all other combinations of cases, landmarks and pumps.}

\item{vestry}{Logical. \code{TRUE} uses the 14 pumps from the map in the Vestry Report. \code{FALSE} uses the 13 pumps from the original map.}

\item{latlong}{Logical.}

\item{case.set}{Character. "observed" or "expected".}

\item{location}{Character. For cases and pumps. "nominal", "anchor" or "orthogonal".}

\item{distance.unit}{Character. Unit of distance: "meter" or "yard".}

\item{time.unit}{Character. "hour", "minute", or "second".}

\item{walking.speed}{Numeric. Walking speed in km/hr.}

\item{include.landmarks}{Logical. Include landmarks as cases.}

\item{anchor.anchor}{Logical. Use anchor cases.}
}
\description{
Compute Euclidean path coordinates from observed case/landmark to nearest/selected pump.
}
