% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fatalities.unstacked}
\alias{fatalities.unstacked}
\title{"Unstacked" amended cholera fatalities data with fatality as unit of observation.}
\format{
A data frame with 3 variable that records the position of the 578 bars on Snow's map.

 \describe{
    \item{\code{case}}{numerical case ID}
    \item{\code{x}}{x-coordinate}
    \item{\code{y}}{y-coordinate}
}
}
\usage{
fatalities.unstacked
}
\description{
An "unstacked" version of the \code{fatalities} dataset. It changes the unit of observation from the case (bar) to the "address", the x-y coordinates of the case at the base of a stack, and assigns the base case's coordinates to all cases in the stack.
}
\note{
\code{\link{unstackFatalities}} documents the code for these data. For details, see \code{vignette}("unstacking.fatalities").
}
\seealso{
\code{\link{caseLocator}}

\code{\link{streetNameLocator}}

\code{\link{streetNumberLocator}}
}
\keyword{datasets}
