% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatalityLocator.R
\name{fatalityLocator}
\alias{fatalityLocator}
\title{Locate fatality by numerical case ID.}
\usage{
fatalityLocator(case, zoom = FALSE, radius = 2, stacked = TRUE)
}
\arguments{
\item{case}{Numeric or Integer. Whole number between 1 and 578.}

\item{zoom}{Logical.}

\item{radius}{Numeric. Controls the degree of zoom.}

\item{stacked}{Logical. TRUE uses \code{fatalities} ("stacked" data); FALSE uses \code{fatalities.address} ("unstacked" data).}
}
\value{
A base R graphics plot.
}
\description{
Plots John Snow's map of the 1854 London cholera outbreak and highlights the
selected case.
}
\examples{
fatalityLocator(290)
fatalityLocator(290, zoom = TRUE)
fatalityLocator(290, stacked = FALSE)
fatalityLocator(290, zoom = TRUE, stacked = FALSE)
}
\seealso{
\code{fatalities}

\code{fatalities.address}
}

