% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unstack.R
\name{fixFatalities}
\alias{fixFatalities}
\title{Fix errors in Dodson and Tobler's digitization of Snow's map.}
\usage{
fixFatalities()
}
\value{
An R data frame.
}
\description{
Fixes two apparent coding errors using three misplaced cases.
}
\seealso{
\code{vignette("duplicate.missing.cases")}
}
