% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{frame.sample}
\alias{frame.sample}
\title{Partitioned map frame points (segment endpoints).}
\format{
A list of 3 vectors length 19, 19 and 18 from cholera::roads$id.
 \describe{
    \item{\code{frame.sample}}{cholera::roads$id}
 }
}
\usage{
frame.sample
}
\description{
Partitioned map frame points (segment endpoints).
}
\keyword{datasets}
