% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolines.R
\name{isolines}
\alias{isolines}
\title{Plot isochrone and isodistance regions (prototype)}
\usage{
isolines(pump.select = 7, post = 50, post.type = "distance",
  palette = "OrRd")
}
\arguments{
\item{pump.select}{Numeric.}

\item{post}{Numeric.}

\item{post.type}{Character. "distance or "time".}

\item{palette}{Character. RColorBrewer palette.}
}
\description{
Plot isochrone and isodistance regions (prototype)
}
