% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoLines.R
\name{isoLines}
\alias{isoLines}
\title{Plot isochrone and isodistance regions (prototype)}
\usage{
isoLines(post = 50, post.type = "distance", palette = "plasma",
  alpha.level = 1/2)
}
\arguments{
\item{post}{Numeric. Distance or time increment.}

\item{post.type}{Character. "distance" or "time".}

\item{palette}{Character.}

\item{alpha.level}{Numeric. Alpha level transparency}
}
\description{
Plot isochrone and isodistance regions (prototype)
}
