% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoVertices.R
\name{isoVertices}
\alias{isoVertices}
\title{Isochrone and isodistance vertices (prototype)}
\usage{
isoVertices(post = 50, post.type = "distance", multi.core = TRUE,
  dev.mode = FALSE)
}
\arguments{
\item{post}{Numeric.}

\item{post.type}{Character. "distance or "time".}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores.}

\item{dev.mode}{Logical. Development mode uses parallel::parLapply().}
}
\description{
Isochrone and isodistance vertices (prototype)
}
