% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landmark.squares}
\alias{landmark.squares}
\title{Centers of city squares.}
\format{
A data frame with 6 variables that records the position of the orthogonal projection of landmarks onto the network of roads.
 \describe{
    \item{\code{name}}{square name}
    \item{\code{x}}{x-coordinate}
    \item{\code{y}}{y-coordinate}
    \item{\code{case}}{numeric case ID}
 }
}
\usage{
landmark.squares
}
\description{
Centers of city squares.
}
\keyword{datasets}
