% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarkData.R
\name{landmarkData}
\alias{landmarkData}
\title{Landmark data.}
\usage{
landmarkData(multi.core = FALSE)
}
\arguments{
\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. On Windows, only \code{multi.core = FALSE} is available.}
}
\description{
Nominal and orthogonal coordinates
}
