% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landmarks}
\alias{landmarks}
\title{Orthogonal projection of landmarks onto road network.}
\format{A data frame with 6 variables that records the position of the orthogonal projection of landmarks onto the network of roads.
 \describe{
    \item{\code{road.segment}}{"address" road segment}
    \item{\code{x.proj}}{orthogonal x-coordinate}
    \item{\code{y.proj}}{orthogonal y-coordinate}
    \item{\code{ortho.dist}}{orthogonal distance to home road segment}
     \item{\code{x}}{nominal x-coordinate}
    \item{\code{y}}{nominal y-coordinate}
    \item{\code{name}}{landmark name}
    \item{\code{case}}{numeric case ID}
 }}
\usage{
landmarks
}
\description{
Orthogonal projection of landmarks onto road network.
}
\note{
\code{\link{landmarkData}} documents the code for these data.
}
\keyword{datasets}
