% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landmarks}
\alias{landmarks}
\title{Landmark coordinates.}
\format{
A data frame of landmark coordinates with 20 observations and 11 variables.
 \describe{
    \item{\code{case}}{numeric case ID}
    \item{\code{road.segment}}{"address" road segment}
    \item{\code{x}}{nominal x-coordinate}
    \item{\code{y}}{nominal y-coordinate}
    \item{\code{x.lab}}{label x-coordinate}
    \item{\code{y.lab}}{label y-coordinate}
    \item{\code{lon}}{longitude}
    \item{\code{lat}}{latitude}
    \item{\code{lon.lab}}{label longitude}
    \item{\code{lat.lab}}{label latitude}
    \item{\code{name}}{landmark name}
 }
}
\usage{
landmarks
}
\description{
Landmark coordinates.
}
\keyword{datasets}
