% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latlong.nearest.pump}
\alias{latlong.nearest.pump}
\title{Walking distance to Nearest Pump (latlong prototype).}
\format{
An object of class \code{list} of length 3.
}
\usage{
latlong.nearest.pump
}
\description{
Walking distance to Nearest Pump (latlong prototype).
}
\keyword{datasets}
