% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latlong.regular.cases}
\alias{latlong.regular.cases}
\title{"Expected" cases (latlong).}
\format{
A data frame with 2 variable that records the position of 20,000 "expected" cases fitted by sp::spsample().
 \describe{
    \item{\code{x}}{x-coordinate}
    \item{\code{y}}{y-coordinate}
 }
}
\usage{
latlong.regular.cases
}
\description{
The result of using sp::spsample() and sp::Polygon() to generate 20,000 regularly spaced simulated Cartesian/geodesic cases within the map's borders.
}
\keyword{datasets}
