% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latlong.regular.cases}
\alias{latlong.regular.cases}
\title{"Expected" cases (latlong).}
\format{
A data frame with 4 variables that records the position of 19,993 "expected" cases fitted by sp::spsample().
 \describe{
    \item{\code{x}}{x-coordinate}
    \item{\code{y}}{y-coordinate}
    \item{\code{lon}}{longitude}
    \item{\code{lat}}{latitude}
 }
}
\usage{
latlong.regular.cases
}
\description{
The result of using sp::spsample() and sp::Polygon() to generate 19,993 regularly spaced simulated Cartesian/geodesic cases within the map's borders.
}
\keyword{datasets}
