% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongEmbed.R
\name{latlongEmbed}
\alias{latlongEmbed}
\title{Embed anchors and pumps into road segments (prototype).}
\usage{
latlongEmbed(vestry = FALSE, multi.core = TRUE)
}
\arguments{
\item{vestry}{Logical.}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. See \code{vignette("Parallelization")} for details.}
}
\description{
Embed anchors and pumps into road segments (prototype).
}
