% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongLandmarks.R
\name{latlongLandmarks}
\alias{latlongLandmarks}
\title{Compute Georeferenced Latitude and Longitude (prototype).}
\usage{
latlongLandmarks(path, orthogonal = FALSE)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/"}

\item{orthogonal}{Logical. Use orthogonal projection coordinates.}
}
\description{
Compute Georeferenced Latitude and Longitude (prototype).
}
\note{
This documents the computation of the latlong version of the landmarks data frame.
}
