% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongNeighborhoodData.R
\name{latlongNeighborhoodData}
\alias{latlongNeighborhoodData}
\title{Compute network graph of roads, cases and pumps.}
\usage{
latlongNeighborhoodData(path, vestry = FALSE)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/"}

\item{vestry}{Logical.}
}
\description{
Assembles cases, pumps and road into a network graph.
}
