% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongPumps.R
\name{latlongPumps}
\alias{latlongPumps}
\title{Compute Georeferenced Latitude and Longitude (prototype).}
\usage{
latlongPumps(path, vestry = FALSE)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/"}

\item{vestry}{Logical.}
}
\description{
Compute Georeferenced Latitude and Longitude (prototype).
}
\note{
This documents the computation of the latlong version of the pumps and pumps.vestry data frames.
}
