% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongSnowMap.R
\name{latlongSnowMap}
\alias{latlongSnowMap}
\title{Plot John Snow's cholera map.}
\usage{
latlongSnowMap(path, stacked = TRUE, asp = 1.65)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/".}

\item{stacked}{Logical. Use stacked fatalities.}

\item{asp}{Numeric. Aspect ratio.}
}
\description{
Plot John Snow's cholera map.
}
