% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongVoronoiB.R
\name{latlongVoronoiB}
\alias{latlongVoronoiB}
\title{Compute Georeferenced Latitude and Longitude of vertices of Voronoi polygons.}
\usage{
latlongVoronoiB()
}
\description{
Compute Georeferenced Latitude and Longitude of vertices of Voronoi polygons.
}
\examples{
snowMap(latlong = TRUE)
cells <- latlongVoronoiB()
invisible(lapply(cells, function(x) polygon(x[, c("lon", "lat")])))
}
