% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongNearestPump.R
\name{latlongWalkingPath}
\alias{latlongWalkingPath}
\title{Plot walking path to nearest pump (prototype).}
\usage{
latlongWalkingPath(case = 1, path, zoom = TRUE)
}
\arguments{
\item{case}{Numeric.}

\item{path}{Character. e.g., "~/Documents/Data/".}

\item{zoom}{Logical or Numeric. A numeric value >= 0 that controls the degree of zoom.}
}
\description{
Plot walking path to nearest pump (prototype).
}
