% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongWalkingPath.R
\name{latlongWalkingPath}
\alias{latlongWalkingPath}
\title{Plot walking path to nearest pump (prototype).}
\usage{
latlongWalkingPath(case = 1, vestry = FALSE)
}
\arguments{
\item{case}{Numeric.}

\item{vestry}{Logical. \code{TRUE} uses the 14 pumps from the map in the Vestry Report. \code{FALSE} uses the 13 pumps from the original map.}
}
\description{
Plot walking path to nearest pump (prototype).
}
