% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{nearestPump}
\alias{nearestPump}
\title{Compute walking distance from anchor cases to nearest pump (or from among selected pumps).}
\usage{
nearestPump(pump.select = NULL, vestry = FALSE, weighted = TRUE,
  observed = TRUE, cores = 1L)
}
\arguments{
\item{pump.select}{Numeric. Default is NULL: all pumps are used. Otherwise, selection by a vector of numeric IDs: 1 to 13 for \code{pumps}; 1 to 14 for \code{pumps.vestry}. Negative selection allowed.}

\item{vestry}{Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 in the original map.}

\item{weighted}{Logical. TRUE computes shortest path in terms of road length. FALSE computes shortest path in terms of the number of nodes.}

\item{observed}{Logical. Observed or expected walking path pump neighborhoods.}

\item{cores}{Numeric. The number logical cores (truncates with as.integer()). Default is 1, which is the only possible value for Windows.}
}
\value{
An R data frame.
}
\description{
Compute walking distance from anchor cases to nearest pump (or from among selected pumps).
}
