% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neighborhood.segments}
\alias{neighborhood.segments}
\title{List of data frames describing the road segments of the 10 observed pump neighborhood.}
\format{Each data frame has 8 variables that describe a road's straight line segments.
\describe{
  \item{\code{street}}{numeric street ID, which range between 1 and 528}
  \item{\code{id}}{characger segement ID}
  \item{\code{name}}{road name}
  \item{\code{x1}}{x-coordinate of first endpoint}
  \item{\code{y1}}{y-coordinate of first endpoint}
  \item{\code{x2}}{x-coordinate of second endpoint}
  \item{\code{y2}}{y-coordinate of second endpoint}
  \item{\code{trimmed}}{logical indicates whether entire road segment is included or is trimmed}
}}
\usage{
neighborhood.segments
}
\description{
List of data frames describing the road segments of the 10 observed pump neighborhood.
}
\keyword{datasets}

