% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{neighborhoodPaths}
\alias{neighborhoodPaths}
\title{Compute walking path pump neighborhoods.}
\usage{
neighborhoodPaths(pump.select = NULL, vestry = FALSE, weighted = TRUE)
}
\arguments{
\item{pump.select}{Numeric. Default is NULL: all pumps are used. Otherwise, selection by a vector of numeric IDs: 1 to 13 for \code{pumps}; 1 to 14 for \code{pumps.vestry}.}

\item{vestry}{Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 in the original map.}

\item{weighted}{Logical. TRUE computes shortest path in terms of road length. FALSE computes shortest path in terms of the number of nodes.}
}
\description{
Group walking paths by neighborhood.
}
