% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoodSnow.R
\name{neighborhoodSnow}
\alias{neighborhoodSnow}
\title{Plot Snow's Broad Street pump walking neighborhood.}
\usage{
neighborhoodSnow(zoom = TRUE, streets = TRUE, add.address = FALSE,
  add.landmarks = TRUE)
}
\arguments{
\item{zoom}{Logical.}

\item{streets}{Logical. TRUE plots streets. FALSE plots orthogonal area.}

\item{add.address}{Logical. Plots the address of cases both in and out of the neighborhood. These are the base case in a stack of fatalities.}

\item{add.landmarks}{Logical. Include landmarks.}
}
\description{
Reproduces Snow's graphic annotation in the Vestry Report.
}
\examples{
neighborhoodSnow()
neighborhoodSnow(streets = FALSE)
neighborhoodSnow(zoom = FALSE, add.address = TRUE); addKernelDensity()
}
