% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoodSnow.R
\name{neighborhoodSnow}
\alias{neighborhoodSnow}
\title{Plot Snow's Broad Street pump walking neighborhood.}
\usage{
neighborhoodSnow(zoom = TRUE, add.address = TRUE, add.landmarks = TRUE)
}
\arguments{
\item{zoom}{Logical.}

\item{add.address}{Logical. Plots the address of cases in the neighborhood, the base case in the stack of fatalities.}

\item{add.landmarks}{Logical. Include landmarks.}
}
\description{
Reproduces Snow's graphic annotation in the Vestry Report.
}
\examples{
neighborhoodSnow(); addKernelDensity()
}

