% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow.R
\name{neighborhoodSnow}
\alias{neighborhoodSnow}
\title{Compute Snow's graphical annotation of Broad Street pump neighborhood.}
\usage{
neighborhoodSnow(latlong = FALSE)
}
\arguments{
\item{latlong}{Logical. Use estimated longitude and latitude.}
}
\description{
Computational approximation.
}
