% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoodVoronoi.R
\name{neighborhoodVoronoi}
\alias{neighborhoodVoronoi}
\title{Plot Voronoi neighborhoods.}
\usage{
neighborhoodVoronoi(selection = NULL, vestry = FALSE, statistic = NULL)
}
\arguments{
\item{selection}{Numeric. Default is NULL: all pumps are used. Ortherwise, selection by a vector of numeric IDs: 1 to 13 for \code{pumps}; 1 to 14 for \code{pumps.vestry}.}

\item{vestry}{Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 in the original map.}

\item{statistic}{Character. NULL plots address points. "address" plots the total count of addresses in a neighborhood at its pump's location. "fatality" plots the total count of fatalities in a neighborhood at its pump's location.}
}
\value{
A base R graphics plot.
}
\description{
Voronoi diagram of John Snow's 1854 London cholera data.
}
\examples{
neighborhoodVoronoi()
neighborhoodVoronoi(statistic = "address")
neighborhoodVoronoi(vestry = TRUE)
neighborhoodVoronoi(selection = 6:7)
neighborhoodVoronoi(selection = -6)
neighborhoodVoronoi(selection = -6, statistic = "fatality")
}
\seealso{
\code{addVoronoi()}
}

