% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walking.R
\name{neighborhoodWalking}
\alias{neighborhoodWalking}
\title{Walking path neighborhoods.}
\usage{
neighborhoodWalking(selection = NULL, vestry = FALSE,
  statistic = "address", weighted = TRUE, multi.core = FALSE)
}
\arguments{
\item{selection}{Numeric. Default is NULL: all pumps are used. Ortherwise, selection by a vector of numeric IDs: 1 to 13 for \code{pumps}; 1 to 14 for \code{pumps.vestry}.}

\item{vestry}{Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 in the original map.}

\item{statistic}{Character. "address" computes the number of addresses in each selected pump neighborhood. "fatality" computes the number of fatalities in pump neighborhoods.}

\item{weighted}{Logical. TRUE uses distance weighted by edge length (i.e., road length). FALSE uses unweighted distance.}

\item{multi.core}{Logical or Numeric. TRUE uses parallel::detectCores(). FALSE uses one, single core. With Numeric, you specify the number logical cores (rounds with as.integer()). On Windows, only "multi.core = FALSE" is available.}
}
\value{
A list of data and parameters of computed walking path neighborhoods.
}
\description{
Data for walking neighbohoods for John Snow's 1854 London cholera data.
}
\section{Notes}{
 This function is computationally intensive (the default configuration takes about 4 minutes to run on a single core). However, two configurations will return pre-computed results: the default set of arguments, and the default set with pump 6 (Little Marlborough Street) excluded: neighborhoodWalking(selection = -6).
}

\examples{
neighborhoodWalking()
neighborhoodWalking(selection = -6)
}
