% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeData.R
\name{nodeData}
\alias{nodeData}
\title{Embed cases ("addresses") and pumps into road network.}
\usage{
nodeData(id = "242-1", output = "nodes", vestry = FALSE)
}
\arguments{
\item{id}{Character. Road segment ID.}

\item{output}{Character. Type of output: nodes or edge list.}

\item{vestry}{Logical. Use Vestry Report pump data}
}
\value{
An R data frame.
}
\description{
Embed cases ("addresses") and pumps into road network.
}
