% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeData.R
\name{nodeData}
\alias{nodeData}
\title{Compute network graph of roads, cases and pumps.}
\usage{
nodeData(embed = TRUE, vestry = FALSE)
}
\arguments{
\item{embed}{Logical. TRUE embeds sites into road network. FALSE returns just the road network.}

\item{vestry}{Logical. Use Vestry Report pump data.}
}
\value{
An R list of nodes, edges and an "igraph" package network graph.
}
\description{
Embed "anchor" cases and pumps into road network graph.
}
