% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeData.R
\name{nodeData}
\alias{nodeData}
\title{Embed cases ("addresses") and pumps into road segment.}
\usage{
nodeData(id = "242-1", type = "nodes", vestry = FALSE)
}
\arguments{
\item{id}{Character. Road segment ID.}

\item{type}{Character. Type of output: nodes or edge list.}

\item{vestry}{Logical. Use Vestry Report pump data.}
}
\value{
An R data frame.
}
\description{
Embed cases ("addresses") and pumps into road segment.
}
\examples{
nodeData("216-1")
nodeData("216-1", "edges")
nodeData()
}
