% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeData.R
\name{nodeData}
\alias{nodeData}
\title{Compute network graph of roads, cases and pumps.}
\usage{
nodeData(embed = TRUE, vestry = FALSE)
}
\arguments{
\item{embed}{Logical. TRUE embeds cases and pumps into road network. FALSE returns just the road network.}

\item{vestry}{Logical. Use Vestry Report pump data.}
}
\value{
An R list of nodes, edges and an 'igraph' network graph.
}
\description{
Assembles "anchor" cases, pumps and road into a network graph.
}
