% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observedNeighborhood.R
\name{observedNeighborhood}
\alias{observedNeighborhood}
\title{Plots walking neighborhoods.}
\usage{
observedNeighborhood(selection = NULL, vestry = FALSE, weighted = TRUE,
  addresses = TRUE, add.landmarks = TRUE)
}
\arguments{
\item{selection}{Numeric. Default is NULL; all pumps are used. Otherwise, selection by a vector of numeric IDs: 1 to 13 for \code{pumps}; 1 to 14 for \code{pumps.vestry}}

\item{vestry}{Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 in the original map.}

\item{weighted}{Logical. TRUE uses shortest path weighted by road distance.}

\item{addresses}{Logical. Use addresses rather than fatalities}

\item{add.landmarks}{Logical. Include landmarks.}
}
\value{
A base R graphics plot.
}
\description{
Plots walking neighborhoods based on the weighted shortest paths between a fatalities and the nearest pump.
}
\details{
Currently computationally intensive (streets = TRUE appox. 90 seocnds; streets = FALSE appox. 75 seconds on Intel Core i7). For better performance, run in batch mode in parallel with multiple cores (appox. 45 and 30 seconds). See \code{vignette}("walking.neighborhoods") for details.
}
\examples{
# observedNeighborhood()
# observedNeighborhood(selection = 6:7)
}
\seealso{
\code{addLandmarks()}

\code{vignette}("walking.neighborhoods")
}

