% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionAddresses.R
\name{openTriads}
\alias{openTriads}
\title{Rotate, stack and partition open triads.}
\usage{
openTriads(subgraphs, census, census.ct)
}
\arguments{
\item{subgraphs}{Object. 'igraph' list of graphs.}

\item{census}{Object. List of graph vertices.}

\item{census.ct}{Object. Count of graph vertices.}
}
\value{
An R data frame.
}
\description{
Rotate, stack and partition open triads.
}
