% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientationPlot.R
\name{orientationPlot}
\alias{orientationPlot}
\title{Orientation plot.}
\usage{
orientationPlot(pumpID = 6, metric = "euclidean")
}
\arguments{
\item{pumpID}{Numeric. Selected pump ID.}

\item{metric}{Character. "eucldidean" or "walking".}
}
\description{
Orientation plot.
}
