% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientationPlot.R
\name{orientationPlot}
\alias{orientationPlot}
\title{Orientation plot.}
\usage{
orientationPlot(pumpID = 6, cases = "ortho", metric = "euclidean",
  pca.line = FALSE, vestry = FALSE)
}
\arguments{
\item{pumpID}{Numeric. Selected pump ID.}

\item{cases}{Character. "address" or "ortho".}

\item{metric}{Character. "eucldidean" or "walking".}

\item{pca.line}{Logical. Add PCA (TSS) line through data.}

\item{vestry}{Logical. \code{TRUE} uses the 14 pumps from the Vestry Report. \code{FALSE} uses the 13 in the original map.}
}
\description{
Orientation plot.
}
