% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oxford.weather}
\alias{oxford.weather}
\title{Oxford monthly weather data, January 1853 - February 2024.}
\format{
A data frame with 9 variables and 2054 observations.
 \describe{
    \item{\code{year}}{Year}
    \item{\code{month}}{Month}
    \item{\code{tmax}}{Mean daily maximum temperature Celsius}
    \item{\code{tmin}}{Mean daily minimum temperature Celsius}
    \item{\code{af}}{Days of air frost }
    \item{\code{rain}}{Total rainfall (mm)}
    \item{\code{sun}}{Total sunshine duration hours}
    \item{\code{prov}}{Provisional data}
    \item{\code{date}}{Date}
 }
}
\usage{
oxford.weather
}
\description{
Extract from UK Met Office (https://www.metoffice.gov.uk/pub/data/weather/uk/climate/stationdata/oxforddata.txt): Lat 51.761 Lon -1.262, 63 metres amsl. Approximate 90 km (55 miles) northwest of Soho.
}
\keyword{datasets}
