% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oxford.weather}
\alias{oxford.weather}
\title{Oxford monthly weather data, January 1853 - November 1860.}
\format{
A data frame with 7 variables and 95 observations.
 \describe{
    \item{\code{year}}{yyyy}
    \item{\code{mo}}{month (mm)}
    \item{\code{tmax}}{maximum temperature degrees Celsius}
    \item{\code{tmin}}{minimum temperature degrees Celsius}
    \item{\code{airfrost}}{days}
    \item{\code{rain}}{milimeters (mm)}
    \item{\code{sun}}{sunshine hours}
 }
}
\usage{
oxford.weather
}
\description{
Extract from UK Met Office (https://www.metoffice.gov.uk/pub/data/weather/uk/climate/stationdata/oxforddata.txt): Lat 51.761 Lon -1.262, 63 metres amsl. Approximate 90 km (55 miles) northwest of Soho.
}
\note{
December 1860 excluded due to missing tmin observation.
}
\keyword{datasets}
