% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionFatalities.R
\name{partitionFatalities}
\alias{partitionFatalities}
\title{Partition fatalities (prototype).}
\usage{
partitionFatalities()
}
\value{
An R list.
}
\description{
Minimize "adjacent" fatalities in stacks.
}
