% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongRoads.R
\name{partitionRoadEndpoints}
\alias{partitionRoadEndpoints}
\title{Partition road endpoints to avoid over-printing of points (prototype).}
\usage{
partitionRoadEndpoints(path)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/".}
}
\description{
Partition road endpoints to avoid over-printing of points (prototype).
}
