% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionRoadEndpoints.R
\name{pdfPartitionRoadEndpoints}
\alias{pdfPartitionRoadEndpoints}
\title{Create PDFs of road endpoints partition (prototype).}
\usage{
pdfPartitionRoadEndpoints(path, multi.core = TRUE)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/".}

\item{multi.core}{Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. See \code{vignette("Parallelization")} for details.}
}
\description{
For georeferencing in QGIS .
}
