% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unstack.R
\name{plot.classifier_audit}
\alias{plot.classifier_audit}
\title{Plot whether case is orthogonal to segment.}
\usage{
\method{plot}{classifier_audit}(x, radius = 0.1, ...)
}
\arguments{
\item{x}{An object of class "classifier_audit" created by classifierAudit().}

\item{radius}{Numeric. Controls the degree of zoom.}

\item{...}{Additional plotting parameters.}
}
\value{
A base R graphic.
}
\description{
Plot method for classifierAudit().
}
\examples{
print(classifierAudit(case = 483, segment = "326-2"))
}
