% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifierAudit.R
\name{plot.classifier_audit}
\alias{plot.classifier_audit}
\title{Plot result of classifierAudit().}
\usage{
\method{plot}{classifier_audit}(x, zoom = TRUE, radius = 0.5,
  unit = "meter", ...)
}
\arguments{
\item{x}{An object of class "classifier_audit" created by \code{classifierAudit()}.}

\item{zoom}{Logical.}

\item{radius}{Numeric. Controls the degree of zoom.}

\item{unit}{Character. Unit of distance: "meter" (the default), "yard" or "native". "native" returns the map's native scale. "unit" is meaningful only when "weighted" is \code{TRUE}. See \code{vignette("roads")} for information on unit distances.}

\item{...}{Additional parameters.}
}
\value{
A base R graphic.
}
\description{
Plot case, segment and orthogonal projector.
}
\examples{
plot(classifierAudit(case = 483, segment = "326-2"))
}
