% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{plot.euclidean}
\alias{plot.euclidean}
\title{Plot method for neighborhoodEuclidean().}
\usage{
\method{plot}{euclidean}(x, ...)
}
\arguments{
\item{x}{An object of class "euclidean" created by \code{neighborhoodEuclidean()}.}

\item{...}{Additional plotting parameters.}
}
\value{
A base R plot.
}
\description{
Plot method for neighborhoodEuclidean().
}
\examples{
\dontrun{

plot(neighborhoodEuclidean())
plot(neighborhoodEuclidean(-6))
plot(neighborhoodEuclidean(pump.select = 6:7))
}
}
