% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclideanDistance.R
\name{plot.euclidean_distance}
\alias{plot.euclidean_distance}
\title{Plot observed and simulated walking neighborhoods.}
\usage{
\method{plot}{euclidean_distance}(x, zoom = TRUE, radius = 0.5, ...)
}
\arguments{
\item{x}{An object of class "euclidean_distance" created by euclideanDistance().}

\item{zoom}{Logical.}

\item{radius}{Numeric. Controls the degree of zoom.}

\item{...}{Additional plotting parameters.}
}
\value{
A base R graphics plot.
}
\description{
Neighborhoods are based on the shortest paths between a fatality's address and its nearest pump.
}
\examples{
plot(euclideanDistance(1))
}
