% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclideanLatlong.R
\name{plot.euclideanLatlong}
\alias{plot.euclideanLatlong}
\title{Plot method for euclideanLatlong()}
\usage{
\method{plot}{euclideanLatlong}(x, type = "star", add = FALSE,
  add.observed.points = TRUE, alpha.level = 0.75, polygon.type = "solid",
  ...)
}
\arguments{
\item{x}{Object.}

\item{type}{Character. "star", "area.points" or "area.polygons". "area" flavors only valid when \code{case.set = "expected"}.}

\item{add}{Logical. Add graphic to an existing plot.}

\item{add.observed.points}{Logical. Add observed fatality "addresses".}

\item{alpha.level}{Numeric. Alpha level transparency for area plot: a value in [0, 1].}

\item{polygon.type}{Character. "perimeter" or "solid".}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for euclideanLatlong()
}
