% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongNeighborhoodWalking.R
\name{plot.latlong_walking}
\alias{plot.latlong_walking}
\title{Plot method for latlongNeighborhoodWalking().}
\usage{
\method{plot}{latlong_walking}(x, ...)
}
\arguments{
\item{x}{An object of class "latlong_walking" created by \code{latlongNeighborhoodWalking()}.}

\item{...}{Additional plotting parameters.}
}
\value{
A base R plot.
}
\description{
Plot method for latlongNeighborhoodWalking().
}
