% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongWalkingPath.R
\name{plot.latlong_walking_path}
\alias{plot.latlong_walking_path}
\title{Plot the walking path between selected cases and/or pumps.}
\usage{
\method{plot}{latlong_walking_path}(x, zoom = TRUE, mileposts = TRUE,
  milepost.unit = "distance", milepost.interval = NULL, alpha.level = 1,
  ...)
}
\arguments{
\item{x}{An object of class "latlong_walking_path" created by latlongWalkingPath().}

\item{zoom}{Logical or Numeric. A numeric value >= 0 that controls the degree of zoom.}

\item{mileposts}{Logical. Plot mile/time posts.}

\item{milepost.unit}{Character. "distance" or "time".}

\item{milepost.interval}{Numeric. Mile post interval unit of distance (yard or meter) or unit of time (seconds).}

\item{alpha.level}{Numeric. Alpha level transparency for path: a value in [0, 1].}

\item{...}{Additional plotting parameters.}
}
\value{
A base R plot.
}
\description{
Plot the walking path between selected cases and/or pumps.
}
