% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongNeighborhoodVoronoi.R,
%   R/voronoiLatlong.R
\name{plot.latlongVoronoi}
\alias{plot.latlongVoronoi}
\title{Plot method for latlongNeighborhoodVoronoi()}
\usage{
\method{plot}{latlongVoronoi}(x, add.pumps = TRUE,
  delaunay.voronoi = "voronoi", euclidean.paths = FALSE, ...)

\method{plot}{latlongVoronoi}(x, add.pumps = TRUE,
  delaunay.voronoi = "voronoi", euclidean.paths = FALSE, ...)
}
\arguments{
\item{x}{Object.}

\item{add.pumps}{Logical.}

\item{delaunay.voronoi}{Character "delaunay", "voronoi", or "both".}

\item{euclidean.paths}{Logical.}

\item{...}{Additional plotting parameters.}
}
\description{
Plot method for latlongNeighborhoodVoronoi()

Plot method for voronoiLatlong()
}
