% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeries.R
\name{plot.time_series}
\alias{plot.time_series}
\title{Plot time series data, with date of removal of pump handle (Vestry report).}
\usage{
\method{plot}{time_series}(x, multi.plot = FALSE,
  main = "Removal of the Broad Street Pump Handle", points = TRUE,
  pump.handle = TRUE, weekend = FALSE, xlab = "Date", ylab = NULL, ...)
}
\arguments{
\item{x}{Object. class "time_series".}

\item{multi.plot}{Logical. Multiple data sets in single plot.}

\item{main}{Character. Title of graph.}

\item{points}{Logical. Plot points.}

\item{pump.handle}{Logical. Plot date of removal of Broad Street pump handle.}

\item{weekend}{Logical. Highlight weekend.}

\item{xlab}{Character. x-axis label.}

\item{ylab}{Character. y-axis label.}

\item{...}{Additional plotting parameters.}
}
\description{
Plot time series data, with date of removal of pump handle (Vestry report).
}
