% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi.R
\name{plot.voronoi}
\alias{plot.voronoi}
\title{Plot Voronoi neighborhoods.}
\usage{
\method{plot}{voronoi}(x, statistic = NULL, ...)
}
\arguments{
\item{x}{An object of class "voronoi" created by neighborhoodVoronoi().}

\item{statistic}{NULL or Character. NULL plots the address, a stack's base case. "address" plots the number of addresses and its Pearson residual in each selected pump neighborhood. "fatality" plots the number of fatalities and its Pearson resiudual in pump neighborhoods.}

\item{...}{Additional plotting parameters.}
}
\value{
A base R graph.
}
\description{
Plot Voronoi neighborhoods.
}
\examples{
dat <- neighborhoodVoronoi()
plot(x = dat)

plot(x = dat, statistic = "fatality")
}
\seealso{
\code{addVoronoi()}
\code{pumpSummary()}
}

