% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walking.R
\name{plot.walking}
\alias{plot.walking}
\title{Plot method for neighborhoodWalking().}
\usage{
\method{plot}{walking}(x, type = "road", ...)
}
\arguments{
\item{x}{An object of class "walking" created by neighborhoodWalking().}

\item{type}{Character. "road", "area.points" or "area.polygons". "area" flavors only valid when case.set = "expected".}

\item{...}{Additional plotting parameters.}
}
\value{
A base R plot.
}
\description{
Plot method for neighborhoodWalking().
}
\section{Note}{
 "area.polygons" is experimental and still under developtment: it may occassionally throw an error.
}

\examples{
# plot(neighborhoodWalking())
# plot(neighborhoodWalking(case.set = "expected"))
# plot(neighborhoodWalking(case.set = "expected"), "area.points")
}
