% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walking.R
\name{plot.walking}
\alias{plot.walking}
\title{Plot method for neighborhoodWalking().}
\usage{
\method{plot}{walking}(x, area = FALSE, ...)
}
\arguments{
\item{x}{An object of class "walking" created by neighborhoodWalking().}

\item{area}{Logical. TRUE returns area plot. FALSE returns walking paths plot. Works only with case.set = "expected" or case.set = "snow".}

\item{...}{Additional plotting parameters.}
}
\value{
A base R plot.
}
\description{
Plot method for neighborhoodWalking().
}
\examples{
# plot(neighborhoodWalking())
# plot(neighborhoodWalking(case.set = "expected"))
# plot(neighborhoodWalking(case.set = "expected"), area = TRUE)
}
