% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walking.R
\name{plot.walking}
\alias{plot.walking}
\title{Plot observed and simulated walking neighborhoods.}
\usage{
\method{plot}{walking}(x, streets = TRUE, observed = TRUE, ...)
}
\arguments{
\item{x}{An object of class "walking" created by neighborhoodWalking().}

\item{streets}{Logical. TRUE plots neighborhoods by street. FALSE plots orthogonal neighborhoods (area).}

\item{observed}{Logical. TRUE uses observed cases. FALSE uses "regular" simulated cases.}

\item{...}{Additional plotting parameters.}
}
\value{
A base R graphics plot.
}
\description{
Neighborhoods are based on the shortest paths between a fatality's address and its nearest pump.
}
\examples{
plot(neighborhoodWalking())
}
