% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkingDistance.R
\name{plot.walking_distance}
\alias{plot.walking_distance}
\title{Plot the walking distance between cases and/or pumps.}
\usage{
\method{plot}{walking_distance}(x, zoom = TRUE, radius = 0.5, ...)
}
\arguments{
\item{x}{An object of class "walking_distance" created by walkingDistance().}

\item{zoom}{Logical.}

\item{radius}{Numeric. Controls the degree of zoom.}

\item{...}{Additional plotting parameters.}
}
\value{
A base R plot.
}
\description{
Plot the walking distance between cases and/or pumps.
}
\examples{
plot(walkingDistance(1))
}
