% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkingPath.R
\name{plot.walking_path}
\alias{plot.walking_path}
\title{Plot the walking path between selected cases and/or pumps.}
\usage{
\method{plot}{walking_path}(x, zoom = TRUE, radius = 0.5,
  unit.posts = "distance", unit.interval = NULL, ...)
}
\arguments{
\item{x}{An object of class "walking_path" created by walkingPath().}

\item{zoom}{Logical.}

\item{radius}{Numeric. Controls the degree of zoom.}

\item{unit.posts}{Character. "distance" for mileposts; "time" for timeposts.}

\item{unit.interval}{Numeric. Sets interval between posts: for "distance", the default is 50 meters; for "time", the default is 60 seconds.}

\item{...}{Additional parameters.}
}
\value{
A base R plot.
}
\description{
Plot the walking path between selected cases and/or pumps.
}
\section{Note}{
 Arrow represent mileposts or timeposts to the destination.
}

\examples{
plot(walkingPath(15))
plot(walkingPath(15), unit.posts = "time")
}
