% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkingPath.R
\name{plot.walking_path}
\alias{plot.walking_path}
\title{Plot the walking distance between cases and/or pumps.}
\usage{
\method{plot}{walking_path}(x, zoom = TRUE, radius = 0.5,
  unit.posts = NULL, unit.interval = NULL, ...)
}
\arguments{
\item{x}{An object of class "walking_path" created by walkingPath().}

\item{zoom}{Logical.}

\item{radius}{Numeric. Controls the degree of zoom.}

\item{unit.posts}{Character. "distance" for mileposts; "time" for timeposts.}

\item{unit.interval}{Numeric. Intervals: meters fot mileposts, seconds for timeposts.}

\item{...}{Additional plotting parameters.}
}
\value{
A base R plot.
}
\description{
Plot the walking distance between cases and/or pumps.
}
