% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{plot.walkingB}
\alias{plot.walkingB}
\title{Plot method for neighborhoodPaths().}
\usage{
\method{plot}{walkingB}(x, area = FALSE, ...)
}
\arguments{
\item{x}{An object of class "walkingB" created by neighborhoodPaths().}

\item{area}{Logical. TRUE returns expedted area plot. FALSE returns expected walking paths.}

\item{...}{Additional plotting parameters.}
}
\value{
A base R plot.
}
\description{
Plot method for neighborhoodPaths().
}
\examples{
# plot(neighborhoodPaths())
# plot(neighborhoodPaths(observed = FALSE))
# plot(neighborhoodPaths(observed = FALSE), area = TRUE)
}
