% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkingB.R
\name{plot.walkingB}
\alias{plot.walkingB}
\title{Plot method for walkingB().}
\usage{
\method{plot}{walkingB}(x, type = "area.points",
  tsp.method = "repetitive_nn", add = FALSE, path.width = 2,
  alpha.level = 0.5, polygon.type = "solid", polygon.col = NULL,
  polygon.lwd = 2, ...)
}
\arguments{
\item{x}{An object of class "walking" created by \code{walkingNominal()}.}

\item{type}{Character. "roads", "area.points" or "area.polygons". "area" flavors only valid when \code{case.set = "expected"}.}

\item{tsp.method}{Character. Traveling salesperson problem algorithm.}

\item{add}{Logical. Add graphic to plot.}

\item{path.width}{Numeric. Set width of paths.}

\item{alpha.level}{Numeric. Alpha level transparency for area plot: a value in [0, 1].}

\item{polygon.type}{Character. "perimeter" or "solid".}

\item{polygon.col}{Character.}

\item{polygon.lwd}{Numeric.}

\item{...}{Additional plotting parameters.}
}
\value{
A base R plot.
}
\description{
Plot method for walkingB().
}
\note{
When plotting area graphs with simulated data (i.e., \code{case.set = "expected"}), there may be discrepancies between observed cases and expected neighborhoods, particularly between neighborhoods. type = "roads" inspired by Shiode et. al. (2015).
}
