% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongAuxilliaryFunctions.R
\name{pointIndex}
\alias{pointIndex}
\title{Index between native and latlong coordinates.}
\usage{
pointIndex(max.ct = 321)
}
\arguments{
\item{max.ct}{Integer. Upper count of observations.}
}
\description{
Index between native and latlong coordinates.
}
