% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointOverlap.R
\name{pointOverlap}
\alias{pointOverlap}
\title{Census of overlapping fatalties}
\usage{
pointOverlap(inter.point.dist = 0.15)
}
\arguments{
\item{inter.point.dist}{Numeric. Ceiling for overlapping points.}
}
\description{
"Resolution" of georeferenced points with pch = 46 (i.e., pch = ".").
}
