% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi.R
\name{print.voronoi}
\alias{print.voronoi}
\title{Print method for neighborhoodVoronoi().}
\usage{
\method{print}{voronoi}(x, ...)
}
\arguments{
\item{x}{An object of class "voronoi" created by \code{neighborhoodVoronoi()}.}

\item{...}{Additional arguments.}
}
\value{
A list of argument values.
}
\description{
Parameter values for neighborhoodVoronoi().
}
\examples{
neighborhoodVoronoi()
print(neighborhoodVoronoi())
}
