% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi.R
\name{print.voronoi}
\alias{print.voronoi}
\title{Print method for neighborhoodVoronoi().}
\usage{
\method{print}{voronoi}(x, ...)
}
\arguments{
\item{x}{An object of class "voronoi" created by neighborhoodVoronoi().}

\item{...}{Additional arguments.}
}
\value{
A data frame with observed and expected counts, observed percentage, and the Pearson residual, (observed - expected) / sqrt(expected).
}
\description{
Return summary statistics for Voronoi neighborhoods.
}
\examples{
neighborhoodVoronoi()
print(neighborhoodVoronoi())
}
\seealso{
\code{addVoronoi()}
\code{plot.voronoi()}
}
