% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi.R
\name{print.voronoi}
\alias{print.voronoi}
\title{Print method for neighborhoodVoronoi().}
\usage{
\method{print}{voronoi}(x, ...)
}
\arguments{
\item{x}{An object of class "voronoi" created by \code{neighborhoodVoronoi()}.}

\item{...}{Additional arguments.}
}
\value{
A vector with observed counts.
}
\description{
Return observed counts for Voronoi neighborhoods.
}
\examples{
neighborhoodVoronoi()
print(neighborhoodVoronoi())
}
\seealso{
\code{addVoronoi()}
\code{plot.voronoi()}
}
