% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walking.R
\name{print.walking}
\alias{print.walking}
\title{Print method for neighborhoodWalking().}
\usage{
\method{print}{walking}(x, ...)
}
\arguments{
\item{x}{An object of class "walking" created by neighborhoodWalking().}

\item{...}{Additional arguments.}
}
\value{
A data frame with observed and expected counts, observed percentage, and the Pearson residual, (observed - expected) / sqrt(expected).
}
\description{
Summary statistics for walking path neighborhoods.
}
\examples{
summary(neighborhoodWalking())
}
