% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkingNominal.R
\name{print.walking}
\alias{print.walking}
\title{Print method for walkingNominal().}
\usage{
\method{print}{walking}(x, ...)
}
\arguments{
\item{x}{An object of class "walking" created by \code{neighborhoodWalking()}.}

\item{...}{Additional parameters.}
}
\value{
A list of argument values.
}
\description{
Parameter values for neighborhoodWalking().
}
