% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walking.R
\name{print.walking}
\alias{print.walking}
\title{Print method for neighborhoodWalking().}
\usage{
\method{print}{walking}(x, ...)
}
\arguments{
\item{x}{An object of class "walking" created by neighborhoodWalking().}

\item{...}{Additional parameters.}
}
\value{
An R vector.
}
\description{
Return count of paths (anchor cases) by pump neighborhood.
}
\section{Note}{
 Printing an object with case.set = "expected" is computationally intensive (you are computing the paths of almost 5000 cases): on a 2.3 GHz Intel i7 this takes approximately 98 seconds on single core and 49 seconds on 4 physical (8 logical) cores.
}

\examples{
# neighborhoodWalking()
# print(neighborhoodWalking())
}
