% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profilePlot.R
\name{profilePerspective}
\alias{profilePerspective}
\title{Rescale data along axis.}
\usage{
profilePerspective(output = "inside", pump = 7, theta = 0,
  multi.core = FALSE)
}
\arguments{
\item{output}{Character."inside" or "outside".}

\item{pump}{Numeric. Selected pump focal point.}

\item{theta}{Numeric. Angle of perspective axis in degrees.}

\item{multi.core}{Logical.}
}
\description{
Rescale data along axis.
}
