% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profilePlot.R
\name{profilePerspective}
\alias{profilePerspective}
\title{Rescale data along axis.}
\usage{
profilePerspective(output = "inside", pump = 7, angle = 0,
  vestry = FALSE, multi.core = FALSE)
}
\arguments{
\item{output}{Character."inside" or "outside".}

\item{pump}{Numeric. Selected pump focal point.}

\item{angle}{Numeric. Angle of perspective axis in degrees.}

\item{vestry}{Logical. TRUE for the 14 pumps from Vestry Report. FALSE for the original 13 pumps.}

\item{multi.core}{Logical or Numeric. TRUE uses parallel::detectCores(). FALSE uses one, single core. You can also specify the number logical cores. On Windows, only multi.core = FALSE is available.}
}
\description{
Rescale data along axis.
}
