% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profilePlot.R
\name{profilePlot}
\alias{profilePlot}
\title{Profile Plot.}
\usage{
profilePlot(pump = 7, theta = 0, multi.core = FALSE, type = "base")
}
\arguments{
\item{pump}{Numeric. Selected pump focal point.}

\item{theta}{Numeric. Angle of perspective axis in degrees.}

\item{multi.core}{Logical or Numeric. TRUE uses parallel::detectCores(). FALSE uses one, single core. You can also specify the number logical cores. On Windows, only "multi.core = FALSE" is available.}

\item{type}{Character. "base" or "ggplot2".}
}
\description{
Profile Plot.
}
