% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pump.case}
\alias{pump.case}
\title{List of the observed fatality "addresses" grouped by walking-distance pump neighborhood.}
\format{A list 321 IDs spread over 13 vectors.
\describe{
 \item{\code{pump.case}}{numerical ID}
}}
\usage{
pump.case
}
\description{
List of the observed fatality "addresses" grouped by walking-distance pump neighborhood.
}
\section{Notes}{
 \code{\link{neighborhoodWalking}} documents the code for these data.

For details, see \code{vignette}("pump.neighborhoods").
}

\examples{
# pumpCase(neighborhoodWalking())
}
\keyword{datasets}
