% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pump.cases}
\alias{pump.cases}
\title{List of regular "simulated" cases by pump neighborhood}
\format{A data frame with 13 observations and 2 variables.
\describe{
 \item{\code{pump.cases}}{numerical ID}
}}
\usage{
pump.cases
}
\description{
List of regular "simulated" cases by pump neighborhood
}
\keyword{datasets}

