% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pump.cases}
\alias{pump.cases}
\title{List of the 578 observed cholera cases grouped by pump neighborhood.}
\format{A list with 13 vectors.
\describe{
 \item{\code{pump.cases}}{numerical ID}
}}
\usage{
pump.cases
}
\description{
List of the 578 observed cholera cases grouped by pump neighborhood.
}
\keyword{datasets}
