% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpCase.R
\name{pumpCase}
\alias{pumpCase}
\title{Extract numeric case IDs by pump neighborhood.}
\usage{
pumpCase(x, case)
}
\arguments{
\item{x}{An object created by \code{neighborhoodEuclidean()}, \code{neighborhoodVoronoi()} or \code{neighborhoodWalking()}.}

\item{case}{Character. "address" or "fatality"}
}
\value{
An R list of numeric ID of cases by pump neighborhoods.
}
\description{
Extract numeric case IDs by pump neighborhood.
}
\examples{
\donttest{

pumpCase(neighborhoodEuclidean())
pumpCase(neighborhoodVoronoi())
pumpCase(neighborhoodWalking())
}
}
