% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpCases.R
\name{pumpCases}
\alias{pumpCases}
\title{Extract numeric case IDs by neighborhood.}
\usage{
pumpCases(obj)
}
\arguments{
\item{obj}{An object created by neighborhoodVoronoi() or neighborhoodWalking().}
}
\value{
A list of pump neighborhoods with the numeric ID of observed cases.
}
\description{
Extract numeric case IDs by neighborhood.
}
\examples{
pumpCases(neighborhoodVoronoi())
pumpCases(neighborhoodWalking())
}
