% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumpLocator.R
\name{pumpLocator}
\alias{pumpLocator}
\title{Locate well by numerical ID.}
\usage{
pumpLocator(id, zoom = FALSE, radius = 2, vestry = FALSE)
}
\arguments{
\item{id}{Numeric or Integer. Whole number between 1 and 14}

\item{zoom}{Logical.}

\item{radius}{Numeric. Controls the degree of zoom.}

\item{vestry}{Logigcal. FALSE for original 13. TRUE for the 14 from Vestry Report: Hanover Square pump and a relocated Broad Street pump.}
}
\value{
A base R graphics plot.
}
\description{
Plots John Snow's map of the 1854 London cholera outbreak and highlights the
selected water pump.
}
\examples{
pumpLocator(7) # Broad Street Pump
pumpLocator(7, zoom = TRUE)
pumpLocator(7, zoom = TRUE, radius = 1)
pumpLocator(14, vestry = TRUE, zoom = TRUE, radius = 1)
}

