% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regular.cases}
\alias{regular.cases}
\title{"Expected" cases.}
\format{A data frame with 2 variable that records the position 4993 "expected" cases.
 \describe{
    \item{\code{x}}{x-coordinate}
    \item{\code{y}}{y-coordinate}
 }}
\usage{
regular.cases
}
\description{
The result of using sp::spsample() and sp::Polygon() to generate 5000 regularly spaced simulated cases within the map's border.
}
\keyword{datasets}

