% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roadLength.R
\name{roadLength}
\alias{roadLength}
\title{Compute road length.}
\usage{
roadLength(road.name = "Oxford Street", unit = NULL)
}
\arguments{
\item{road.name}{Character. Road name.}

\item{unit}{Character. Unit of measurement.}
}
\value{
An R vector of length one.
}
\description{
Computes length of selected road.
}
