% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roadSegments.R
\name{roadSegments}
\alias{roadSegments}
\title{Reshape 'roads' data frame into 'road.segments' data frame.}
\usage{
roadSegments(latlong = FALSE)
}
\arguments{
\item{latlong}{Logical. Use estimated longitude and latitude.}
}
\value{
An R data frame.
}
\description{
Used to integrate pumps and cases into road network when computing walking neighborhoods.
}
\note{
This function documents the code that generates \code{\link[cholera]{road.segments}}.
}
