% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roadSegments.R
\name{roadSegments}
\alias{roadSegments}
\title{Reshape 'roads' data frame into 'road.segments' data frame.}
\usage{
roadSegments()
}
\value{
An R data frame.
}
\description{
For use in integrating pumps and cases into road network needed to compute walking neighborhoods.
}
\section{Notes}{
 This function documents the code that generates \code{\link[cholera]{road.segments}}.
}

