% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotatePoint.R
\name{rotatePoint}
\alias{rotatePoint}
\title{Rotate points (prototype).}
\usage{
rotatePoint(id = 66, unique.coords = TRUE)
}
\arguments{
\item{id}{Numeric. Road segment endpoint ID.}

\item{unique.coords}{Logical. Use unique coordinates.}
}
\description{
Rotate points (prototype).
}
