% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongAuxilliaryFunctions.R
\name{rotatePoint}
\alias{rotatePoint}
\title{Rotate points (prototype).}
\usage{
rotatePoint(id = 1, dataset = "roads", unique.coords = TRUE)
}
\arguments{
\item{id}{Numeric. Road segment endpoint ID.}

\item{dataset}{Character. "roads", "fatalities", "fatalities.address", "pumps", or "pumps.vestry".}

\item{unique.coords}{Logical. Use unique coordinates.}
}
\description{
Rotate points (prototype).
}
