% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim.ortho.proj}
\alias{sim.ortho.proj}
\title{Orthogonal projection of simulated "expected" cases onto road network.}
\format{A data frame with 5 variable that records the position of the orthogonal projection of the 4993 cases onto the network of roads. Excluding isolates (Pump 2 at  Adam and Eve Court; Falconberg Court and Falconberg Mews), there are 4972 cases.
\describe{
    \item{\code{road.segment}}{"address" road segment}
    \item{\code{x.proj}}{x-coordinate}
    \item{\code{y.proj}}{y-coordinate}
    \item{\code{ortho.dist}}{orthogonal distance to home road segment}
    \item{\code{case}}{numeric case ID}
 }}
\usage{
sim.ortho.proj
}
\description{
Orthogonal projection of simulated "expected" cases onto road network.
}
\section{Notes}{
 \code{\link{simulateFatalities}} documents the code for these data.
}

\keyword{datasets}
