% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim.pump.case}
\alias{sim.pump.case}
\title{List of "simulated" fatalities grouped by walking-distance pump neighborhood.}
\format{A list 4972 IDs spread over 13 vectors.
\describe{
 \item{\code{sim.pump.case}}{numerical ID}
}}
\usage{
sim.pump.case
}
\description{
List of "simulated" fatalities grouped by walking-distance pump neighborhood.
}
\note{
\code{\link{neighborhoodWalking}} documents the code for these data. For details, see \code{vignette}("pump.neighborhoods").
}
\examples{
\donttest{

pumpCase(neighborhoodWalking(case.set = "expected"))
}
}
\keyword{datasets}
