% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snow.neighborhood}
\alias{snow.neighborhood}
\title{Snow neighborhood cases.}
\format{A vector with 384 observations.
\describe{
  \item{\code{snow.neighborhood}}{numeric case ID}
}}
\usage{
snow.neighborhood
}
\description{
Data frame of the numeric IDs of cases, taken from from Dodson and Tobler, that fall within Snow's Broad Street pump neighborhood.
}
\keyword{datasets}

