% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowColors.R
\name{snowColors}
\alias{snowColors}
\title{Create a set of colors for pump neighborhoods.}
\usage{
snowColors(vestry = FALSE)
}
\arguments{
\item{vestry}{Logical. TRUE uses the 14 pumps in the Vestry Report. FALSE uses the original 13.}
}
\value{
A character vector of colors.
}
\description{
Uses RColorBrewer::brewer.pal().
}
