% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowMap.R
\name{snowMap}
\alias{snowMap}
\title{Plot John Snow's cholera map.}
\usage{
snowMap(add.landmarks = TRUE, vestry = FALSE, stacked = TRUE)
}
\arguments{
\item{add.landmarks}{Logical. Add landmarks.}

\item{vestry}{Logical. TRUE uses the 14 pumps from the map in the Vestry Report. FALSE uses the 13 pumps from the original map.}

\item{stacked}{Logical. Use stacked fatalities.}
}
\value{
A base R graphics plot.
}
\description{
Dodson and Tobler's digitization of Snow's map with added landmarks.
}
\examples{
snowMap()
snowMap(vestry = TRUE, stacked = FALSE)
}
\seealso{
\code{\link{addLandmarks}},
\code{\link{addKernelDensity}},
\code{\link{addLandmarks}},
\code{\link{addPlaguePit}},
\code{\link{addSnow}},
\code{\link{addVoronoi}}.
\code{\link{addWhitehead}}
}
