% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowNeighborhood.R
\name{snowNeighborhood}
\alias{snowNeighborhood}
\title{Data for plotting Snow's annotated neighborhood.}
\usage{
snowNeighborhood()
}
\value{
An R list of edge IDs and simulated case IDs (sim.ortho.proj).
}
\description{
Computes "missing" and split road segments data, and area plot data.
}
