% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionAddresses.R
\name{starGraph}
\alias{starGraph}
\title{Decompose and partition star graphs.}
\usage{
starGraph(subgraphs, group = c("25", "45"))
}
\arguments{
\item{subgraphs}{Object. 'igraph' list of graphs.}

\item{group}{Character. Vector of group IDs.}
}
\value{
An R data frame.
}
\description{
Used for graphs with 1 core and 3 peripheral (4) vertices.
}
