% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetHighlight.R
\name{streetHighlight}
\alias{streetHighlight}
\title{Highlight road by name.}
\usage{
streetHighlight(road.name, col = "red", lwd = 3)
}
\arguments{
\item{road.name}{Character vector. The functions tries to correct for case and to remove extra spaces.}

\item{col}{Character. Highlight color.}

\item{lwd}{Numeric. Line width.}
}
\value{
A base R graphics segment(s).
}
\description{
Highlight road by name.
}
\examples{
snowMap()
streetHighlight("Broad Street")
}
