% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetLength.R
\name{streetLength}
\alias{streetLength}
\title{Compute street length.}
\usage{
streetLength(road.name = "Oxford Street", unit = NULL)
}
\arguments{
\item{road.name}{Character. Road name.}

\item{unit}{Character. Unit of measurement: "meter" or "yard". Default is NULL, which returns the map's native scale.}
}
\value{
An R vector of length one.
}
\description{
Computes length of selected named street.
}
\examples{
streetLength("Oxford Street")
streetLength("oxford street")
streetLength("oxford street", unit = "meter")
}
\seealso{
\code{\link{roads}}, code{\link{road.segments}}, \code{\link{streetNameLocator}}, \code{\link{streetNumberLocator}}, \code{vignette("road.names")}
}
