% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetNameLocator.R
\name{streetNameLocator}
\alias{streetNameLocator}
\title{Locate road by name.}
\usage{
streetNameLocator(road.name, zoom = FALSE, radius = 1, unit = NULL,
  all.cases = FALSE)
}
\arguments{
\item{road.name}{Character vector. Note that \code{streetNameLocator}() tries to correct for case and to remove extra spaces.}

\item{zoom}{Logical.}

\item{radius}{Numeric. Controls the degree of zoom. For "radius" <= 5, the numeric ID of all cases or just the anchor case is plotted.}

\item{unit}{Character. Unit of measurement: "meter" or "yard". Default is NULL, which returns the map's native scale.}

\item{all.cases}{Logical. When zoom = TRUE and radius <= 5, all.cases = TRUE plots the numeric ID of all cases; when all.cases = FALSE only the numeric ID of the anchor case is shown.}
}
\value{
A base R graphics plot.
}
\description{
Highlights the selected road and its cases. See the list of road names in \code{vignette}("road.names").
}
\examples{
streetNameLocator("Oxford Street")
streetNameLocator("oxford street")
streetNameLocator("Cambridge Street", zoom = TRUE)
streetNameLocator("Cambridge Street", zoom = TRUE, radius = 0)
}
\seealso{
\code{\link{roads}}, \code{\link{road.segments}}, \code{\link{streetNumberLocator}}, \code{vignette("road.names")}
}
