% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streetNumberViewer.R
\name{streetNumberViewer}
\alias{streetNumberViewer}
\title{Plot road segment by numerical ID.}
\usage{
streetNumberViewer(road.number, zoom = FALSE, radius = 1)
}
\arguments{
\item{road.number}{Numeric or integer. Whole number between 1 and 528.}

\item{zoom}{Logical.}

\item{radius}{Numeric. Controls the degree of zoom. Default value is 1. For values <= 5, the anchor case number is plotted rather than a point for each case.}
}
\value{
A base R graphics plot.
}
\description{
Plots John Snow's map of the 1854 London cholera outbreak and highlights the
selected road segment. See \code{vignette}("road.names") for details.
}
\examples{
streetNumberViewer(243)
streetNumberViewer(243, zoom = TRUE)
streetNumberViewer(243, zoom = TRUE, radius = 0)
}
\seealso{
\code{roads}

\code{\link[cholera]{streetNameViewer}}
}

