% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latitudeLongitudeFatality.R
\name{subsetFatalitiesPDF}
\alias{subsetFatalitiesPDF}
\title{Create PDFs of stratified non-anchor cases.}
\usage{
subsetFatalitiesPDF(path)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/"}
}
\description{
For QGIS geo-referencing.
}
