% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latitudeLongitudeFatality.R
\name{subsetFatalitiesPDF}
\alias{subsetFatalitiesPDF}
\title{Create PDFs of stratified non-anchor cases.}
\usage{
subsetFatalitiesPDF(path, pch = 15, cex = 0.2)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/"}

\item{pch}{Integer. R pch.}

\item{cex}{Numeric.}
}
\description{
For QGIS geo-referencing.
}
