% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latitudeLongitudeFrame.R
\name{subsetFramePDF}
\alias{subsetFramePDF}
\title{Create PDFs of map frame points.}
\usage{
subsetFramePDF(path)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/"}
}
\description{
For QGIS geo-referencing.
}
