% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongAuxilliaryFunctions.R
\name{subsetPDF}
\alias{subsetPDF}
\title{Create subsetted PDFs (prototype).}
\usage{
subsetPDF(path)
}
\arguments{
\item{path}{Character. e.g., "~/Documents/Data/"}
}
\description{
Reduce over-printing of points.
}
