% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclideanDistance.R
\name{summary.euclidean_distance}
\alias{summary.euclidean_distance}
\title{Plot observed and simulated walking neighborhoods.}
\usage{
\method{summary}{euclidean_distance}(object, ...)
}
\arguments{
\item{object}{An object of class "euclidean_distance" created by euclideanDistance().}

\item{...}{Additional parameters.}
}
\description{
Neighborhoods are based on the shortest paths between a fatality's address and its nearest pump.
}
