% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi.R
\name{summary.voronoi}
\alias{summary.voronoi}
\title{Compute summary statistics for Voronoi neighborhoods.}
\usage{
\method{summary}{voronoi}(object, ...)
}
\arguments{
\item{object}{An object of class "voronoi" created by neighborhoodVoronoi().}

\item{...}{Additional arguments.}
}
\value{
A data frame with observed and expected counts, observed percentage, and the Pearson residual, (observed - expected) / sqrt(expected).
}
\description{
Compute summary statistics for Voronoi neighborhoods.
}
\examples{
summary(neighborhoodVoronoi())
}
\seealso{
\code{addVoronoi()}
\code{plot.voronoi()}
}
