% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi.R
\name{summary.voronoi}
\alias{summary.voronoi}
\title{Summary method for neighborhoodVoronoi().}
\usage{
\method{summary}{voronoi}(object, ...)
}
\arguments{
\item{object}{Object. An object of class "voronoi" created by \code{neighborhoodVoronoi()}.}

\item{...}{Additional arguments.}
}
\value{
A vector of counts by neighborhood.
}
\description{
Return computed counts for Voronoi neighborhoods.
}
\examples{
summary(neighborhoodVoronoi())
}
\seealso{
\code{addVoronoi()}
\code{plot.voronoi()}
}
