% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walking.R
\name{summary.walking}
\alias{summary.walking}
\title{Compute summary statistics for walking path neighborhoods.}
\usage{
\method{summary}{walking}(object, ...)
}
\arguments{
\item{object}{An object of class "walking" created by neighborhoodWalking().}

\item{...}{Additional arguments.}
}
\value{
A data frame with observed and expected counts, observed percentage, and the Pearson residual, (observed - expected) / sqrt(expected).
}
\description{
Compute summary statistics for walking path neighborhoods.
}
\examples{
summary(neighborhoodWalking())
}
