% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walking.R
\name{summary.walking}
\alias{summary.walking}
\title{Summary method for neighborhoodWalking().}
\usage{
\method{summary}{walking}(object, ...)
}
\arguments{
\item{object}{Object. An object of class "walking" created by \code{neighborhoodWalking()}.}

\item{...}{Additional parameters.}
}
\value{
An R vector.
}
\description{
Return computed counts for walking neighborhoods.
}
\examples{
\donttest{
summary(neighborhoodWalking())
}
}
