% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanakaContourPlot.R
\name{tanakaContourPlot}
\alias{tanakaContourPlot}
\title{Tanaka contour plot.}
\usage{
tanakaContourPlot(add = FALSE)
}
\arguments{
\item{add}{Logical. Add to existing plot.}
}
\description{
Soho elevation data.
}
