% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{time.series.vestry}
\alias{time.series.vestry}
\title{Vestry Report Time Series Data.}
\format{A data frame with 61 observations and 3 variables.
\describe{
  \item{\code{date}}{calendar date}
  \item{\code{fatalities}}{fatal attacks}
  \item{\code{deaths}}{deaths}
}}
\usage{
time.series.vestry
}
\description{
Time series of recorded fatalities from Vestry Report Appendix B. Useful for studying the effect of removing the handle on the Broad Street pump on 08 September 1854.
}
\keyword{datasets}
