% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timePosts.R
\name{timePosts}
\alias{timePosts}
\title{Compute walking path timeposts.}
\usage{
timePosts(pump.select, timepost.interval = 60, walking.speed = 5,
  multi.core = FALSE)
}
\arguments{
\item{pump.select}{Numeric.}

\item{timepost.interval}{Numeric. Timepost interval in seconds.}

\item{walking.speed}{Numeric. Walking speed in km/hr.}

\item{multi.core}{Logical or Numeric. TRUE uses parallel::detectCores(). FALSE uses one, single core. You can also specify the number logical cores. On Window, only "multi.core = FALSE" is available.}
}
\value{
A list of data frames.
}
\description{
Return coordinates, post labels and post angles.
}
