% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitMeter.R
\name{unitMeter}
\alias{unitMeter}
\title{Convert nominal map distance to yards or meters.}
\usage{
unitMeter(x, unit = NULL)
}
\arguments{
\item{x}{Numeric. Nominal map distance.}

\item{unit}{Character. Unit of measurement: "meter" or "yard". Default is NULL, which returns the map's native scale. See \code{vignette("roads")} for information on conversion.}
}
\description{
A best guess estimate.
}
