% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitMeter.R
\name{unitMeter}
\alias{unitMeter}
\title{Convert nominal map distance to meters or yards.}
\usage{
unitMeter(x, output.unit = "meter")
}
\arguments{
\item{x}{Numeric. Nominal map distance.}

\item{output.unit}{Character. Unit of distance: "meter", "yard" or "nominal". "nominal" returns the map's nominal scale. See \code{vignette("roads")} for information on conversion.}
}
\description{
A best guess estimate.
}
