% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoiPlot.R
\name{voronoiPlot}
\alias{voronoiPlot}
\title{Plot Voronoi diagram.}
\usage{
voronoiPlot(select = NULL, vestry = FALSE, output = NULL)
}
\arguments{
\item{select}{Numeric vector. Selection by numeric id: 1 to 13 for "pumps"; 1 to 14 for "pumps.vestry". Default is NULL; all pumps are used.}

\item{vestry}{Logical. Use the 14 pumps from the Vestry Report rather than the 13 in the original map.}

\item{output}{Character. NULL plots address points. "addresses" plots the total count of addresses at pumps' locations. "fatalities" plots the total count of fatalities at pumps' locations.}
}
\value{
A base R graphics plot.
}
\description{
Plot  Voronoi diagrams of John Snow's 1854 London cholera data.
}
\examples{
voronoiPlot()
voronoiPlot(output = "addresses")
voronoiPlot(vestry = TRUE)
voronoiPlot(select = 6:7)
voronoiPlot(select = -6)
voronoiPlot(select = -6, output = "fatalities")
}
\seealso{
\code{addVoronoi()}
}

