% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkingDistance.R
\name{walkingDistance}
\alias{walkingDistance}
\title{Compute the shortest walking distance between cases and/or pumps.}
\usage{
walkingDistance(origin, destination = NULL, type = "case-pump",
  observed = TRUE, weighted = TRUE, vestry = FALSE, unit = "meter",
  time.unit = "second", walking.speed = 5)
}
\arguments{
\item{origin}{Numeric or Integer. Numeric ID of case or pump.}

\item{destination}{Numeric or Integer. Numeric ID(s) of case(s) or pump(s). Exclusion is possible via negative selection (e.g., -7). Default is \code{NULL}, which returns closest pump or "anchor" case.}

\item{type}{Character "case-pump", "cases" or "pumps".}

\item{observed}{Logical. Use observed or "simulated" expected data.}

\item{weighted}{Logical. \code{TRUE} computes shortest path in terms of road length. \code{FALSE} computes shortest path in terms of nodes.}

\item{vestry}{Logical. \code{TRUE} uses the 14 pumps from the Vestry report. \code{FALSE} uses the 13 in the original map.}

\item{unit}{Character. Unit of distance: "meter", "yard" or "native". "native" returns the map's native scale. "unit" is meaningful only when "weighted" is \code{TRUE}. See \code{vignette("roads")} for information on unit distances.}

\item{time.unit}{Character. "hour", "minute", or "second".}

\item{walking.speed}{Numeric. Walking speed in km/hr.}
}
\value{
An R data frame.
}
\description{
Compute the shortest walking distance between cases and/or pumps.
}
\note{
The function uses a case's "address" (i.e., "anchor" case of a stack) to compute distance. Time is computed using \code{distanceTime()}. Adam and Eve Court, and Falconberg Court and Falconberg Mews, are disconnected from the larger road network and form two isolated subgraphs. This has two consequences: first, only cases on Adam and Eve Court can reach pump 2 and those cases cannot reach any other pump; second, cases on Falconberg Court and Mews cannot reach any pump. Unreachable pumps will return distances of "Inf".
}
\examples{
\dontrun{

# distance from case 1 to nearest pump.
walkingDistance(1)

# distance from case 1 to pump 6.
walkingDistance(1, 6)

# exclude pump 7 from consideration.
walkingDistance(1, -7)

# distance from case 1 to case 6.
walkingDistance(1, 6, type = "cases")

# distance from pump 1 to pump 6.
walkingDistance(1, 6, type = "pumps")
}
}
\seealso{
\code{\link{fatalities}}, \code{vignette("pump.neighborhoods")}
}
