% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkingPath.R
\name{walkingPath}
\alias{walkingPath}
\title{Plot the shortest walking path between cases and/or pumps.}
\usage{
walkingPath(origin, destination = NULL, type = "case-pump", zoom = TRUE,
  radius = 0.5, weighted = TRUE, vestry = FALSE, unit = NULL)
}
\arguments{
\item{origin}{Numeric or Integer. Numeric ID of case or pump.}

\item{destination}{Numeric or Integer. Numeric ID(s) of case(s) or pump(s). Exclusion is possible via negative selection (e.g., -7). Default is NULL, which returns closest pump or "anchor" case.}

\item{type}{Character. "case-pump", "cases" or "pumps".}

\item{zoom}{Logical.}

\item{radius}{Numeric. Controls the degree of zoom.}

\item{weighted}{Logical. TRUE computes shortest path weighted by road distance. FASLE computes shortes path by number of nodes.}

\item{vestry}{Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 in the original map.}

\item{unit}{Character. Unit of measurement: "meter" or "yard". Default is NULL, which returns the map's native scale. Meaningful only when "weighted" = TRUE. See \code{vignette("roads")} for information on unit distances.}
}
\value{
A base R graphics plot.
}
\description{
Beta v.1
}
\examples{
# path from case 1 to nearest pump.
walkingPath(1)

# path from case 1 to nearest pump in meters (appox).
walkingPath(1, unit = "meter")

# path from case 1 to pump 6.
walkingPath(1, 6)

# exclude pump 7 from consideration.
walkingPath(1, -7)

# path from case 1 to case 6.
walkingPath(1, 6, type = "cases")

# path from pump 1 to pump 6.
walkingPath(1, 6, type = "pumps")
}
\seealso{
\code{\link{fatalities}}, \code{vignette("pump.neighborhoods")}
}
