% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkingPath.R
\name{walkingPath}
\alias{walkingPath}
\title{Plot the walking path from a case to the nearest pump.}
\usage{
walkingPath(x, obs = TRUE, zoom = TRUE, radius = 0.5, weighted = TRUE,
  vestry = FALSE, selection = NULL, unit = NULL)
}
\arguments{
\item{x}{Numeric or Integer. Observed cases must be a whole number between 1 and 578. "Simulated" cases must be a whole number between 1 and 4993 with the following exceptions: 3334, 4427, 4428, 4499, 4500, 4501, 4570, 4571, 4572, 4573, 4574, 4643, 4644, 4645, 4646, 4647, 4716, 4717, 4718, 4719, 4720. See "walking.neighborhoods" vignette for details.}

\item{obs}{Logical. TRUE for observed cases; FALSE for "regular" simulated cases.}

\item{zoom}{Logical.}

\item{radius}{Numeric. Controls the degree of zoom.}

\item{weighted}{Logical. Shortest path weighted by road distance.}

\item{vestry}{Logical. TRUE uses the 14 pumps from the Vestry Report. FALSE uses the 13 in the original map.}

\item{selection}{Numeric. Default is NULL and all pumps are used. Ortherwise, selection by a vector of numeric IDs: 1 to 13 for \code{pumps}; 1 to 14 for \code{pumps.vestry}.}

\item{unit}{Character. Default is NULL, which returns the graph's unit scale. "yard" returns distance in yards. "meter" returns distance in meters. Either implies "weighted" is TRUE.}
}
\value{
A base R graphics plot.
}
\description{
Plot shortest walking path from an observed or "simulated" case to the nearest pump.
}
\examples{
walkingPath(1)
walkingPath(1, obs = FALSE)
walkingPath(1, selection = -7) # exclude pump 7
walkingPath(1, selection = 6)  # only consider pump 6
walkingPath(1, unit = "meter", selection = 3) # distance from case 1 to pump 3.
}
\seealso{
\code{fatalities}
\code{fatalities.address}
}

